% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{add_css}
\alias{add_css}
\title{add_css writes out CSS commands for inclusion in htmlfile}
\usage{
add_css(htmlfile)
}
\arguments{
\item{htmlfile}{the name of the htmlfile being worked on, either
the homepage or one of the tab html files reflecting the
different categories of output}
}
\value{
nothing but it does add text to different html files
}
\description{
add_css write the css commands into the htmlfile being
    generated by make_html. This provides the 'pretty' formatting
    using standard cascading style sheets syntax. It is added
    directly to the html file instead of being in a separate file
    to simplify writing the html (no need to create a separate
    file, and then point to it). It also simplifies maintenance of
    the code.
}
