% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{addpops}
\alias{addpops}
\title{addpops adds the populations from a single replicate together}
\usage{
addpops(invar, nyrs, reps)
}
\arguments{
\item{invar}{the summable variable eg catch, recruits, matureB}

\item{nyrs}{the number of years of data}

\item{reps}{the number of replicates}
}
\value{
a 2D matrix of yrs x reps
}
\description{
addpops adds the populations from a single replicate together
    to form zone totals. It can only be used on matureB, exploitB, catch,
    acatch, and recruits. the sum of acatch is already available as TAC in
    the projection dynamics object 'zoneDP'
}
\examples{
print("wait on data")
}
