% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{addrecvar}
\alias{addrecvar}
\title{addrecvar adds recruitment variation to end of conditioning step}
\usage{
addrecvar(
  zoneDD,
  zoneC,
  glob,
  condC,
  ctrl,
  varyrs,
  calcpopC,
  fleetdyn,
  hsargs,
  sigR = 1e-08,
  sigB = 1e-08,
  lastsigR = 0.3
)
}
\arguments{
\item{zoneDD}{the dynamic zone after historical fishery data has been used
    to finalize the conditioning
#param zoneDDR the empty dynamic zone object ready for expansion and variation}

\item{zoneC}{the zone's constants object for each population}

\item{glob}{the object containing the global constants}

\item{condC}{the object containing the historical fishery data}

\item{ctrl}{the control object}

\item{varyrs}{the number of years at the end of the historical period to
which recruitment variation is to be added}

\item{calcpopC}{a function that takes the output from hcrfun and generates
the actual catch per population expected in the current year.}

\item{fleetdyn}{a function that calculates the distribution of catch across
the sau and populations. Currently not needed by Tas but needed by SA}

\item{hsargs}{the constants used to define the workings of the hcr}

\item{sigR}{the initial recruitment variation default=1e-08}

\item{sigB}{the initial biomass cpuie variation default = 1e-08}

\item{lastsigR}{the recruitment variation to be added to the final varyrs}
}
\value{
an initialized dynamics zone object for the projections with the
    first year populated
}
\description{
addrecvar replicates the historical depletion step in zoneDD
    reps times and copies the first Nyrs - varyrs of values into each
    replicate. Then it steps through each replicate adding recruitment
    variation to the dynamics for the last varyrs years. This is so that
    when the projections begin each replicate will already have recruitment
    variability fully developed and there will be no time lag on the
    recruitment and subsequent dynamics. Take especial note that prior to the
    projections the proposed harvest strategy needs to be applied to the
    historical fishery data (in Tasmania this is currently catches and CPUE).
    This application occurs in lines 60 - 65 in the addrecvar function. This
    WILL NEED ATTENTION with other jurisdictions.
}
\examples{
print("wait on suitable data-sets")
}
