% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{addrepvar}
\alias{addrepvar}
\title{addrepvar adds variation to the start of a generic simulation run}
\usage{
addrepvar(zoneC, zoneDR, inHt, glb, ctrl, nyrs = 20)
}
\arguments{
\item{zoneC}{the constant zone object. Its selectivity patterns should be
altered to reflect the projection conditions using modzoneCSel}

\item{zoneDR}{the dynamic zone object expanded to include all replicates,
using makezoneDR}

\item{inHt}{the initial harvest strategy that was used to deplete the zone}

\item{glb}{the globals object}

\item{ctrl}{the ctrl object}

\item{nyrs}{defaults to 20, which appears to be enough to produce the
required variation}
}
\value{
a revised zoneDR object
}
\description{
addrepvar conducts projyrs of simulation at the same constant
    harvest rate used to deplete each population to the desired level, only
    it does this with recruitment variability turned on, which sets up a
    series of replicates with different starting positions.
}
\examples{
print("wait on suitable data")
}
