% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{addwarning}
\alias{addwarning}
\title{addwarning carefully adds a single line to a predefined warning file}
\usage{
addwarning(warnfile, issue)
}
\arguments{
\item{warnfile}{the full path and name of a text file that aims to hold any
warnings that are reported to it.}

\item{issue}{a character string describing the issue of any warning that one
wants to issue.}
}
\value{
nothing but it does append a character string as a newline to a file
}
\description{
addwarning addresses the need of capturing any warnings that are
    issued during a simulation run. A predefined, full path warning file is
    input along a single line of text describing the issue. In the case of
    aMSE, the name of the warning file is carried in the globals object. At
    the finish of the run, any warning messages are included in the runnotes
    that are printed to the home tab in the output.
}
\examples{
print("wait for me to make one up")
# syntax:  addwarning(glb$warnfile,"")
}
