% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docondition.R
\name{adjustavrec}
\alias{adjustavrec}
\title{adjustavrec re-fits the AvRec saudatafile values in the MSE}
\usage{
adjustavrec(rundir, glb, ctrl, calcpopC, verbose = TRUE)
}
\arguments{
\item{rundir}{the full path to the directory in which all files relating to a
particular run are to be held.}

\item{glb}{the globals object}

\item{ctrl}{the control object - contains datafile and controlfile names}

\item{calcpopC}{a function that takes the output from hcrfun and generates
the actual catch per population expected in the current year.}

\item{verbose}{Should progress comments be printed to console, default=TRUE}
}
\value{
nothing though it will write results to the console if verbose=TRUE
}
\description{
adjustavrec is used during the conditioning of the MSE. Once the
    different SAU have been characterized and the data file values settled,
    when do_condition is run the fit of the predicted cpue and
    size-composition data can often be improved by re-fitting the AvRec
    constants in each SAU to adjust their values to better suit the population
    characteristics once random variation is added to each population's
    properties. adjustavrec does this automatically. Be wary of this function
    because it directly alters the values of AvRec inside the saudata.csv
    file.
}
\examples{
\dontrun{
  # needs at least libraries, aMSE, hutils, makehtml, TasHS
  dropdir <- getDBdir()
  prefixdir <- paste0(dropdir,"A_codeUse/aMSEUse/scenarios/")
  startime <- Sys.time()
  postfixdir <- "S21"
  verbose <- TRUE
  rundir <- filenametopath(prefixdir,postfixdir)
  controlfile <- paste0("control",postfixdir,".csv")
  confirmdir(rundir)
  data(zone1)
  data(saudat)
  rewritecontrolfile(rundir,zone1,controlfile=controlfile)
  rewritedatafile(rundir,zone1,saudat)
  rewritecompdata(rundir,zone1)
  hsargs <- list(mult=0.1,wid = 4,targqnt = 0.55,
                 maxtarg = c(150,150,150,150,150,150,150,150),
                 pmwts = c(0.65,0.25,0.1),
                 hcr = c(0.25,0.75,0.8,0.85,0.9,1,1.05,1.1,1.15,1.2),
                 startCE = 1992)
  out <- do_condition(rundir,controlfile,calcpopC=calcexpectpopC,
                      verbose = TRUE,doproduct = TRUE,dohistoric=TRUE,
                      mincount=120)
  makeoutput(out,rundir,postfixdir,controlfile,hsfile="TasHS Package",
             doproject=FALSE,openfile=TRUE,verbose=FALSE)
  adjustavrec(rundir,out$glb,out$ctrl,calcpopC=calcexpectpopC,verbose=TRUE)
 # Now repeat the do_condition and makeoutput steps to see the final result
}
}
