% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docondition.R
\name{adjustqest}
\alias{adjustqest}
\title{adjustqest recalibrates qest when conditioning for instantaneous Fs}
\usage{
adjustqest(out)
}
\arguments{
\item{out}{the output from either do_condition or do_MSE}
}
\value{
invisibly a list of obsce, predce, compce, cemult, qest, newqest,
    and ssq
}
\description{
adjustqest is required when preparing or conditioning aMSE to
    use instantaneous Fs instead of annual harvest rates. Because the
    dynamics of the operation are very different the absolute predicted
    values of CPUE are changed but mostly run parallel to those produced
    when conditioned on harvest rates. This function runs through each SAU
    and optimizes the match between predicted and observed CPUE solely by
    changing the qest values. This should bring the conditioning very close
    to where it was with the harvest rate conditioning. Further improvement
    could be produced by using optiimzerecdevs (which, with hind sight,
    should have been called adjustrecdevs - sigh).
}
\examples{
print("Run do_MSE on the Harvest Rate conditioned zone.")
print("Then run adjustqest on the 'out' object from do_MSE.")
print("That will write an extra line into the saudata...csv file.")
print("put: harvest, 0, use harvest rates = 0, instant F = 1.")
print("at the bottonm of the START section in the controlfile")
print("Switch to become harvest = 1, and the qestF will be put into")
print("the qest part of the saudat information and that should all work.")
}
\seealso{
{
  \link{optimizerecdevs}, \link{adjustavrec}
}
}
