% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMSY.R
\name{allMSY}
\alias{allMSY}
\title{allMSY estimates MSY by population, SAU, and Zone for each LML in data}
\usage{
allMSY(
  msydir,
  rundir,
  ctrlfile,
  uplim = 0.375,
  inc = 0.005,
  verbose = TRUE,
  newLML = NULL
)
}
\arguments{
\item{msydir}{full path to the directory in which the results are to be
stored. Best to use confirmdir(msydir,ask=FALSE) to be sure it exists.}

\item{rundir}{the rundir from the original analysis}

\item{ctrlfile}{the name of the controlfile, which will be added to the
rundir so the file can be read}

\item{uplim}{upper limit to the harvest rates to search for productivity}

\item{inc}{the increment from 0 up to uplim. The smaller this is the longer
it will take.}

\item{verbose}{should comments and warnings go to the console. Default=TRUE}

\item{newLML}{the numerical value of a previously unused LML for which the
productivity is required. Not yet implemented!}
}
\value{
a list of the glb, ctrl, warning file path, and total time take
    (all required by makehtml), plus summarymsy a list of nlml lists of
    sauprod, zoneprod, and sauprodzone for each LML. Finally, fro each LML,
    inside outproduct is a copy of zoneC, zoneD, product, and glb.
}
\description{
allMSY discovers all values for the LML (nlml) in the
    conditioning data and then estimates the productivity and MSY details
    at the population, SAU, and Zone scales. It generates a webpage output
    using makehtml.
}
\examples{
print("no chance")
}
