% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{alltosau}
\alias{alltosau}
\title{alltosau sums population properties to form SAU totals}
\usage{
alltosau(zoneDR, glb)
}
\arguments{
\item{zoneDR}{the dynamics projeciton object}

\item{glb}{the global constants object}
}
\value{
a list of some SAU properties
}
\description{
alltosau sums the projected values for matureB, exploitB,
    acatch, catch, and recruit, by population, into totals by SAU, using
    sauindex from glb
}
\examples{
print("wait on appropriate internal data sets")
}
