% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{anyclosed}
\alias{anyclosed}
\title{anyclosed determines if any SAU experienced closure in the projections}
\usage{
anyclosed(zoneDP, glb)
}
\arguments{
\item{zoneDP}{the zoneDP object following projections}

\item{glb}{the globals object}
}
\value{
a list of a boolean, closed, and a list of the replicates within
    each SAU that had a closure where the components of that list are
    either NULL or a vector of indices.
}
\description{
anyclosed is used to determine whether any of the SAU have a
    closure during any of the replicates. It identifies which replicates for
    each SAU experienced a closure so that their dynamics within zoneDP can
    be compared with the dynamics of those replicates that did not have a
    closure. Currently, the assumption is that there can only ever be a
    single closure in any single replicate.
}
\examples{
# syntax outclosed <- anyclosed(out$zoneDP,out$glb)
}
