% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{aszone}
\alias{aszone}
\title{aszone sums the various SAU dynamics and fishery properties}
\usage{
aszone(sauzone, zoneCP)
}
\arguments{
\item{sauzone}{the output from the applyharvest strategy function, a large
list of arrays or results}

\item{zoneCP}{the constant part of the projection zone}
}
\value{
a list of 8 nyrs x reps matrices, summarizing the fishery outputs at
    the geographical scale of the zone
}
\description{
aszone calculates zone wide totals by summing the mature and
    exploitable biomass across SAU, it also sums the catches and recruitment
    levels. It calculates zone wide harvest rates by dividing the catches by
    the available exploitable biomass. It also calculates the depletion
    levels of both the mature and exploitable biomass by dividing their
    zone totals through time by the initial zone wide B0 and ExB0. Finally,
    it calculates a zone-wide cpue by catch-weighting each of the SAU cpue
    values and then summing each set of SAU for each year and iteration.
}
\examples{
print("wait on data")
}
