% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{biology_plots}
\alias{biology_plots}
\title{biology_plots generates a series of stored plots and tables}
\usage{
biology_plots(resdir, runlabel, glb, zoneC)
}
\arguments{
\item{resdir}{the results directory}

\item{runlabel}{the runlabel, ideally from the ctrl object}

\item{glb}{the globals list}

\item{zoneC}{the zonal constants by population, zoneC}
}
\value{
nothing but it does plot 3 plots and put one table into
    resdir
}
\description{
biology_plots generates the yield vs spawning biomass,
    weight-at-lenght, and emergence plots for all populations. In
    addition, it also tabulates the biological properties of each
    population and SAU and the total zone
}
\examples{
print("this will be quite long when I get to it")
}
