% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{blockE13}
\alias{blockE13}
\title{blockE13 is an abalone data-set for testing performance measures}
\format{
A data.frame of abalone fishery data
\describe{
  \item{year}{the year of fishing}
  \item{coeff}{the back-transformed coefficients from a standardization}
  \item{scaled}{the coefficients scaled to a mean of one}
  \item{cpue}{the coefficients scaled to the nominal geometric mean of the
      time series, to place it on the nominal scale}
  \item{catch}{the related catch from the eastern parts of block 13}
}
}
\source{
Mundy, C. and J.M. McAllister (2020) Tasmanian Abalone Assessment
    2019. IMAS, University of Tasmania.
}
\description{
blockE13 is a fishery data-set for blacklip abalone
    (\emph{Haliotis rubra}) from block 13 in the eastern zone, this is
    Tasmania's Block 13. It constitutes three time-series of the same cpue
    data in different formats (see below). It is for use when testing the
    performance measures within Tasmania's MCDA, although it could be used
    for other purposes, such as illustrating the typical linear relationship
    between catch and CPUE. Note there will only ever be one less PM value
    than there are cpue data in the time-series.
}
\section{Subjects}{

 \itemize{
   \item performance measure estimation
   \item relationship between catch and CPUE
 }
 @export
}

\examples{
data(blockE13)
blockE13
}
