% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{calccatchHSPM}
\alias{calccatchHSPM}
\title{calccatchHSPM calculates all the HSPMs for the actual catch data}
\usage{
calccatchHSPM(catch, glbc, scenes, aavcyrs = 10)
}
\arguments{
\item{catch}{a list of the actual catches from the out$sauout$zonePsau$catch
array from each scenario's out object}

\item{glbc}{a list of the globals objects from each scenario}

\item{scenes}{the names of the different scenarios being compared}

\item{aavcyrs}{the number of projection years to use for the aavc calculation
the default = 10}
}
\value{
a list of  the aavc x sau x scene, and the sum5 and sum10 for the
    same sau and scenes. Plus the medians across all replicates for each
    sau and zone for each of the three PMs.
}
\description{
calccatchHSPM calculates all the Harvest Strategy Performance
    Measures for the actual catch data. The number of years over which the
    aavc is calculated can be modified using the aavcyrs argument. The output
    data can be analysed separately or plotted (eg as boxplots)
}
\examples{
print("wait on data sets")
}
\seealso{
{
   \link{getprojyrs}, \link{getprojyraavc},  \link{catchHSPM}
}
}
