% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{calcprojsel}
\alias{calcprojsel}
\title{calcprojsel generates the selectivity and selectivity x weight}
\usage{
calcprojsel(zoneC, projC, glb)
}
\arguments{
\item{zoneC}{the constants object}

\item{projC}{the projection object from zone1}

\item{glb}{the globals object}
}
\value{
a list of projSel and projSelWt
}
\description{
calcprojsel is used to produce each projection years'
    selectivity and selectivity x weight-at-size. projSel is an
    Nclass x projyrs matrix, a separate selectivity for each projection year,
    which allows for changing the selectivity during the projection. The
    projSelWt is a Nclass x projyrs x numpop array, required because each
    population's weight-at-size relationship will be slightly different. Now
    appears to be redundant.
}
\examples{
 print("wait on new example data")
}
