% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{calcsau}
\alias{calcsau}
\title{calcsau compares an input variable to a constant for each SAU}
\usage{
calcsau(invar, saunames, ref0)
}
\arguments{
\item{invar}{either zoneDP$matureB or zoneDP$exploitB}

\item{saunames}{the names of each SAU}

\item{ref0}{either a vector of B0 or ExB0 for each SAU}
}
\value{
an array of projyrs x nSAU x reps
}
\description{
calcsau divides an input variable, such as matureB or exploitB
    by their respective constant, such as B0 or ExB0, from a vector of
    values for each SAU. This is used to calculate the depletion levels of
    mature and exploitable biomass
}
\examples{
print("wait on new data")
}
