% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS-HCR.R
\name{calibrateMCDA}
\alias{calibrateMCDA}
\title{calibrateMCDA uses historical CPUE to calibrate the MCDA}
\usage{
calibrateMCDA(histCE, saunames, hsargs, mult = 1.1)
}
\arguments{
\item{histCE}{this is a matrix of nominal scale CPUE from the fishery of
interest. Usually standardized. Obtainable from the projC object but
also the condC object}

\item{saunames}{the names of each SAU (from zone1$SAUnames)}

\item{hsargs}{the arguments used within the Harvest strategy's HCR. See the
description for details.}

\item{mult}{the multiplier on the bounds to expand them upwards and
downwards. default value = 1.1 = 10 percent increase either side}
}
\value{
a list of the three PM scores, and the total scores, the TAC
    multipliers, and the Target cpue target and limit.
}
\description{
calibrateMCDA uses the condC$histCE to apply the getgrad1,
    getgrad4, and targscore functions so as to provide vectors of
    performance measure scores to enhance the calibration of the MCDA before
    any projections are done. This is not absolutely required but improves
    the reality of the process. It also means that the productivity of each
    SAU needs to be scaled to the current cpue (see Conditioning the MSE).
    This uses getgrad1 and getgrad4 instead of getgradone and getgradwid as
    it needs to process whole vectors of cpue per sau rather than used a
    row of cpue per sau.
}
\examples{
print("wait on example data being available")
}
