% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{catchvsMSY}
\alias{catchvsMSY}
\title{catchvsMSY by sau divides the projected catches by the sau MSY value}
\usage{
catchvsMSY(catch, glbc, prods, scenes)
}
\arguments{
\item{catch}{the projected catches array of all years x nsau x replicates}

\item{glbc}{the globals object for each scenario}

\item{prods}{a list of the production statistics matrix for each scenario,
this is the out$sauprod matrix stored for each scenario.}

\item{scenes}{the names of the different scenarios being compared in the
same order in which the out objects are decomposed into their components}
}
\value{
a list of 3D arrays of the ratio of catch/MSY for each sau in each
    scenario
}
\description{
catchvsMSY is used when calculating the ratio of the projected
    catches to the MSY. This uses the estimate of MSY for each sau that is
    given in out$sauprod object from each scenario, each of which is placed
    into the 'prods' list.
}
\examples{
print("wait on data sets")
# catch=catch;glbc=resout$glbc;prods=prods;scenes=scenes
}
