% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{catchweightCE}
\alias{catchweightCE}
\title{catchweightCE uses historical catch-by-sau to make weighted zone cpue}
\usage{
catchweightCE(cedat, cdat, nsau)
}
\arguments{
\item{cedat}{historical cpue from condC}

\item{cdat}{historical catches from condC}

\item{nsau}{the number of SAU, from glb}
}
\value{
a vector of catch-weghted cpue for the zone from SAU cpue estimates
}
\description{
catchweightCE is used when characterizing the historical
    fisheries data. It uses matching catch-by-sau to generate catch-weighted
    estimates of zone-wide CPUE through time.
}
\examples{
print("wait on suitable data sets")
}
