% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{CBmsyphaseplot}
\alias{CBmsyphaseplot}
\title{CBmsyphaseplot makes a zone scale phase plot of Cdivmsy vs BdivBmsy}
\usage{
CBmsyphaseplot(rundir, zone, prods, glbc, category = "a", console = TRUE)
}
\arguments{
\item{rundir}{the directory in which all results are held for a scenario or
comparison of scenarios}

\item{zone}{the list of zone outputs for each scenario}

\item{prods}{a list of the production statistics matrix for each scenario,
this is the out$sauprod matrix stored for each scenario.}

\item{glbc}{the globals object for each scenario (needed in case the hyrs
differ between scenarios)}

\item{category}{the tab name within the internal website into which to
place he plot, default = 'a' to ensure it has something}

\item{console}{should each plot go to the console or be saved to rundir.
default = TRUE ie go to console}
}
\value{
an invisible list of lists - Cvsmsy, BvsBmsy, and goodplace
}
\description{
CBmsyphaseplot is used within do_comparison to compare harvest
    strategy scenarios at the zone level by using phase plots of predicted
    catches divided by MSY against predicted mature biomass divided by Bmsy.
    Some HS can recommend taking catches >= MSY even though the biomass
    available is not sufficient for such catches to be sustainable or allow
    for rebuilding. The percent of replicates within the dangerous region
    where catches >= MSY yet biomass is below Bmsy is given topleft. Top
    right is the goodplace, with the  numbers being the minimum, median,
    and maximum number of years it takes for both catches >= MSY AND
    mature biomass >= Bmsy.
}
\examples{
print("wait on example data sets")
# CBmsyphaseplot(rundir=rundir,zone=zone,prods=prods,glbc=glbc,
#                category="zone",console=FALSE)
}
\seealso{
\link{do_comparison}
}
