% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docondition.R
\name{changecolumn}
\alias{changecolumn}
\title{changecolumn alters a selected column of recdevs in the control file}
\usage{
changecolumn(rundir, filename, linerange, column, newvect, verbose = FALSE)
}
\arguments{
\item{rundir}{the rundir for the scenario}

\item{filename}{the character name of the control file}

\item{linerange}{the linerange within the control file containing the rows
representing the selected years of recruitment deviates. Obtained by
using 'getrecdevcolumn}

\item{column}{the sau + 1 to account for the initial year in each row of
deviates}

\item{newvect}{the new values to replace those present. If using optim these
would be exp(ans$par), if 'nlm' exp(ans$estimate)}

\item{verbose}{should console reports of before and after be made?
default = FALSE}
}
\value{
nothing although values within the control file will be changed and,
    if verbose=TRUE, it will write to the console.
}
\description{
changecolumn is designed to be used when conditioning the OM
    when optimizing ad hoc recruitment deviates by calculating the SSQ
    between the predicted and observed CPUE for a selected range of years.
    If the length of new values is not the same as the selected linerange
    the function will stop with a warning. When a line is changed then all
    -1 values are changed to +1. Used in optimizerecdevs
}
\examples{
print("wait on suitable internal data sets")
}
