% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{changeline}
\alias{changeline}
\title{changeline replaces a given line in a given file with new text}
\usage{
changeline(rundir, filename, linenumber, newline, verbose = FALSE)
}
\arguments{
\item{rundir}{the directory path in which to find the text file.}

\item{filename}{the full name of the text file in quotations.}

\item{linenumber}{the line number of the text file to be changed}

\item{newline}{the character string with which to replace the line}

\item{verbose}{should confirmation be output to the console. default=FALSE}
}
\value{
nothing but it does alter a line in a text file. Optionally it may
    confirm the action to the console
}
\description{
changeline enables a text file to be changed line by line.
    One identifies a given line, perhaps by using 'findlinenumber', and
    this can then be replaced by an input character string. Obviously this
    is a fine way to mess up a data file so use with care.
}
\examples{
print("wait on an example")
}
\seealso{
{
 \link{findlinenumber}, \link{changevar}
}
}
