% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{changevar}
\alias{changevar}
\title{changevar can alter the value in a single line in, eg, the control file}
\usage{
changevar(filename, rundir, varname, newvalue, prompt = FALSE, verbose = TRUE)
}
\arguments{
\item{filename}{name of the file to be changed. eg 'control.csv'}

\item{rundir}{name of the scenario directory in which the file is to be found}

\item{varname}{the name of the object whose value is to be changed}

\item{newvalue}{the new value of the object}

\item{prompt}{should allowable names first be listed with current values?}

\item{verbose}{should concluding remarks to console be made. default=TRUE}
}
\value{
nothing but a file is changed - be careful
}
\description{
changevar is a DANGEROUS function for lazy people. I say that
    because it can damage your control files if not used carefully. It speeds
    changing a single value within, say, the control.csv file. For example,
    if one wanted to conduct a retrospective analysis on what would have
    happened should we have introduced the HS sooner, one could use
    something like the following to change the value to 44 implying to start
    projections in 2017.
    changevar(varname="CATCHES",newvalue=44,filename="controlsau.csv",
              rundir=rundir)
     This can only change a single value on a single line but can be used to
     alter the values of the variables listed in 'goodnames'. Currently, this
     only works for Windows machine. Let me know if others are required.
}
\examples{
print("Wait on some time passing")
}
\seealso{
{
 \link{findlinenumber}, \link{changeline}
}
}
