% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{checkmsedata}
\alias{checkmsedata}
\title{checkmsedata contains some tests of the input MSE saudata file}
\usage{
checkmsedata(intxt, rundir, verbose = TRUE)
}
\arguments{
\item{intxt}{the data file from readLines used in readsaudatafile}

\item{rundir}{the rundir for the scenario}

\item{verbose}{should test results be put to the console as well as filed,
default=TRUE}
}
\value{
nothing but will write a file to rundir and may write to the console
}
\description{
checkmsedata contains some tests of the input MSE saudata file
}
\examples{
print("wait on example data sets")
}
\seealso{
{
 \link{readsaudatafile}
}
}
