% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{checkresdir}
\alias{checkresdir}
\title{checkresdir checks resdir contains the required csv files}
\usage{
checkresdir(resdir, ctrlfile = "control.csv")
}
\arguments{
\item{resdir}{the directory in which all all files relating to a
particular run are to be held.}

\item{ctrlfile}{default="control.csv", the filename of the control
file present in resdir containing information regarding the
run.}
}
\value{
the control list for the run
}
\description{
checkresdir checks resdir contains the required csv
    files including the named control file, which then contains
    the names of the region data file, and the population data
    file. The run stops if any are not present or are misnamed.
}
\examples{
resdir <- tempdir()
ctrlfiletemplate(resdir)
regionfiletemplate(resdir)
datafiletemplate(6,resdir,filename="reg1smu2pop6.csv")
ctrl <- checkresdir(resdir)
ctrl
}
