% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{checksizecompdata}
\alias{checksizecompdata}
\title{Title}
\usage{
checksizecompdata(rundir, controlfile, verbose = TRUE)
}
\arguments{
\item{rundir}{the directory in which all files relating to a
particular run are to be held.}

\item{controlfile}{default="control.csv", the filename of the control
file present in rundir containing information regarding the run.}

\item{verbose}{Should progress comments be printed to console, default=TRUE#'}
}
\value{
nothing but it does plot a graph
}
\description{
Title
}
\examples{
print("wait on data sets")
}
