% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{comparedynamics}
\alias{comparedynamics}
\title{comparedynamics plots medians and quantiles of dynamic variables}
\usage{
comparedynamics(rundir, dyn, glbc, scenes)
}
\arguments{
\item{rundir}{to location for all results of the comparisons}

\item{dyn}{a list of the dynamics objects from each scenario's out object}

\item{glbc}{a list of the globals objects from each scenario}

\item{scenes}{a vector of the names for each scenario}
}
\value{
nothing but it does add four plots to rundir and four lines to the
    resultTable.csv file in rundir ready for a makecompareoutput
}
\description{
comparedynamics plots, for each sau, the medians and 90th
    quantiles of the distributions of the reps for each scenario for four
    dynamic variables: cpue, catch (actual catches not aspirational),
    mature biomass, and harvets rate.
}
\examples{
\dontrun{
  print("wait on an example")
# rundir=rundir;dyn=dyn;glbc=glbc;scenes=scenes
}
}
