% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{comparefinalscores}
\alias{comparefinalscores}
\title{comparefinalscores compares the median final HS scores from scenarios}
\usage{
comparefinalscores(
  rundir,
  scores,
  scenes,
  legloc = "bottomright",
  filen = "",
  category = "Tables",
  resol = 300
)
}
\arguments{
\item{rundir}{to location for all results of the comparisons}

\item{scores}{derived from the out$scores object saved from each scenario}

\item{scenes}{the scenario names}

\item{legloc}{the location of the legend, default = bottomright}

\item{filen}{the name of the png file if it is to be stored, default=""}

\item{category}{in what tab should the tables be stored. default="Tables"}

\item{resol}{default = 300, the dpi for the png plots}
}
\value{
nothing but it does generate a plot.
}
\description{
comparefinalscores takes the 'scores' object from each 'out'
    object generated by do_MSE and plots, for each sau, the median final
    HS score for each scenario within each sau, using the scenario name as
    the legend entry. 'scores' is a list object made up of two list objects
    generated in do_MSE using both 'getcpueHS' and 'plotfinalscores'. Here we
    use only the outmed list and not the outHS list. The outmed list contains
    the medians of each of the replicates scores held in outHS, and this
    function plots the median final scores for each scenario against each
    other for each sau.
}
\examples{
print("wait on data")
}
