% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{compzoneN}
\alias{compzoneN}
\title{compzoneN compares numbers-at-size before/after depletion}
\usage{
compzoneN(unfN, curN, glb, yr, depl, ssc = 5, LML = 0, rundir = "")
}
\arguments{
\item{unfN}{the unfished numbers-at-size from getzoneprops}

\item{curN}{the current numbers-at-size from getzoneprops}

\item{glb}{the global object}

\item{yr}{the year of the dynamics}

\item{depl}{the depletion level of the current n-a-s}

\item{ssc}{index for starting size class. thus 1 = 2, 2 = 4, 5 = 10, etc.
default = 5 for it plots size classes from 10mm up}

\item{LML}{the legal minimum length in the comparison year}

\item{rundir}{the results directory, default = "" leading to no
.png file, just a plot to the screen.}
}
\value{
invisibly the filename ready for logfilename
}
\description{
compzoneN generates a plot comparing the unfished
    numbers-at-size with those for a given level of depletion.
}
\examples{
print("still to be developed")
# unfN=unfN; curN=depN;glb=glb; yr=1; depl=0.3993; LML=132; rundir=rundir
}
