% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{condDat}
\alias{condDat}
\title{condDat is an abalone data-set for conditioning the MSE}
\format{A list of hypothetical conditioning data for the MSE
\describe{
  \item{constants}{the PDFs used to describe the biological parameters}
  \item{Lt}{the length at tagging}
  \item{Dt}{The time interval between tagging and recapture, in this
      instance they are all listed as 1 year}
  \item{DL}{the growth increment in mm}
}}
\description{
condDat is an abalone data-set for blacklip abalone
    (\emph{Haliotis rubra}) that can be used to condition the
    operating model as an example when running the aMSE function
    examples. It describes a region containing two hypothetical
    Tasmanian blocks (the level of SMU) with a total of six
    populations. If this were a data-set in a file it would require
    the readdatafile function to read the file. The function
    datafiletemplate will generate a template of the required
    format to be read by readdatafile, which can be edited
    to suit a given fishery.
}
\examples{
 data(condDat)
 str(condDat)
}
