% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{confirmdir}
\alias{confirmdir}
\title{confirmdir checks to see if a directory exists and makes it if not}
\usage{
confirmdir(x, make = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{the directory to be checked and created if necessary}

\item{make}{should the directory be created if it does not already exist?
default=TRUE}

\item{verbose}{should responses be sent to the console? default=TRUE}
}
\value{
nothing but it can create a directory
}
\description{
confirmdir enables one to be sure a selected directory exists.
    If it has not bee created then confirmdir will create it if it does not
    already exist. This is useful when defining output directories on the
    hard drive where large objects may be stored.
}
\examples{
x <- tempdir()
confirmdir(x)
}
