% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{constants}
\alias{constants}
\title{constants is conditioning data for 16 populations and 8 SAU}
\description{
constants is a data.frame of parameters for blacklip
    abalone (\emph{Haliotis rubra}) used to condition the
    operating model as an example when running the aMSE function
    examples. It describes a zone containing eight hypothetical
    Tasmanian blocks (the level of SAU) with a total of sixteen
    populations (two per SAU). If this were a data-set in a CSV file it
    would require the readdatafile function to read the file. The
    function datafiletemplate will generate a template of the required
    format to be read by readdatafile, which can be edited to suit a
    given fishery. once transposed and with random variation included,
    this data becomes the contents of popdefs.
}
\section{contents}{

\itemize{
  \item popnum the index to the population
  \item SAU the spatial management unit number = blocks in Tasmania
  \item DLMax maximum growth increment used in the inverse logistic
  \item sMAxDL the sd variability of DLMax, Normal variation
  \item L50  the shell length at 50% of maximum growth increment
  \item sL50 the sd variability of L50, Normal variation
}
}

\examples{
 data(constants)
 print(constants)
}
