% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{constructTasHSout}
\alias{constructTasHSout}
\title{constructTasHSout remakes outputs from the hcrfun for all projections}
\usage{
constructTasHSout(hcrfun, ce, acatch, glb, hsargs)
}
\arguments{
\item{hcrfun}{the same as used in do_MSE = mcdahcr from TasHS package}

\item{ce}{the 3D array of cpue, which, for Tas, will include all years back
to 1992, the start of defensible cpue data.}

\item{acatch}{the projected aspirational catches required by mcdahcr}

\item{glb}{the global object for the given scenario  out$glb}

\item{hsargs}{the hsargs used in the scenario out$hsargs}
}
\value{
invisibly a list of multTAC, grad1, grad4,score1, score4, scoret,
    scoretot, and targrefpts. Each is a 3D array of years x sau x reps,
    except for the targrefpts which is the sau x 4 matrix of refpts for the
    final year in each replicate.
}
\description{
constructTasHSout uses the scenario's hcrfun to reconstrcut the
    scores and reference points for the Tasmanian HS.
}
\examples{
print("wait on data sets")
}
