% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{copyto}
\alias{copyto}
\title{copyto copies a vector of files from one scenario directory to another}
\usage{
copyto(prefixdir, fromdir, todir, filelist, makenew = TRUE, verbose = TRUE)
}
\arguments{
\item{prefixdir}{the directory (path) containing the different scenarios}

\item{fromdir}{just the name of the directory from which to source the
files listed in filelist (no path)}

\item{todir}{just the name of the new destination directory (no path)}

\item{filelist}{a vector of filenames to be copied, as character.}

\item{makenew}{if the 'todir' does not exist should it be created using
dir.create? default = TRUE}

\item{verbose}{should details be printed to the console, default=TRUE}
}
\value{
Nothing but it will transfer files and change their names. This
    information will be printed to the console if verbose = TRUE
}
\description{
copyto copies a vector of files (see examples) from one
    scenario's directory to another. If a filename includes the
    name of the scenario, eg controlM15h75.csv, is found in the scenario
    M15h75, and it is to be copied to, say, M15h5, then the filename will
    be changed automatically. Care should be taken when using this
    function as it writes files and potentially new directories to your
    storage drives.
}
\examples{
\dontrun{
 prefixdir <- "c:/Users/User/DropBox/A_codeUse/aMSEUse/scenarios/"
 vectfiles <- c("controlM15h75.csv","saudataM15h75.csv","lf_WZ90-20.csv")
 copyto(prefixdir=prefixdir,fromdir="M15h75,todir="M15h5,filelist=vectfiles)
}
#  fromdir="EG";todir="EGtest"; makenew = TRUE;verbose=TRUE
}
