% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{cpueboxbysau}
\alias{cpueboxbysau}
\title{cpueboxbysau generates boxplots of sau x years to max median cpue}
\usage{
cpueboxbysau(rundir, cpue, glbc, scenes, filen = "", startyr = 0, maxval = 0)
}
\arguments{
\item{rundir}{the directory in which comparisons are being made. It is best
to be a separate directory form any particular scenario.}

\item{cpue}{a list of the cpue arrays from the out$sauout$zonePsau$cpue
arrays from each scenario's out object}

\item{glbc}{a list of the globals objects from each scenario}

\item{scenes}{the names of the different scenarios being compared}

\item{filen}{the filename for saving the plot, the default="", which sends
the plot to the console.}

\item{startyr}{what year to start the search for the maximum median cpue,
default = 0, this means that only the projection years are used.}

\item{maxval}{the maximum y-axis value for the years to max median cpue,
default=0, which means the maximum for each boxplot is taken from the
data for that plot.}
}
\value{
a matrix of the boxplot statistics for the years to max median cpue
    for each scenario and sau considered.
}
\description{
cpueboxbysau is used to generate boxplots of the years taken to
    reach the maximum median cpue for each sau x senario. It also provides
    the summary statistics from the boxplots for each sau.
}
\examples{
print("wait on data sets")
}
\seealso{
{
   \link{getprojyrs}, \link{getprojyraavc},  \link{catchHSPM}
}
}
