% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{cpueHSPM}
\alias{cpueHSPM}
\title{cpueHSPM generates boxplots of cpue HSPMs and output statistics}
\usage{
cpueHSPM(rundir, cpue, glbc, scenes, filen = "", startyr = 0)
}
\arguments{
\item{rundir}{the directory in which comparisons are being made. It is best
to be a separate directory form any particular scenario.}

\item{cpue}{a list of the cpue arrays from the out$sauout$zonePsau$cpue
arrays from each scenario's out object}

\item{glbc}{a list of the globals objects from each scenario}

\item{scenes}{the names of the different scenarios being compared}

\item{filen}{the filename for saving the plot, the default="", which sends
the plot to the console.}

\item{startyr}{what year to start the search for the maximum median cpue,
default = 0, this means that only the projection years are used.}
}
\value{
a list of the boxplot statistics for the three HSPMs x the number of
    scenarios compared
}
\description{
cpueHSPM is used to generate boxplots of the year of the maximum
    median cpue for all projections. The number of years over which the
    aavc is calculated can be modified using the startyr argument. In
    addition to the boxplots, the function outputs the statistics describing
    the statistics from the boxplots.
}
\examples{
print("wait on data sets")
}
\seealso{
{
   \link{getprojyrs}, \link{getprojyraavc},  \link{getprojyrC}
}
}
