% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{ctrlfiletemplate}
\alias{ctrlfiletemplate}
\title{ctrlfiletemplate generates a template input control file}
\usage{
ctrlfiletemplate(indir, filename = "controlsau.csv", devrec = -1)
}
\arguments{
\item{indir}{directory in which to place the control.csv file}

\item{filename}{the name for the generated ctrlfile, a character
string that defaults to controlsau.csv.}

\item{devrec}{what form should the recdevs take? valid values are -1, the
default (= random deviates imposed dependent upon withsigR), 0.0, where
a set of conditioned values between 1980 and 2016 are imposed that
improve the fit to the observed CPUE as inserted by ctrlfiletemplate,
or 1.0, which prepares the recdevs for a round of conditioning ready to
improve the observed fit to cpue.}
}
\value{
invisibly the fill path and name of the control file. More
    importantly, it write a control file template to that directory.
}
\description{
ctrlfiletemplate generates a standardized control file
    template. Generate this and then modify the contents to suit
    the system you are attempting to simulate. Defaults to 100
    replicates. There needs to be as many recdevs as there are conditioning
    catches (in the template = 58 rows. If they are all set to -1,
    the default, then they will have no effect and recdevs will be taken off
    the stock-recruitment curve with random deviates defined by withsigR
    found in the ctrl file. If devrec = 1.0 then the recdevs from 1980 to
    2016 will be set to 1.0 which means there will be no random variation
    and the system is ready to be conditioned on those recdevs to improve
    the predicted CPUE during the historical conditioning period. If
    devrec = 0.0, then the recdevs for 1980 - 2016 will be set to values
    that have already conditioned the model and provides a reasonable fit to
    the observed CPUE trends.
}
\examples{
\dontrun{
 yourdir <- tempdir()
 datafiletemplate(nSAU=8,yourdir,"saudata_test.csv")
 ctrlfiletemplate(yourdir,filename="testctrl.csv",devrec=-1)   #
 control <- readctrlfile(yourdir,"testctrl.csv")
 str(control,max.level=1)
}
}
\seealso{
{
  \link{datafiletemplate}, \link{readsaudatafile}, \link{readctrlfile}
}
}
