% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{ctrlzonetemplate}
\alias{ctrlzonetemplate}
\title{ctrlzonetemplate generates a template input control file}
\usage{
ctrlzonetemplate(indir, filename = "control.csv")
}
\arguments{
\item{indir}{directory in which to place the control.csv file}

\item{filename}{the name for the generated ctrlfile, a character
string that defaults to control.csv.}
}
\value{
invisibly the fill path and name of the control file. More
    importantly, it write a control file template to that directory.
}
\description{
ctrlzonetemplate generates a standardized control file
    template. Generate this and then modify the contents to suit
    the system you are attempting to simulate. Defaults to 10
    replicates, and with
    effectively no variation in recruitment, mature biomass or cpue
    calculations.
}
\examples{
\dontrun{
# changes needed now I have simplified to two files and changed their formats
 yourdir <- tempdir()
 ctrlzonetemplate(yourdir,filename="testctrl.csv")   #
 control <- readctrlfile(yourdir,"testctrl.csv")
 str(control,max.level=1)
}
}
