% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{datafiletemplate}
\alias{datafiletemplate}
\title{datafiletemplate generates a template input datafile}
\usage{
datafiletemplate(glob, dirdata, filename = "tmpdat.csv", verbose = FALSE)
}
\arguments{
\item{glob}{the global constants obtained from the region file}

\item{dirdata}{the directory into which to place the data template}

\item{filename}{the name for the generated datafile, a character
string, defaults to tmpdat.csv}

\item{verbose}{print the filename to the console but default=FALSE}
}
\value{
a standard definition data file, to be read by readdatafile
}
\description{
datafiletemplate generates a standard input datafile
    to use as a template, go in and edit it appropriately to suit
    your own needs. It contains the probability distributions that
    are sampled to provide the necessary biological constants for
    each population.
}
\examples{
\dontrun{
 yourdir <- tempdir()
 datafiletemplate(yourdir,"tmpdat.csv")
 data(region1)
 glb <- region1$globals
 constants <- readdatafile(yourdir,"tmpdat.csv",glb)
 str(constants,max.level=1)
}
}
