% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{defineBlock}
\alias{defineBlock}
\title{defineBlock subdivides defined number of populations into blocks}
\usage{
defineBlock(numblk, blknum, numpop)
}
\arguments{
\item{numblk}{number of blocks in the simulated zone, from 'constants'}

\item{blknum}{number of populations in each block, from 'constants'}

\item{numpop}{the total number of populations in the zone}
}
\value{
a vector of length numpop, indexing which block each pop is in.
}
\description{
subdivides the defined number of populations into blocks
    in the numbers defined in the constant blkpop read into 'constants'.
    It allocates the populations sequentially to numblk blocks; if
    numblk does not divide into numpop exactly, any remainder is put
    into the final block. For block can read SAU = spatial assessment
    unit.
}
\examples{
  nblock <- 4
  blkpop <- c(3,4,8,2)
  blockI <- defineBlock(nblock,blkpop,numpop=17)
  print(blockI)
  cat(length(blockI),sum(blkpop),"\n")
}
