% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{definepops}
\alias{definepops}
\title{definepops makes a vector of parameters for each population}
\usage{
definepops(innblock, inblockI, condDat)
}
\arguments{
\item{innblock}{the number of blocks defined in the zone; scalar}

\item{inblockI}{a vector containng the block index for each population}

\item{condDat}{a matrix with rows containing the PDF parameters for each
of the biological parameters to be sampled and columns reflecting
values for each block of populations}
}
\value{
a matrix with a row for each population  and whose columns are
    the parameters defiing the biological parameters for each population.
}
\description{
definepops makes a vector of parameters for each
   population. Each population will have slightly different biological
   properties sampled from probability density distributions defined in
   bConst; a matrix read in by 'parsefile'. In the non-library version of
   the software the matrix bConst was assumed to be global.
}
\examples{
\dontrun{
 data(bConst)
 nblock <- 5
 blkpop <- c(4,9,6,4,2)
 numpop <- 25
 blockI <- defineBlock(nblock,blkpop,numpop)
 popdefs <- definepops(nblock,blockI,bConst)
 print(popdefs)

 nblock <- 3
 filename <- datafileTemplate(numblock=nblock,filename="block3.csv")
 condDat <- readdataFile(filename)
 bConst <- condDat$constants
 blkpop <- condDat$blkpop
 numpop <- condDat$numpop
 blockI <- defineBlock(nblock,blkpop,numpop)
 popdefs <- definepops(nblock,blockI,bConst)
 print(popdefs)
}      # innblock=nblock; inblockI=blockI; condDat=condDat
}
