% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{diagnostics}
\alias{diagnostics}
\title{diagnostics generates an array of diagnostic plots}
\usage{
diagnostics(inzone, plot = TRUE)
}
\arguments{
\item{inzone}{the simulated zone}

\item{plot}{plot the output or not; default = TRUE}
}
\value{
invisibly returns the unique values of LML
}
\description{
diagnostics can generates an array of diagnostic plots,
    however, currently it only plots the selectivity
}
\examples{
\dontrun{
  print("stiull to devise an example.")
}
}
