% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{diagnosticsproj}
\alias{diagnosticsproj}
\title{diagnosticsproj plots a series of diagnostics to DiagProj}
\usage{
diagnosticsproj(zonePsau, glb, rundir, nrep = 3)
}
\arguments{
\item{zonePsau}{the SAU scale object containing the dynamics results}

\item{glb}{the global constants object}

\item{rundir}{the rundir for the scenario}

\item{nrep}{the number of replicate trajectories to plot; default=3}
}
\value{
Nothing but it does add some plots to rundir
}
\description{
diagnosticsproj provides a series of plots and results that
    illustrate the properties of the projections. These include the
    residuals between SAU actual catches and their predicted catches. But
    also a series of plots of the projections for only nrep trajectories to
    illustrate that the dynamic variables are changing through time in a
    plausible or 'realistic' manner.
}
\examples{
print("wait on suitable data sets")
}
