% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docomparison.R
\name{do_comparison}
\alias{do_comparison}
\title{do_comparison is a wrapper function that compares scenarios}
\usage{
do_comparison(
  rundir,
  postfixdir,
  outdir,
  files,
  pickfiles,
  verbose = TRUE,
  intensity = 100,
  zero = FALSE,
  Q90 = TRUE,
  altscenes = NULL,
  juris = "",
  ribbonleg = "topleft",
  scencol = NULL,
  resol = 300,
  ...
)
}
\arguments{
\item{rundir}{the complete path to the directory into which all results and
plots from the comparisons will be placed.}

\item{postfixdir}{the name of the final sub-directory into which the results
will be placed. This is used as the name of the website generated to
display the results}

\item{outdir}{the full path to the directory holding all the required .RData
files. If set to '' then files is assumed to include the full path}

\item{files}{the vector of RData file names to be used. If outdir set to ''
then files is assumed to include the full path as well as the file name}

\item{pickfiles}{a vector of indices selecting the files to be used.}

\item{verbose}{should progress updates be made to the console, default=TRUE}

\item{intensity}{is the density of the rgb colours used in the ribbon plots.
the default = 100}

\item{zero}{should the phase plots have an origin at zero, default=FALSE}

\item{Q90}{default = TRUE which means the inner 90th quantiles will be
plotted. If FALSE then the 95th quantiles will be used.}

\item{altscenes}{a vector of alternative scenario names for use in plots and
tables where scenario names are very long and confuse results. default=
NULL, which implies using the base scenario names (runlabels).}

\item{juris}{which jurisdiction is being examined? default="". This is
an argument so that if there are jurisdiction specific functions for
examining outputs that may be of particular interest to a particular
jurisdiction a suitable function can be written and input, along with
any extra arguments it might have, in the final ellipsis.}

\item{ribbonleg}{location of the legend in the ribon plots, default=topleft}

\item{scencol}{what sequence of colours should each scenario have. The
default=NULL which implies the colour will reflect the sequence in
which they are read in.}

\item{resol}{default = 300, what resolution for the plots.}

\item{...}{the ellipsis is here in case a jurisdiction specific function or
functions is/are written to perform extra analyses, plots, and tables.}
}
\value{
nothing but it does conduct a comparison of at least two scenarios
    and places tables and plots into a given sub-directory
}
\description{
do_comparison provides a simplified interface for when making
    comparisons between multiple scenarios. It uses the saved .RData files
    from each scenario and it is up to the user to put those filenames into
    a vector of characters. The juris argument is an attempt at generalizing
    this comparison function and will allow for different jurisdictions to
    enable future additions to do_comparison that will suit their specific
    requirements. Currently, the only option that has been implemented is
    for juris='TAS', but others can be added now that the basic framework
    is implemented.
}
\examples{
\dontrun{
  suppressPackageStartupMessages({
  library(aMSE)
  library(TasHS)
  library(codeutils)
  library(hplot)
  library(makehtml)
  library(knitr)
  })
  dropdir <- getDBdir()
  prefixdir <- paste0(dropdir,"A_codeUse/aMSEUse/scenarios/")
  postfixdir <- "BC_compare"
  rundir <- filenametopath(prefixdir,postfixdir)
  # normally one would use code to select the files, outdir is used if all
  # RData files are in one directory, otherwise set outdir='' and files
  # should contain the full paths as well as the filenames
  files=c("BC.RData","BCmeta.RData","BC541.RData")
  do_comparison(rundir,postfixdir,outdir,files,pickfiles=c(1,2,3))
}
}
\seealso{
{
   \link{scenebyvar}, \link{scenebyzone}, \link{getfilestocompare}
}
}
