% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domse.R
\name{do_MSE}
\alias{do_MSE}
\title{do_MSE an encapsulating function to hold the details of a single run}
\usage{
do_MSE(
  rundir,
  controlfile,
  hsargs,
  hcrfun,
  sampleCE,
  sampleFIS,
  sampleNaS,
  getdata,
  calcpopC,
  makeouthcr,
  fleetdyn = NULL,
  scoreplot,
  plotmultflags,
  interimout = "",
  varyrs = 7,
  startyr = 42,
  verbose = FALSE,
  ndiagprojs = 3,
  cutcatchN = 56,
  matureL = c(70, 200),
  wtatL = c(80, 200),
  mincount = 100,
  includeNAS = FALSE,
  depensate = 0,
  kobeRP = c(0.4, 0.2, 0.15),
  nasInterval = 5,
  minsizecomp = c(100, 135),
  uplimH = 0.4,
  incH = 0.005,
  deleteyrs = 0,
  selectyr = 0,
  calcFIS = NULL,
  prepareforfis = NULL,
  makefisproj = NULL
)
}
\arguments{
\item{rundir}{the full path to the directory in which all files relating to a
particular run are to be held.}

\item{controlfile}{the filename of the control file present in rundir
containing information regarding the particular run.}

\item{hsargs}{the constants used to define the workings of the hcr}

\item{hcrfun}{the name of the harvest control rule that is used to
calculate the aspirational catches by SAU or TAC by zone for the
following year}

\item{sampleCE}{a function from jurisdictionHS.R that generates the CPUE
statistics from the projected data}

\item{sampleFIS}{a function from jurisdictionHS.R that generates the FIS
statistics}

\item{sampleNaS}{a function from jurisdictionHS.R that generates the
Numbers-at-size samples}

\item{getdata}{a function that gathers all the data required by the hcrfun
and combines it into an hcrdata object ready for the hcrfun}

\item{calcpopC}{a function that takes the output from hcrfun (either the
aspirational catch x SAU or TAC x zone) and generates the actual catch
per population in each SAU expected in the current year.}

\item{makeouthcr}{Another JurisdictionHS.R function that generates (or not)
an object that is continually updated by the hcrfun. The intent is that
it will contain all the HS score values..}

\item{fleetdyn}{a function that calculates the distribution of catch across
the sau and populations. Currently not needed by Tas but needed by SA}

\item{scoreplot}{the function that plots the hcr scores and final scores. In
Tas = plotfinalscores, in SA = SAplotfinalscores. If more inputs than
outhcr, zoneDP, sau, and filen=filename are needed then define your own
function using the ellipsis at the end of the argument list.}

\item{plotmultflags}{the function that plots the TAC/acatch multipliers, and
meta rule flags. In Tas = plotmultandflags. If more inputs than
outhcr, zoneDP, sau, and filen=filename are needed then define your own
function using the ellipsis at the end of the argument list.}

\item{interimout}{should results be saved after projections have finished?
the default="", which means nothing is saved if no risk of crashing in
subsequent analyses. But if trying new stuff which may waste lots of time
if the projections need to be repeated, then set
interimout = outdir, which should be a character string identifying the
fullpath to a directory, which is usually defined in the run_aMSE you are
using, and that will save the main objects needed for later analysis into
the directory used to save final results. Importantly, if your simulation
runs succeed there is no need to save the interimout results. Using
interimout automatically saves the numbers-at-size so, if you want the
numbers-at-size and your simulations work, then turn off interimout, ie
set it = "", and set includeNAS=TRUE.}

\item{varyrs}{how many years at the end of the conditioning on the fishery,
data into zoneDD, to which to add recruitment variation, default = 7,
which suits the Tasmanian west coast. Used in prepareprojection}

\item{startyr}{the index for the first year of the conditioning dynamics to
include in the plotted cpue trajectories, to give startyr:indexoflastyear.
used in sauplots.}

\item{verbose}{Should progress comments be printed to console, default=FALSE}

\item{ndiagprojs}{the number of replicate trajectories to plot in the
diagnostics tab to illustrate ndiagprojs trajectories to ensure that
such projections appear realistic; default=3}

\item{cutcatchN}{to reduce the size of the final array of numbers-at-size
in the catch one can remove all the empty cells below a given size
class. In the default there are 105 2mm size classes and setting
cutcatchN = 56 removes size classes 1:55 2 - 110 mm; only from catchN}

\item{matureL}{is a vector of 2, default = c(70,200), that define the x-axis
bounds on the biology maturity-at-Length plots}

\item{wtatL}{is a vector of 2, default = c(80,200), that define the x-axis
bounds on the biology weight-at-Length plots}

\item{mincount}{given size-composition data what minimum sample size will
be deemed acceptable for inclusion in the plots and conditioning.
default=100.}

\item{includeNAS}{should the NAS projections be included in the output. If
they are then the final output size is greatly increased. default=FALSE}

\item{depensate}{should depensation occur. If > 0, eg depensate=0.2, then
depensation will occur when depletion is < whatever value one puts in for
depensate. Implemented inside the function 'oneyearrec'}

\item{kobeRP}{reference points for the eHS pseudo kobe plot. A vector of the
targdepl, default=0.4, limdepl, default=0.2, and the limH, default=0.15}

\item{nasInterval}{default = 5, which selects the year just before
projections and then the size-composition data for every such interval
in years along the projection period. The years selected always include
the last year of projection and the year just before projections started.
Used by predsizecomp function inside do_MSE.}

\item{minsizecomp}{a vector of two values, the first being the minimum size-
class to be used when plotting the predicted size-composition of the
stock, and the second being the minimum size-class when plotting the
numbers-at-size in the catch. The default=c(100,135)}

\item{uplimH}{defines the upper limit of harvest rate used when estimating
the productivity (also important when initial depletion is not 1.0). The
default = 0.4}

\item{incH}{defines the interval between H steps when estimating productivity
default = 0.005}

\item{deleteyrs}{default = 0, meaning delete no years from the sizecomp data.
if there are years that are to be removed then this should be a matrix
of years to delete vs sau, ie years as rows and sau as columns. All
sau need to be included. to complete the matrix 0 values can be used.}

\item{selectyr}{which year's LML should be used to estimate the LML. This
should be a year index (eg in Tas 1-58). If set = 0, then the LML for
the glb$hyrs + glb$pyrs, the last year of projections data will be used.
default = 0}

\item{calcFIS}{default = NULL. jurisdictional function required to calculate
the annual fisindex, only used when FIS data available}

\item{prepareforfis}{default = NULL. jurisdictional function required to
prepare for using FIS data, only used when FIS data available}

\item{makefisproj}{default = NULL. jurisdictional function required to
prepare the constants needed for using FIS data in theprojections, only
used when FIS data available.}
}
\value{
a vary large list containing tottime, projtime, starttime, glb, ctrl,
    zoneDD, zoneDP, projC, condC, sauout, and outzone
}
\description{
do_MSE is a form of meta-function that contains the code and
    function calls that constitute a single scenario run for aMSE. This is
    the reason the argument list is as long as it is. The code has been
    encapsulated like this to simplify development, maintenance, and
    distribution. Already with the number of variant conditioning examples
    produced as tests of the code-base, errors have arisen through a failure
    to propagate all changes to the code now in this function to all the
    separate MSE run files used to manage each scenario. A potential flaw
    lies with the need to apply the jurisdictionHS functions to any
    historical fishery data used to condition the operating model, ready for
    the first year of the projections. Currently, the code used is unique to
    the Tasmanian case and this obviously still requires generalization.
    This may require a new function to be produced in each jurisdictionHS.R
    file, but that is still under consideration. Part of the development of
    this meta-function will be to further articulate the generation of
    results, and the inclusion of error traps on the argument list entries.
    To aid in the conditioning the option of not calculating the productivity
    has been added. By default it will occur and when projecting it will also
    always occur. HSstats is a list that is always saved and contains
    performance statistics for the HS used (currently contains sum5 and sum10
    the cumulative catches to 5 and 10 years into the projections).
}
\examples{
print("wait on suitable data sets in data")
}
\seealso{
{
 \link{makeequilzone}, \link{dohistoricC}, \link{prepareprojection},
 \link{doprojections}, \link{getprojyrC}, \link{oneyearrec}
}
}
