% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{dodepletion}
\alias{dodepletion}
\title{dodepletion resets regionD to an input depletion level}
\usage{
dodepletion(regC, regD, glob, depl, product, len = 15)
}
\arguments{
\item{regC}{the constants components of the simulated region}

\item{regD}{the dynamic components of the simulated region}

\item{glob}{the general global variables}

\item{depl}{the target depletion proportion for the whole region}

\item{product}{the production curve matrix from doproduction}

\item{len}{the number of intervals in the trial harvest rates}
}
\value{
a revised regionD object
}
\description{
dodepletion resets the depletion level of the whole
    regionand does this by searching for the harvest rate that
    leads to the sum of the mature biomass, across populations,
    divided by the sum of the effective B0 across populations is
    as close as possible to the desired depletion level. This means
    the individual populations will likely vary around the target
    depletion, but across the region it will be correct. The
    depletion is measured relative to the effective B0 as that
    takes account of any larval dispersal. This function uses the
    production curve matrix to search for harvest rates that bound
    the target depletion and then re-searches across those bounds
    using len intervals.
}
\examples{
\dontrun{
  data(region1)
  glb <- region1$globals
  data(constants)
  ans <- makeregionC(region1,constants)
  regionC <- ans$regionC
  ans <- makeregion(glb,regionC)
  regionC <- ans$regionC
  regionD <- ans$regionD
  ans2 <- findunfished(regionC,regionD,glb)
  regionC <- ans2$regionC  # region constants
  regionD <- ans2$regionD  # region dynamics
  product <- ans2$product
  regDD <- dodepletion(regionC,regionD,glb,depl=0.3,product)
  sum((regDD$matureB[1,]/sum(regDD$matureB[1,]))*regDD$deplsB[1,])
  mean(regDD$deplsB[1,])
}
}
