% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{doproduction}
\alias{doproduction}
\title{doproduction - estimates production curve for all populations}
\usage{
doproduction(inpop, uplim = 0.4)
}
\arguments{
\item{inpop}{the population to be assessed; an 'abpop'}

\item{uplim}{the upper limit on teh harvest rate to test, defaults to
0.4; a scalar}
}
\value{
a list made up of the MSY, the depletion level at MSY, and a
   matrix with six columns: ExB, MatB, AnnH, Catch, Deplet, and RelCE.
}
\description{
doproduction estimates a surplus production curve for all
   populations. This estimates a zone's production by estimate MSY for
   each population separately and then combining. do.zoneProd is more
   accurate for a zone being fished as a whole. 'doproduction' is
   usually applied sequentially to all populations in a zone when using
   the makeZone function.
}
\examples{
\dontrun{
txt1 <- 'generate a zone and apply do.zoneProd and print the results'
str(txt1)
}
}
