% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{doprojection}
\alias{doprojection}
\title{doprojection applies the Tasmanian MCDA to a simulated zone}
\usage{
doprojection(
  zoneCP,
  zoneDP,
  glob,
  ctrl,
  projyrs,
  applyHS,
  hsargs,
  projpms,
  inityrs,
  ...
)
}
\arguments{
\item{zoneCP}{the modified zoneC object ready for projections}

\item{zoneDP}{the modified zoneD object ready for projections}

\item{glob}{the global variables object}

\item{ctrl}{the ctrl object}

\item{projyrs}{the number of years of projection}

\item{applyHS}{the function describing the harvest strategy to be used}

\item{hsargs}{a list of any arguments used by applyHS}

\item{inityrs}{the number of years kept from the period of conditioning,}

\item{...}{the ellipsis should contain any additional arguments}
}
\value{
the filled in zoneDP list
}
\description{
doprojection applies the Tasmanian MCDA to a simulated zone. It
    is still in need of optimization as 1000 iterations would currently take
    about 20 minutes.
}
\examples{
print("wait on data files")
# zoneCP=zoneCP;zoneDP=zoneDR;glob=glb;ctrl=ctrl;projyrs=projC$projyrs;inityrs=projC$inityrs
# wid=4;targqnt=0.55;pmwts=c(0.65, 0.25,0.1);hcr = c(0.25,0.75,0.8,0.85,0.9,1,1.05,1.1,1.15,1.2)
}
