% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{doprojections}
\alias{doprojections}
\title{doprojections conducts the replicate model runs for Tasmania}
\usage{
doprojections(
  ctrl,
  zoneDP,
  zoneCP,
  glb,
  hcrfun,
  hsargs,
  sampleCE,
  sampleFIS,
  sampleNaS,
  getdata,
  calcpopC,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{ctrl}{the ctrl object from readctrlfile}

\item{zoneDP}{the object used to contain the dynamics from the replicate
model runs}

\item{zoneCP}{the object used to contain the constants for each population
used in model dynamics}

\item{glb}{the object containing the global constants for the given run}

\item{hcrfun}{the name of the harvest control rule that is used to
calculate the multiplier for the previous aspirational catches (possibly
for each SAU but possibly the TAC for the whole zone) so as to
estimate the aspirational catches or TAC or the following year}

\item{hsargs}{the constants used to define the workings of the hcr}

\item{sampleCE}{a function that generates the CPUE statistics}

\item{sampleFIS}{a function that generates the FIS statistics}

\item{sampleNaS}{a function that generates the Numbers-at-size samples}

\item{getdata}{a function that gathers all the data required by the hcrfun
and combines it into an hcrdata object ready for the hcrfun}

\item{calcpopC}{a function that takes the output from hcrfun and generates
the actual catch per population expected in the current year.}

\item{verbose}{should the iterations be counted on the console?}

\item{...}{the ellipsis used in case any of the functions hcrfun, sampleCE,
sampleFIS, sampleNas, and getdata require extra arguments not included
in the default named collection}
}
\value{
a list containing the full dynamics across all years zoneDP, and the
    final output from the HCR as hcrout
}
\description{
doprojections conducts the replicate model runs for
    Tasmania using the mcdahcr and the input hsargs. The model dynamics and
    the numbers-at-size variables have been separated and returned separately
    because the size components are very large and mean saving such objects
    is both slow and space limiting.
}
\examples{
print("wait on suitable internal data sets")
}
\seealso{
{
 \link{oneyearsauC}, \link[makehtml]{make_html}
}
}
