% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{doquantribbon}
\alias{doquantribbon}
\title{doquantribbon generates a ribbon of multiple quantile values}
\usage{
doquantribbon(
  varq,
  varname,
  yrnames,
  scenes,
  q90,
  intens = 127,
  addleg = "bottomright"
)
}
\arguments{
\item{varq}{an array of quantiles with dimensions the 5 quantiles of the
variable in question (2.5, 5, 50, 95 97.5), by years, the number of
years for which quantiles are available}

\item{varname}{The names of the variable being summarized, used as the
y-label on the plot}

\item{yrnames}{the numeric values of the years to be plotted. Used as the
x-axis as well as the x-axis labels}

\item{scenes}{the names given to the different scenarios}

\item{q90}{should the 90th or 95th quantiles be plotted, TRUE = 90th}

\item{intens}{if polys=TRUE then intens signifies the intensity of colour on
a scale of 0 - 255. 127 is about 50 percent dense.}

\item{addleg}{add a legend? default="bottomright"}
}
\value{
nothing but it does add polygonsto a plot
}
\description{
doquantribbon is used when trying to present a summary of multiple
    scenarios applied across a sequence of years (for example in the
    projection years of an MSE). It assumes that for each scenario quantiles
    have been calculated for each year so that the input array if 5 x nyrs x
    nscenes, where nyrs is the number of years and nscenes is the number of
    scenarios. Five quantiles are assumed with values
    probs=c(0.025, 0.05, 0.5, 0.95 0.975), and the option exists of plotting
    either the 90th (0.05 and 0.95) or the 95th (0.025, 0.975). If the input
    quantiles have different values then the 'q90' argument,if TRUE, will
    reference the 2nd and 4th row and the 1st and 5th if FALSE. If polys is
    TRUE then filled transparent polygons will be plotted, with colours in
    the same sequence as the scenarios, alternatively, if polys=FALSE, then
    lines will be plotted instead. This function is currently only used by
    sauribbon.
}
\examples{
print("wait on datasets")
}
\seealso{
{
   \link{sauribbon}, \link{RGB}
}
}
