% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{dosau}
\alias{dosau}
\title{dosau plots the conditioning history for the dynamics}
\usage{
dosau(inzone, glb, picksau, histCE, yrnames, extra = FALSE)
}
\arguments{
\item{inzone}{the conditioned zone (zoneDD) after it has been converted to
sau scale by using getsauzone}

\item{glb}{the global constants object}

\item{picksau}{which sau should be plotted}

\item{histCE}{the historical cpue series from condC}

\item{yrnames}{the years of historical catches eg 1973:2019}

\item{extra}{should exploitable B and depleB be plotted? default=FALSE}
}
\value{
nothing but it does generate a plot
}
\description{
dosau plots the deplsB, cpue, matureB, catch, harvestR, and
    recruits for a given SAU during the years of conditioning. This aims to
    assist the conditioning process by illustrating the state of the sau
    during and at the end of conditioning on the fishery. If extra is TRUE
    then it also plots
}
\examples{
print("wait on suitable data-sets")
}
