% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{dosauplot}
\alias{dosauplot}
\title{dosauplot generates the plot of the chosen variable for each sau}
\usage{
dosauplot(
  ylabel,
  postrep,
  glb,
  startyr,
  addCI = FALSE,
  CIprobs = c(0.05, 0.5, 0.95),
  histCE = NULL,
  addCE = FALSE
)
}
\arguments{
\item{ylabel}{the variable name to be plotted, used as a y-axis label}

\item{postrep}{the replicate values from the replicated projection zoneDP for
the variable selected taken from the output of 'zonetosau'.}

\item{glb}{the global constants object}

\item{startyr}{the index for the first year of the conditioning dynamics to
include}

\item{addCI}{should quantile confidence bounds be included on the plots,
default=FALSE}

\item{CIprobs}{the quantiles used to generate the confidence intervals. The
default = c(0.05,0.5,0.95)}

\item{histCE}{the historical cpue data if included, default = NULL}

\item{addCE}{should historical cpue be included on the plots, default=FALSE}
}
\value{
invisibly, a list of CI and median for each SAU
}
\description{
dosauplot takes the results of the prepareproject and the
    projection function, it combines the trajectories across the years of
    conditioning and the projection years, for all replicates, and plots all
    replicates for the years from 'startyr' to the final year of the
    projections. It includes the median and inner 90% CI, and, for the CPUE
    plot, if the historical conditioning CPUE is included in histCE, it also
    includes the original cpue series. Beware of trying to plot the
    historical cpue in years where there are no historical cpue data,
    although there are error captures at work to limit the pot to available.
}
\examples{
print("wait on suitable built in data sets")
}
