% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{doTASprojections}
\alias{doTASprojections}
\title{doTASprojections conducts the replicate model runs for Tasmania}
\usage{
doTASprojections(ctrl, zoneDP, zoneCP, histCE, glb, mcdahcr, hsargs)
}
\arguments{
\item{ctrl}{the ctrl object from readctrlfile}

\item{zoneDP}{the object used to contain the dynamics from the replicate
model runs}

\item{zoneCP}{the object used to contain the constants for each population
used in model dynamics}

\item{histCE}{the historical CPUE data used to condition the model prior to
running the replicates. This is appended to new predicted cpue data from
each replicate to calibrate the HCR within the current HS}

\item{glb}{the object containing the global constants for the given run}

\item{mcdahcr}{the name of the harvest control rule that is used to
calculate the multiplier for the previous aspirational catches so as to
estimate the aspirational catches for the following year}

\item{hsargs}{the constants used to define the workings of the hcr}
}
\value{
a replacement for zoneDP containing the dynamics of all replicates
    for all projection years
}
\description{
doTASprojections conducts the replicate model runs for
    Tasmania using the mcdahcr and the input hsargs
}
\examples{
print("wait on suitable internal data sets")
}
