% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_alone.R
\name{draftnumbersatsize}
\alias{draftnumbersatsize}
\title{draftnumbersatsize plots details of initial numbers-at-size}
\usage{
draftnumbersatsize(rundir, glb, Nt, ssc = 5)
}
\arguments{
\item{rundir}{the results directory output files will be sent here}

\item{glb}{the globals list}

\item{Nt}{the numbers-at-size at equilibrium = zoneD$Nt[,1,]}

\item{ssc}{index for starting size class. thus 1 = 2, 2 = 4, 5 = 10, etc.
default = 5 for it plots size classes from 10mm up}
}
\value{
nothing but it does add numpops + 1 plots to the results directory
}
\description{
draftnumbersatsize plots up the initial unfished numbers-
    at-size distribution, including the totals for each sau, and the
    numbers-at-size for each population within each sau.
}
\examples{
print("this will be quite long when I get to it")
}
