% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{driftrec}
\alias{driftrec}
\title{driftrec adjusts the recruitment allowing for larval drift}
\usage{
driftrec(recs, recp)
}
\arguments{
\item{recs}{the vector of recruits calculated internally to oneyear}

\item{recp}{the recruitment proporiton lost. Either a constant (eg
0.02 for 2 percent) or a vector, a value for each population.}
}
\value{
a vector of revised recruitment levels accounting for
    larval drift.
}
\description{
driftrec adjusts the recruitment levels in the stock
    dynamics allowing for larval drift. The input drift levels can
    be a constant for each population, or a vector, with a value
    for each population if their rates are assumed to different. If
    a constant valuer is input it is extended to a vector of the
    same length as the number of populations. Each population is
    assumed to connect only to the populations either side of it.
    The populations on the end only connect into the simulated
    populations and so only lose half the recp value for that
    population, which is sent to the population it is beside in the
    conditioning. If the zone is not conditioned using yield or
    productivity from spatial data then productivity will be
    allocated randomly and the recp (recruitment proportion lost)
    may as well be set to zero.
}
\examples{
  recs <- c(243699,49747,1285137,1492742,923282,273427)
  newrec <- driftrec(recs,0.04)
  newrec0 <- driftrec(recs,0.0)
  recs
  newrec
  newrec0
  sum(recs)
  sum(newrec)
  sum(newrec0)
}
