% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{fillzoneDef}
\alias{fillzoneDef}
\title{fillzoneDef characterizes simulated Zone; holds zone properties}
\usage{
fillzoneDef(inzone, prod)
}
\arguments{
\item{inzone}{the zone to be characterized.}

\item{prod}{the production object from doproduction}
}
\value{
A list of 10 objects containing the properties of the zone;
    has the class zoneDefinition - for S3 methods

\itemize{
  \item Production the productivity of each populations separately
  \item defpop each population's defining parameters
  \item Struct the zone's hierarchical structure
  \item nBlock the number of blocks
  \item numpop the number of populations
  \item SummaryPop biological properties of populations
  \item SummaryZone summary of the zone's properties
  \item BlockProp summary of the block properties
  \item Date date/time of fillzoneDef activation; if the function is run
     soon after a zone's creation this can be a proxy for the zone
     creation.
}
}
\description{
fillzoneDef Characterizes the simulated Zone; holds the
    zone properties. Lists the date and time the zone was made, its size
    and basic biological properties: MSY, B0, Average bLML, Average SaM,
    and more details
}
\examples{
# needs summaryPop and summaryZone
txt2 <- 'always use examples rather than example'
}
