% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{finalcondyeardepletion}
\alias{finalcondyeardepletion}
\title{finalcondyeardepletion plots and tabulates final depletion after conditioning}
\usage{
finalcondyeardepletion(rundir, sauzone, glb, deplvar = "sB", console = TRUE)
}
\arguments{
\item{rundir}{the scenario's directory}

\item{sauzone}{the output from zonetosau, found in sauplots. This is
the projection results summarized into their respective sau rather than
the individual populations}

\item{glb}{the globals object}

\item{deplvar}{which depletion variable to use, either 'sB' or 'eB' are the
only options. Any other input will use 'sB'}

\item{console}{should the plot go to the console or be saved into rundir?
default=TRUE}
}
\value{
invisibly returns the matrix of quantiles for each sau
}
\description{
finalcondyeardepletion generates histograms and tabulations of
    either the spawning biomass or exploitable biomass depletion levels in the
    final year of conditioning. This includes the varyrs of recruitment
    variation prior to projections. It plots a histogram for each sau and
    generates a table of quantiles at probs of 0, 0.05, 0.5, 0.95, and 1.0.
}
\examples{
print("wait on data sets")
}
\seealso{
{
   \link{zonetosau}, \link{sauplots}
}
}
