% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{findF1}
\alias{findF1}
\title{findF1 approximates the F0.1 harvest rate for each population}
\usage{
findF1(product)
}
\arguments{
\item{product}{the 3D array from doproduction}
}
\value{
either the index within the production array or the vector
    of gradients
}
\description{
findF1 generates estimates of the annual harvest rate
    that would be equivalent to the F0.1 for each population. It
    does this by calculating the gradient of the production curve
    for the selected pop number and selecting the best approximation
    by searching for the gradient that is cloest to 0.1.
}
\examples{
data(product) #for pop=1, a 28percent drop from Hmsy leads to a
findF1(product=product) # loss of 3 tonnes, 4 percent of MSY
findmsy(product)  # compare the AnnH, Deplet, and RelCE levels.
}
