% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{findFs}
\alias{findFs}
\title{findFs uses an iterative strategy for finding each year's F values}
\usage{
findFs(cyr, Nyr, sel, aaw, M, reps = 8, wtdiv = 1000)
}
\arguments{
\item{cyr}{a vector of known catches in a given year for nfleets}

\item{Nyr}{the numbers at size at the start of the given year or end of the
year before}

\item{sel}{a matrix of the selectivity of the fishing gears}

\item{aaw}{the weight-at-age or weight-at-size}

\item{M}{the instantaneous natural mortality rate}

\item{reps}{how many internal loops to use finding each F, default = 8}

\item{wtdiv}{divisor for aaw to scale wata to tonnes of catch, default=1000}
}
\value{
the fully selected fishing mortality rates
}
\description{
findFs is used when searching for the instantaneous F that will
    produce the required catch in a given year in an age- or size-structured
    model when there are multiple gears, with different selectivity being
    used. The outcome when applying instantaneous fishing mortality rates
    across each year will be different from when one applies annual harvest
    rates. Using instantaneous rates implies that fishing mortality occurs
    continuously throughout each year, and annual harvest rates imply that
    all fishing occurs in the middle of the year. With abalone these
    assumptions also include with instantaneous rates the numbers-at-size in
    the population undergo somatic growth before fishing occurs, then natural
    and fishing mortality occurs through each year. With harvest rates we
    have growth plus half of natural mortality occurring before fishing,
    during which no natural mortality occurs, the fishing is followed by the
    remaining half of the natural mortality. These different implications
    can lead to real differences. With abalone, neither set of assumptions
    best matches the fishery.
}
\examples{
print("wait on example data sets")
# cyr=obsC;Nyr=Nt[,yr-1];sel=as.matrix(sel[,pickft]);aaw=aaw;M=M;reps=reps
# cyr=incat;Nyr=NumNe;sel=as.matrix(selyr);aaw=pop$WtL;M=pop$Me;reps=8;
# wtdiv=1e6
}
