% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{findlinenumber}
\alias{findlinenumber}
\title{findlinenumber prints out the contents of a text file with line numbers}
\usage{
findlinenumber(rundir, filename, inc = 20)
}
\arguments{
\item{rundir}{the directory path in which to find the text file.}

\item{filename}{the full name of the text file in quotations.}

\item{inc}{the number of lines on screen}
}
\value{
nothing, but it does print the contents of a text file with the
    respective line numbers on the console.
}
\description{
findlinenumber solves the problem of finding the line number in
    a given text file when one wants to change a specific line using the
    function changeline. After inc lines are printed the function stops and
    waits for any character to be input (a space will suffice) before printing
    the next inc lines
}
\examples{
print("wait on a suitable example")
}
