% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{findmsy}
\alias{findmsy}
\title{findmsy identifies the closest productivity value to MSY}
\usage{
findmsy(product)
}
\arguments{
\item{product}{array of productivity values output from doproduction}
}
\value{
a matrix of numpop rows containing the approximate MSY
    related productivity values
}
\description{
findmsy for each population in the region, identifies
    the closest productivity value to MSY and returns the vector
    of productivity values for the selected harvest rate. The Catch
    in each populaiton = MSY and the other variables relate to the
    value at MSY
}
\examples{
print("wait on data sets")
}
