% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{fishery_plots}
\alias{fishery_plots}
\title{fishery_plots generates a set of plots relating to the fishery properties}
\usage{
fishery_plots(rundir, glb, select, histyr, projLML, rge = 50:90)
}
\arguments{
\item{rundir}{the rundir for the particular scenario}

\item{glb}{the object contianing the global constants}

\item{select}{the selectivity matrix from zoneCP, with all years of selectivity}

\item{histyr}{the object from condC containing the historical LML}

\item{projLML}{the object containing the LML in the projections}

\item{rge}{the range of the midpts for the size classes so that the view can
be limited so that the lower arm set to zero and the upper set to 1 need
not all be viewed and to give greater separation to any set of LML, which
will usually be close to each other.}
}
\value{
invisibly, the matrix of unique selectivities
}
\description{
fishery_plots produces a series of plots relating to the fishery
    and its properties. The selectivity curves reflect the LML through time
    and so this needs to be represented.
}
\examples{
print("wait on suitable internal data sets")
}
