% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{getabdevs}
\alias{getabdevs}
\title{getabdevs absolute deviations between a vector and its moving average}
\usage{
getabdevs(x, years, smooth = TRUE, n = 7)
}
\arguments{
\item{x}{a vector of numbers to be tested, perhaps a single replicate
projection}

\item{years}{the loess or moving average is relative to a set of years or
some other sequence. It must hve the same length as x}

\item{smooth}{default = TRUE, which implies a standard loess curve will be
fitted to the curve before finding the deviations. If set FALSE, then
a moving average of length n will be used.}

\item{n}{default = 7, the length of a moving average if one is used.}
}
\value{
a list of the avdev and sumdev across the vector
}
\description{
getabdevs is used to calculate the absolute deviations between
    a vector of numbers and either a loess fit to the vector or a moving
    average across that vector. This provides a measure of variation akin
    to a Average Annual Variation (aav) suitable for oscillating projections.
    This function can be used with apply to summarize across replicates.
}
\examples{
x <- runif(98,min=1,max=25)
outdev <- getabdevs(x,1:98)
}
