% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getavrec}
\alias{getavrec}
\title{getavrec pulls out just the AvRec values for each sau for a scenario}
\usage{
getavrec(rundir, datafile, nsau)
}
\arguments{
\item{rundir}{the directory in which one finds the saudata file for the
given scenario}

\item{datafile}{the exact name of the saudata file}

\item{nsau}{the number of SAU to be found}
}
\value{
a numeric vector of the average recruitment for each SAU
}
\description{
getavrec extracts the AvRec values for each SAU for a given
    scenario. It does this by reading in the saudata file and using grep
    to search for the correct line and returning the line as is. Care is
    required to only take the first record of 'AvRec' so as not to
    consider the variability in sAvRec.
}
\examples{
\dontrun{
  rundir <- "c:/Users/User/DropBox/A_codeUse/aMSEUse/scenarios/MhLML/M1h5/"
  getavrec(rundir,"saudataM1h5.csv",8)
}
}
