% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{getcompout}
\alias{getcompout}
\title{getcompout extracts a component by scenario from output of do_comparison}
\usage{
getcompout(dyn, glb, scenes, pickvar = "matureB", projonly = TRUE)
}
\arguments{
\item{dyn}{a list of the dynamics objects from each scenbario}

\item{glb}{the globals object. The assumption is that all glb objects are
very similar between scenarios, although different reps can be compared}

\item{scenes}{the names of the scenarios selected for comparison}

\item{pickvar}{the name of the variable of interest, must be in quotes.}

\item{projonly}{should only the projection years be used, default=TRUE}
}
\value{
a list of the selected variable for each scenario. Each list
    component will be a 3-D array of years x sau x replicate
}
\description{
getcompout is used to extract a given variable from the dynamics
    included in the output from the do_comparison function. It first isolates
    the 'dyn' component from the 'result' object, then extracts and collates
    the selected variable. There are ten valid variable names that this
    function will accept: matureB, exploitB, midyexpB, catch, acatch,
    harvestR, cpue, recruit, deplsB, and depleB. Anything else will throw an
    error message. The outputs can be used by plotdynphase.
}
\examples{
print("wait on example data")
}
\seealso{
[\link{plotdynphase} for how this may be used]
}
