% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_alone.R
\name{getdraftpops}
\alias{getdraftpops}
\title{getdraftpops generates a draft copy of the population properties}
\usage{
getdraftpops(rundir, ctrlfile, verbose = TRUE, uplim = 0.35, incH = 0.005)
}
\arguments{
\item{rundir}{the directory containing the control csv files. It
can/will also act to store results in a manner that will allow them
to be displayed using makehtml.}

\item{ctrlfile}{the main file that controls the particular run. It contains
the name of the data file that is used to biologically condition the
numpop populations}

\item{verbose}{Should progress comments be printed to console, default=TRUE}

\item{uplim}{defines the upper limit of harvest used when estimating the
productivity (also important when initial depletion is not 1.0). The
default = 0.35}

\item{incH}{defines the interval between H steps when estimating productivity
default = 0.005}
}
\value{
a list of the draft population properties needed by makezoneC,
    an object containing zoneC, zoneD, product, and glb, and the condC
}
\description{
getdraftpops is used to generate a first draft of the population
    properties from the sau data using a standard sau based control and data
    file. In am sau based run of aMSE the input properties for each sau are
    expanded in a random manner across the populations, all that is except
    the AvRec, which is determined by the use of proprec in the datafile.
    However, all other properties are assigned randomly. The outcome is
    characterized in terms of each population's productivity, which can then
    be edited to more closely sorrespond to the productivity as observed in
    the GPS data.
}
\examples{
print("wait on example") # verbose=TRUE;uplim=0.35;incH=0.005
}
