% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getdynamics}
\alias{getdynamics}
\title{getdynamics extracts the sau dynamics for a sau}
\usage{
getdynamics(zoneC, zoneDD, condC, glb, sau = 1)
}
\arguments{
\item{zoneC}{the constants object for all populations}

\item{zoneDD}{the dynamic object for all populations}

\item{condC}{the conditioning fishery data includeding the catches and cpue}

\item{glb}{the globals object}

\item{sau}{which sau to extract, default = 1}
}
\value{
amatrix of the dynamics for the given sau
}
\description{
getdynamics assists with conditioning by generating the sauZone
    and extracting the fishery dynamics for the historical conditioning
    period ready for printing.
}
\examples{
print("wait on data sets")
}
