% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getgrad1}
\alias{getgrad1}
\title{getgrad1 calculates the one year gradient score}
\usage{
getgrad1(vectce)
}
\arguments{
\item{vectce}{vector of cpue for a given spatial scale}

\item{maxGradient}{the scaling factor used in a particular zone}
}
\value{
a vector of gradient1 for each year, starting from year 2
}
\description{
getgrad1 calculates the one year gradient score for a
   vector of CPUE. The equation used is [CE(y) / CE(y-1)] - 1,
   which provides the annual proportional change in CPUE.
}
\examples{
data(blockE13)
nyr <- length(blockE13$year)
grad1 <- getgrad1(blockE13$cpue)
score1 <- getscore1(grad1)
cbind(blockE13$year[2:nyr],grad1,score1)
}
