% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS-HCR.R
\name{getgradone}
\alias{getgradone}
\title{getgradone calculates the one year gradient score}
\usage{
getgradone(vectce, yr)
}
\arguments{
\item{vectce}{vector of cpue for a given spatial scale}
}
\value{
a vector of gradient1 for each year, starting from year 2
}
\description{
getgradone calculates the one year gradient score for a
   vector of CPUE. The equation used is [CE(y) / CE(y-1)] - 1,
   which provides the annual proportional change in CPUE.
}
\examples{
data(blockE13)
nyr <- length(blockE13$year)
grad1 <- getgrad1(blockE13$cpue,yr = 10)
print(grad1)
}
