% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS-HCR.R
\name{getgradwid}
\alias{getgradwid}
\title{getgradwid returns a single gradient across a cpue vector}
\usage{
getgradwid(vectce, yr, wid = 4)
}
\arguments{
\item{vectce}{the vector of cpue. it must be at least yr long}

\item{yr}{the last yr of the wid long time-series whose gradient is wanted}

\item{wid}{the number of years across which the gradient is to be estimated}
}
\value{
a single gradient
}
\description{
getgradwid calculates the regression gradient of a time-series
    of cpue values of length wid, using the analytical formulae rather than
    lm (for speed)
}
\examples{
vectcpue <- c(0.867,0.885,0.877,1.008,1.169,1.231,1.226,1.237,1.238,1.156)
getgradwid(vectcpue,yr=10)   #  -0.01704731
}
