% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getLFdata}
\alias{getLFdata}
\title{getLFdata reads the observed LF-composition data}
\usage{
getLFdata(rundir, filename)
}
\arguments{
\item{rundir}{the run directory containing the control and data files for a
particular run}

\item{filename}{the complete filename for the length-composition data file}
}
\value{
a list containing the lfs array (lengths x years x sau) and palfs
    matrix of the count of observations in each year for each sau
}
\description{
getLFdata reads the observed LF-composition data, which needs
    to be stored in a csv file with the following format. Each SAU must use
    the same sequence of length-classes as rows, and the same sequence of
    years as columns, even if there are empty rows and columns. In addition,
    the first column must list the length-class and the second column lists
    the SAU. See the 'data-file-description.docx', and the example data files
    to see examples of the required format.
}
\examples{
print("wait on suitable internal data files")
}
