% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docondition.R
\name{getLFlogL}
\alias{getLFlogL}
\title{getLFlogL calculates the multi-nomial log-likelihood for sizecomp data}
\usage{
getLFlogL(catchN, obsLFs, glb, wtsc, sau)
}
\arguments{
\item{catchN}{the predicted numbers-at-size in the catch for all years}

\item{obsLFs}{the observed NAS, after cleaning by the preparesizecomp
function}

\item{glb}{the golbals list}

\item{wtsc}{the weighting given to the size-composition data}

\item{sau}{which sau is the subject of analysis}
}
\value{
the weighted ssq for the filtered observed size-composition data
}
\description{
getLFlogL when comparing the observed size-composition data from
    commercial catches with those rpedicted during the historical fishing
    period one can quantify any differences using the multi-nomial log-
    likelihood. This function calculates this, and the value can then be
    combined with am ssq value from a comparison of the predicted and observed
    CPUE ready to be minimized if conditioning the OM on a fisheries real
    data.  As always, the operatig model is running at a population level
    while data arrives at an SAU level, so this is only an approximate
    approach, but it should improve over using purely the results from the
    sizemod R package.
}
\examples{
print("wait on data sets")
}
