% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getline}
\alias{getline}
\title{getline extracts a vector of numbers from a txt or csv files}
\usage{
getline(rundir, filen, varname, nobs)
}
\arguments{
\item{rundir}{the directory in which one finds the saudata or control file
for the given scenario}

\item{filen}{the exact name of the file being examined}

\item{varname}{what variable name is required? Take care with spelling}

\item{nobs}{how many numbers to return, usually nSAU}
}
\value{
a numeric vector for each SAU or of length nobs
}
\description{
getline can be used to extract the AvRec and MaxCEpars values
    for each SAU for a given scenario. It does this by reading in the
    saudata file and using grep to search for the varname, which gives the
    first linenumber with that name. One can extract from the saudata file
    as well as from the control file. The idea is to use values from that
    line when conditioning the operating model.
}
\examples{
\dontrun{
  rundir <- "c:/Users/User/DropBox/A_codeUse/aMSEUse/scenarios/MhLML/M1h5/"
  getline(rundir,"saudataM1h5.csv","AvRec",8)
}
}
