% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getlistVar}
\alias{getlistVar}
\title{getlistVar Extracts a vector or maxtrix of vectors from zone}
\usage{
getlistVar(inzone, indexVar)
}
\arguments{
\item{inzone}{the zone being explored; a 'zone'}

\item{indexVar}{the name of the variable to be extracted; character}
}
\value{
either a vector or matrix of values depending on the variable
}
\description{
getlistVar Extracts a vector or maxtrix of vectors from
   zone if a vector the names relate to populations, if a matrix the
   columns relate to populations the rows to the years of simulation.
   Only Me, R0, B0, steeph, ExploitB, MatureB, HarvestR, Catch,
   popdef, MSY, LML, bLML, SaM, cpue, Recruit, and ExB0, are
   currently valid choices. The indexVar = popdef would generate a
   listing of all the constants which would obviate the need for Me,
   R0, and steeph.
}
\examples{
\dontrun{
 data(condDat)
 out <- makeZone(condDat)
 zone <- out$zone
 getlistVar(zone,"MSY")
 getlistVar(zone,"B0")
}
}
