% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS-HCR.R
\name{getlmcoef}
\alias{getlmcoef}
\title{getlmcoef is a greatly simplified way of gasining regression coefs}
\usage{
getlmcoef(y, x)
}
\arguments{
\item{y}{the dependent variable}

\item{x}{the independent variable}
}
\value{
a vector of the intercept and gradient, in that order.
}
\description{
getlmcoef is a simplified replacement for the lm function that
    is much faster than lm but only returns the linear regression
    coefficients. This is limited to a simple y ~ x model. When in use be
    sure to give it the values of y required.
}
\examples{
y <- c(1.226,1.237,1.238,1.156)
x <- 1:4
yp <- y/y[1]  # to make the cpue relative to the first in the series
getlmcoef(yp,x)
getgradwid(yp,4)
}
