% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getmaxCE}
\alias{getmaxCE}
\title{getmaxCE identifies peak median CPUE during projections for each SAU}
\usage{
getmaxCE(ceCI, glb, targetCE)
}
\arguments{
\item{ceCI}{is the sauout$outCI$cpue object which contains the CI for the
replicate projections}

\item{glb}{the globals object}

\item{targetCE}{the target CE from the hsargs}
}
\value{
a matrix of nSAU x 5: 5p, 50p, 95p year, and difference between
    the target and the maximum median value, all for each SAU
}
\description{
getmaxCE is one of the HS performance statistics. Because it
    uses the targetCE from the TasmanianHS it is obviously Tasmanian
    specific.
}
\examples{
\dontrun{
  ceCI <- out$sauout$outCI$cpue
  glb <- out$glb
  targCE <- hsargs$maxtarg
  getmaxCE(ceCI=ceCI,glb=glb,targetCE=targCE)
}
}
