% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getnas}
\alias{getnas}
\title{getnas gets the numbers-at-size for all populations and the zone}
\usage{
getnas(zoneD, yr, glob)
}
\arguments{
\item{zoneD}{the dynamic portion of the zone}

\item{yr}{which yr from the range available should be summarized}

\item{glob}{the global variables object}
}
\value{
a matrix of numpop + 1 columns of numbers-at-size for each population
    and for the zone
}
\description{
getnas extracts numbers-at-size for all populations,
    SAUs, and the zone. There will therefore be numpop + 1
    columns of numbers-at-size fr the particular year given.
}
\examples{
print("wait on suitable data sets")
}
\seealso{
{
 \link{getLFdata}, \link{getzoneLF}, \link{prepareprojection},
 \link{doprojections}
}
}
