% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getprojyraavc}
\alias{getprojyraavc}
\title{gerprojyraavc applies the aavc function to the input catches}
\usage{
getprojyraavc(catches, glb)
}
\arguments{
\item{catches}{the three dimensional array of catches from the projection
dynamics, years x sau x replicates}

\item{glb}{the gloabls object for the scenario}
}
\value{
a replicates x nsau matrix of aavc for whatever number of years were
    input to the function, the default is all years.
}
\description{
gerprojyraavc takes in a 3D array of catches (or whatever) and
    applies the aavc function to obtain the average annual variation of
    catches across the second array dimension. The expected dimensions are
    years x sau x replicates. The function uses 'apply' to obtain the aavc
    along the year dimension for every replicate. To contrain the number of
    years analysed one needs to constrain the input array, eg catch[1:10,,]
}
\examples{
print("Wait on data sets")
}
