% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getprojyrC}
\alias{getprojyrC}
\title{getprojyrC returns cumulative catch from selected projection years}
\usage{
getprojyrC(catsau, glb, period = 10)
}
\arguments{
\item{catsau}{the time series of catches summed across populations within
each SAU. This array (reps x SAU x Allyears), is found within the 'out'
object from the do_MSE function (as in out$catsau).}

\item{glb}{the globals object. Again found in 'out', as in out$glb}

\item{period}{how many years to cumulate; default = 10}
}
\value{
a reps x SAU matrix of summed catches
}
\description{
getprojyrC is used to calculate the cumulative catch taken
    in each SAU and across the zone for selected years of each replicate.
    Thus, the output would be a replicates x nSAU matrix of total catches
    across the first 'period' years of the projection under the given
    harvest strategy.
}
\examples{
print("wait on suitable internal data sets")
# catsau=catch[[1]]; glb=glbc[[1]];period=5
}
