% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getprojyrs}
\alias{getprojyrs}
\title{getprojyrs i sused to truncate an array to only the projected year}
\usage{
getprojyrs(sauarr, glb, withlast = TRUE)
}
\arguments{
\item{sauarr}{the projection data summarized to the sau scale. This might be
matureB, exploitB, midyexpB, catch, acatch, harvestR, cpue, recruit,
deplsB, or depleB. CatchN and Nt need different treatment}

\item{glb}{the global object = out$glb}

\item{withlast}{should the last year of observations be included from the
simulation or only the projection years used? default=TRUE}
}
\value{
an array of the input data truncated to the projection years with,
    optionally, just the last year of the observed period included.
}
\description{
getprojyrs is used when calculating the harvest strategy
    performance measures. To do that requires us to focus on the projection
    years, and possibly include the last year of observations for continuity.
}
\examples{
print("wait on data sets")
}
