% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getrateofchange}
\alias{getrateofchange}
\title{getrateofchange estimates annual change rate of a variable in dynamics}
\usage{
getrateofchange(dyn, whichvar, glb)
}
\arguments{
\item{dyn}{the dynamics objects from each scenario collected into a single
dyn object from the output of do_comparison}

\item{whichvar}{which variable within the dyn object to characterize. The
variables it can work with include: matureB, exploitB, midyrexpB, catch,
acatch, harvestR, cpue, recruit, deplsB, and depleB.}

\item{glb}{the globals object needed for the number and names of the sau,
and the years and their names.}
}
\value{
a list of the differences, the percentager differences, and the
    medians
}
\description{
getrateof change
}
\examples{
# syntax  getrateofchange(dyn=dyn,whichvar="catch",glb=glb)
}
