% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docondition.R
\name{getrecdevcolumn}
\alias{getrecdevcolumn}
\title{getrecdevcolumn extracts a column of recdevs from a control file}
\usage{
getrecdevcolumn(rundir, filename, yearrange, sau, verbose = FALSE)
}
\arguments{
\item{rundir}{the rundir for the scenario}

\item{filename}{the character name of the control file}

\item{yearrange}{the range of years of rec devs to be selected}

\item{sau}{which SAU or block is to be fitted?}

\item{verbose}{should details of the run be made. default=FALSE}
}
\value{
a list of the vewctor of recdevs, the vector of the yearrange, and
    a vector of the linenumbers within the control file that are to be
    changed
}
\description{
getrecdevcolumn is used when conditioning the operating model
    by modifying the recruitment deviates. getrecdevcolumn extracts a column
    of years selected from a selected controlfile. It does this to use those
    values as the starting parameter vector for optim.
}
\examples{
print("wait on suitable internal data-sets")
}
