% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getregionLF}
\alias{getregionLF}
\title{getregionLF extracts all LF data from a zone across pops and years}
\usage{
getregionLF(regD, glb)
}
\arguments{
\item{regD}{the dynamic part of the region}

\item{glb}{the global constants}
}
\value{
an Nclass x Nyrs matrix containing LF data across all
    populations by year
}
\description{
getregionLF extracts all LF data from a region across
    all populations for each year. Thus an Nclass x Nyrs X numpop
    matrix is compressed into a Nclass x Nyrs matrix
}
\examples{
\dontrun{
print("An example needs to be written")
}
}
