% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getregionprops}
\alias{getregionprops}
\title{getregionprops extracts the depletion level for a given year}
\usage{
getregionprops(regC, regD, glb, year = 1)
}
\arguments{
\item{regC}{the constants for the region being simulated}

\item{regD}{the dynamic part of the region being simulated}

\item{glb}{the global constants}

\item{year}{which years information is wanted, default=1}
}
\value{
a matrix of population properties with a column of totals
}
\description{
getregionprops extracts a set of properties for each
    population and summarizes them for the region as well. These
    properties include effB0, matureB, legalmatB, propprot, MSY,
    effexB0, exploitB, SpBDepl, ExBDepl, legalDepl, MSYDepl, LML,
    and bLML. See the model documentation for the meaning of each
    of these.
}
\examples{
data(testregC)
data(testregD)
data(region1)
glb <- region1$globals
round(getregionprops(testregC,testregD,glb),4)
}
