% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getsaudata}
\alias{getsaudata}
\title{getsaudata reads in the sau properties data from teh control.csv file}
\usage{
getsaudata(datadir, infile)
}
\arguments{
\item{datadir}{the full path to where the zone's and sau's property datafile
is to be found}

\item{infile}{the name of the datafile, usually zone1$ctrl$datafile, which is
obtained by first running 'readctrlfile'.}
}
\value{
a matrix of properties by sau
}
\description{
getsaudata is used when attempting to condition the aMSE zone by
    conducting formal stock assessments on the data available for single
    sau. We use 'readctrlfile' to obtain the fishery data held in the object
    condC, and use getsaudata to get the growth, maturity, and selectivity
    data required by any size-based stock assessment model. We cannot use
    'readsaudatafile' because that also distributed the properties among the
    designated populations for each sau, and outputs the population
    propoerties. Maybe I should change the sequence, first to read in the
    saudata, and only then distribute those among the populations.
}
\examples{
\dontrun{
  getsaudata(datadir=datadir,infile=ctrl$datafile)
}
}
\seealso{
{
 \link{readctrlfile}
}
}
