% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getsaunas}
\alias{getsaunas}
\title{getsaunas extracts mean number-at-size at the SAU scale}
\usage{
getsaunas(nas, glb)
}
\arguments{
\item{nas}{the 4D array of the zone level catchN or Nt}

\item{glb}{the globals object}
}
\value{
a 3D array of numbers-at-size vs year vs SAU
}
\description{
getsaunas summarizes the mean number-at-size across replciates
    at the SAU scale from the zone scale outputs. The input nas can be
    either the catchN or the population Nt from the out put NAS object.
}
\examples{
print("wait on data")
}
