% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getscore1}
\alias{getscore1}
\title{getscore1 calculates the scores for the grad1 PM}
\usage{
getscore1(grad1)
}
\arguments{
\item{grad1}{the values derived from the function getgrad1}
}
\value{
a vector of scores to be included in the MCDA
}
\description{
getscore1 calculates the scores for the grad1
    performance measure. It does this by re-scaling the range of
    the
}
\examples{
data(blockE13)
nyr <- length(blockE13$year)
grad1 <- getgrad1(blockE13$cpue)
score1 <- getscore1(grad1)
cbind(blockE13$year[2:nyr],grad1,score1)
}
