% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getsingleNum}
\alias{getsingleNum}
\title{getsingleNum find a line of text and extracts a single number}
\usage{
getsingleNum(varname, intxt, context = "")
}
\arguments{
\item{varname}{the name of the variable to get from intxt}

\item{intxt}{text to be parsed, usually obtained using readLines}

\item{context}{the function name within which getsingleNum is
being used, defaults to empty string}
}
\value{
a single number
}
\description{
getsingleNum uses grep to fine an input line. The
    variable 'context' is used to identify within which function
    getsingleNum is being called. This can then be used in a
    the stop function if grep fails to find the 'varname' inside the
    'intxt', which is usually obtained using readLines.
}
\examples{
\dontrun{
 # Not exported, prefix with aMSE:::
 context = "Function Example"
 txtlines <- c("replicates, 100","Some_other_text, 52")
 getsingleNum("replicates",txtlines,context=context)
 getsingleNum("eeplicates",txtlines,context=context)
 getsingleNum("other",txtlines,context=context)
}
}
