% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getsmureg}
\alias{getsmureg}
\title{getsmureg summarizes regionD into MSU and region}
\usage{
getsmureg(regD)
}
\arguments{
\item{regD}{the regionD after nyrs of dynamics}
}
\value{
a list of six matrices of nSMU columns of SMU summaries,
    and one column for the region
}
\description{
getsmureg rowsums the matrices of matureB, exploitB,
    catch, and recruit from the input regD into the separate SMUs
    and the total into the region. The harvestR is simply the
    respective catch divided by the exploitB, and the cpue are the
    individual population cpue weighted relative to the catch taken
    from each population in either the SMU or the complete region.
}
\examples{
print("wait on an example")
}
