% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docondition.R
\name{getssqparts}
\alias{getssqparts}
\title{getssqparts calculates the ssq for the cpue and size-comp data}
\usage{
getssqparts(rundir, controlfile, calcpopC, mincount = 100, wtsc = 0.02)
}
\arguments{
\item{rundir}{the directory in which all the files and outputs are kept}

\item{controlfile}{the name of the controlfile for the MSE run}

\item{calcpopC}{the function from the harvest strategy that is used to
calculate the expected catch by population within each say each year. The
annual catch by sau is known but in the MSE that needs to be distributed,
in this case without error, across each population.}

\item{mincount}{the minimum count used in the size-composition data selection
default = 100, this can be a vector of length nsau with a value for each
sau if you want to give a different value to each sau.}

\item{wtsc}{the weighting given to the size-composition data ssq,
default= 0.02. This can be a vector of length nsau if you want to give
each sau a different weighting.}
}
\value{
a vector of the total ssq, the cpue ssq, and the size-comp ssq
}
\description{
getssqparts is used during the conditioning of the operating
    model to improve the fit of the model parameters to the observed SAU
    scale data on cpue and size-composition. It only uses sum-of squared
    deviations so an arbitrary weight is given to the size-composition data,
    and that value should be such that changes in the ssq for the cpue are of
    the same order of magnitude for the size-composition data.
}
\examples{
print("wait on internal data sets")
}
