% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{getStr}
\alias{getStr}
\title{getStr obtains a string from an input text line}
\usage{
getStr(inline, nb)
}
\arguments{
\item{inline}{input text line}

\item{nb}{number of bits to return}
}
\value{
a vector of character strings
}
\description{
getStr obtains a string from an input text line
}
\examples{
\dontrun{
  print("to be developed" )
}
}
