% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getsum}
\alias{getsum}
\title{getsum sums each of the main dynamics within regionD}
\usage{
getsum(inmat, index)
}
\arguments{
\item{inmat}{what matrix within regionD to sum into SMU and region}

\item{index}{a vector containing an index of populations within SMU}
}
\value{
an nSMU+1 column matrix summarizing each SMU and the region
}
\description{
getsum is used by getsmureg to sum each of the main
    dynamic variables within regionD
}
\examples{
print("wait on an example")
}
