% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getunFished}
\alias{getunFished}
\title{getunFished - extracts all data relating to year 1; unfished.}
\usage{
getunFished(inzone, glb)
}
\arguments{
\item{inzone}{the zone whose unfished state is wanted}

\item{glb}{contains the global variables used everywhere}
}
\value{
a list of multiple components relating to the unfished stock
}
\description{
getunFished - extracts all data relating to year 1;
   unfished. Of course, this will only work as long as movezoneYear
   hasn't been called, which would disrupt the contents of the first
   year. But if applied to zone rather than zone1 etc, this should be
   fine. The outputs include: R0, A0, B0, B0crypt, popdef, MSY,
   MSYDepl, LML, ExB0, Nt, and Ncrypt
}
\examples{
\dontrun{
 data(condDat)
 glb <- condDat$globals
 out <- makeZone(condDat,uplim=0.4)
 zone <- out$zone
 unfish <- getunFished(zone,glb)
 str(unfish,max.level=1)
}
}
