% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getvect}
\alias{getvect}
\title{getvect extracts invar from the popdef vector in zoneC}
\usage{
getvect(zoneC, invar)
}
\arguments{
\item{zoneC}{the constants object for the zone}

\item{invar}{a character variable eg. "steeph", "DLMax"}
}
\value{
a numpop vector of invar from the numpop popdefs in zoneC
}
\description{
getvect extracts a numpop vector of invar from the
    popdef vector in zoneC. Still need to use sapply to pull out
    complete vectors such as popdef or maturity etc.
}
\examples{
data(testzoneC)
getvect(testzoneC,"steeph")
}
