% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getyr2maxce}
\alias{getyr2maxce}
\title{getyr2maxce gives a nsau x reps matrix of years to max(cpue)}
\usage{
getyr2maxce(cpue, glb)
}
\arguments{
\item{cpue}{this is the 3D array of the out$sauout$zonePsau$cpue produced by
each scenario. The out objects can be from a saved RData file from
each scenario. The idea being that each cpue array should be processed in
turn and the matrices of the reuslts processed in whatever function calls
this one.}

\item{glb}{the globals object for the scenario}
}
\value{
a reps x nsau matrix of the year in which the maximum cpue occurs in
    the projections
}
\description{
getyr2maxce uses apply and 'which.closest' from codeutils to
    identify the year of maximum cpue in each replicate for each sau in the
    cpue array out of the projected dynamics.
}
\examples{
\dontrun{
  glb <- out$glb  # alternative might be ans[[i]]$glb
  cpue <- out$sauout$zonePsau$cpue
  ce <- getprojyrs(ce,glb$hyrs,glb$pyrs,startyr=glb$indexCE)
  maxce <- getyr2maxce(ce,glb)
  str1(maxce)
}
}
\seealso{
{
    \link{which.closest}
}
}
