% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getzonechangerate}
\alias{getzonechangerate}
\title{getzonechangerate gets annual change rate of a variable across zones}
\usage{
getzonechangerate(zone, whichvar, glb)
}
\arguments{
\item{zone}{the zone summary objects from each scenario collected into a
single zone object from the output of do_comparison}

\item{whichvar}{which variable within the zone object to characterize. The
variables it can work with include: matureB, exploitB, midyrexpB, catch,
acatch, harvestR, cpue, recruit, deplsB, and depleB.}

\item{glb}{one of the globals object from a scenario the number of
projection years MUST be the same.}
}
\value{
a list of the differences, the percentager differences, and the
    medians
}
\description{
getzonechangerate estimates annual change rate of a variable
    across scenarios by zone. It uses the zone object within docomparison
}
\examples{
# syntax  deltas <- getzonechangerate(zone=zone,whichvar="catch",glb=glb)
}
