% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getzoneLF}
\alias{getzoneLF}
\title{getzoneLF extracts all LF data from a zone across pops and years}
\usage{
getzoneLF(zoneD, glb)
}
\arguments{
\item{zoneD}{the dynamic part of the zone}

\item{glb}{the global constants}
}
\value{
an Nclass x Nyrs matrix containing LF data across all
    populations by year
}
\description{
getzoneLF extracts all LF data from a zone across
    all populations for each year. Thus an Nclass x Nyrs X numpop
    matrix is compressed into a Nclass x Nyrs matrix
}
\examples{
\dontrun{
print("An example needs to be written")
}
}
