% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getzoneprod}
\alias{getzoneprod}
\title{getzoneprod zone scale summary of product matrix from doproduction}
\usage{
getzoneprod(product)
}
\arguments{
\item{product}{The productivity array from doproduction containing the
range of imposed harvest rates, and the resulting outputs for each
population}
}
\value{
a matrix containing the approximate productivity matrix for the zone
}
\description{
getzoneprod takes in the product matrix from doproduction and
    sums the mature and exploitable biomassand the total catches. Then it
    recalculates, for the zone, the approximate annual harvest rate, the
    depletion level, and the relative catch rate. The annual harvest rate and
    relative catch rate are simply the arithmetic mean of the values for all
    populations, which the depletion is the column of mature biomass divided
    by the first value = B0
}
\examples{
data(product)
zoneprod <- getzoneprod(product)
head(zoneprod,20)
}
