% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getzoneProps}
\alias{getzoneProps}
\title{getzoneProps extracts the depletion level for a given year}
\usage{
getzoneProps(regC, glb, year = 1)
}
\arguments{
\item{glb}{the global constants}

\item{year}{which years information is wanted}

\item{inzone}{the zone being simulated}
}
\value{
a list of size objects
}
\description{
getzoneProps extracts the depletion level along with an
    array of other statistics for a given year.
}
\examples{
\dontrun{
 data(condDat)
 glb <- condDat$globals
 out <- makeZone(condDat,uplim=0.4)
 zone <- out$zone
 zprop <- getzoneProps(zone,glb,year=1)
 str(zprop,max.level=1)
}
}
