% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{historicalplots}
\alias{historicalplots}
\title{historicalplots a wrapper function to call plots of fishery plots}
\usage{
historicalplots(rundir, condC, glb, commonscale = FALSE, proportion = FALSE)
}
\arguments{
\item{rundir}{the directory where all results are stored}

\item{condC}{the R object contaiining the historical fishery data generated
by the readctrlfile function}

\item{glb}{the global variables object}

\item{commonscale}{should the SAU plots share a common y-axis scale,
default =  FALSE}

\item{proportion}{should the plot be of proportion relative to the maximum
catch through each time-series or as raw tonnes x year, default=FALSE}
}
\value{
nothing but it does add plot files to the rundir
}
\description{
historicalplots is a wrapper used by do_MSE and do_condition,
    to plot out details of the historical fishery data used in the
    conditioning. It currently plots out the historical catches and CPUE.
}
\examples{
print("wait on suitable data sets")
}
