% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{HSphaseplot}
\alias{HSphaseplot}
\title{HSphaseplot makes a phase plot of predicted mature depletion vs H}
\usage{
HSphaseplot(
  dyn,
  glb,
  sau,
  rundir = "",
  startyr = 1992,
  console = TRUE,
  targdepl = 0.4,
  limdepl = 0.2,
  limH = 0.175
)
}
\arguments{
\item{dyn}{a list of the dynamics for every sau. This function only requires
the mature biomass depletion and the harvest rates}

\item{glb}{the globals object}

\item{sau}{which sau index (1 - 8 in Tasmania) to use?}

\item{rundir}{the directory for all files and results, default=""}

\item{startyr}{default = 1992. Otherwise could use condC$yearCE[1]}

\item{console}{should the plot go to the console or be saved? Default=TRUE}

\item{targdepl}{what target depletion is wanted, default=0.4}

\item{limdepl}{what is the limit depletion reference point, default = 0.2}

\item{limH}{what is the maximum H of limit fishing mortality, default = 0.2
#'}
}
\value{
a list of the median mature biomass depletion and median harvest rate
    by year invisibly, it also plots these vectors.
}
\description{
HSphaseplot generates, for a single sau, a Kobe or phase plot
    of the predicted harvest rate against the predicted mature biomass
    depletion. It highlights the startyr of cpue conditioning (1992 in
    Tasmania) the end of conditioning (usually 2020 in Tasmania), and the
    final year (2050 if projected for 30 years). The default limit on harvest
    rate is set at 0.175, this may not be precautionary for all sau.
}
\examples{
print("wait on a data set")
}
