% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{initiateHS}
\alias{initiateHS}
\title{initiateHS applies HS to the last years of the historical conditioning}
\usage{
initiateHS(zoneDP, zoneCP, inN, acatch, sigmar, sigmab, glb)
}
\arguments{
\item{zoneDP}{the empty dynamics object used in the projections}

\item{zoneCP}{the constant object after modification for projections}

\item{inN}{an array of the numbers-at-length for each population from the
last year of the conditioning on the fishery or historical fishery data}

\item{acatch}{the aspirational catches per SAU derived from the conditioning
of the HS based on the historical or fishery conditioning data}

\item{sigmar}{the recruitment variability (sd) to be used in the projections}

\item{sigmab}{the variability (sd) applied to the exploitable biomass
estimates to introduce a form of implementation error on the distribution
of catches among SAU}

\item{glb}{the globals object}

\item{exb}{a vector of exploitable biomass for each population from the last
year of the conditioning on the fishery or historical fishery data}
}
\value{
a revised zoneDP object with the first year filled in for each
    replicate
}
\description{
initiateHS uses the last years of fishery conditioning to run
    the HS for the first year of the projections. This fills in the first
    row of each of the variables in the dynamic object for each of the
    proposed iterations.
}
\examples{
print("wait on suitable internal data sets")
}
