% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invC}
\alias{invC}
\title{invC matrix inversion using RcppArmadillo}
\usage{
invC(x)
}
\arguments{
\item{x}{the matrix to be inverted}
}
\description{
invC uses BLAS to conduct a matrix inversion and
    appears to be about 73 percent faster than R's solve when used
    in a length-based model context.
}
