% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{lf10}
\alias{lf10}
\title{lf10 contains three years of length-composition data for block 10}
\format{
A data.frame of abalone size-composition data
\describe{
  \item{year}{the year of sampling}
  \item{sau}{the block or SAU}
  \item{length}{the measured size in mm, with 2mm size classes}
  \item{counts}{the counts at each length class}
  \item{propcounts}{the proportion of each length class of the total}
}
}
\source{
Thanks to the Institute of Marine and Antarctic Science,
    which is part of the University of Tasmania, and especially to
    Dr Craig Mundy, leader of the Abalone Group, for permission to use
    this data.
}
\description{
lf10 is a data.frame of length-composition of commercial catch
    from block 10 on the west coast of Tasmania in a longdat format as
    derived from the commlf function makelongdat.This makes it useful for
    illustrating the use of makewidedat.
}
\section{Subjects}{

 \itemize{
   \item selectivity
   \item size-based stock assessment model fitting
   \item size-distributions of commercial catches
 }
 @export
}

\examples{
 data(lf10)
 head(lf10,20)
 mids <- seq(138,210,2)
 answer <- makewidedat(lf10,mids)
 answer[1:20,]
}
