% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{lfs}
\alias{lfs}
\title{lfs is a 3-D array of size-composition data for use in examples}
\description{
lfs is a 38 x 31 x 8 array of values for size-composition data
    from the west coast of Tasmania made up of 38 size classes, 31 years,
    from 1990 - 2020, and 8 statistical blocks. This is purely for use with
    examples within the aMSE R package and requires quality control checks
    before use anywhere else.
}
\examples{
 data(lfs)
 lfs[,10:20,6]
}
