% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{make_html}
\alias{make_html}
\title{make_html create a HTML files to view figures in a browser.}
\usage{
make_html(
  replist = NULL,
  resdir = NULL,
  plotFileTable = "plotFileTable",
  width = 500,
  openfile = TRUE,
  runnotes = NULL,
  verbose = TRUE
)
}
\arguments{
\item{replist}{Object created by an aMSE run}

\item{resdir}{Directory where a particular run's resulst are all held.
It must contain a sub-directory called plots where *.png files
are saved}

\item{plotFileTable}{CSV file with info on PNG files. The resdir/plot
directory will be sequentially searched for filenames included
in this csv file, which includes a figure legend and a category=tab,
which identifies in which tab the plot or table will be placed.}

\item{width}{Width of plots (in pixels).}

\item{openfile}{Automatically open index.html in default browser?}

\item{runnotes}{Add additional notes to home page.}

\item{verbose}{Display more info while running this function?}
}
\description{
make_html writes a set of HTML files with tabbed navigation
    between them, with each tab containing output relating to a particular
    set of results or diagnostics from a particular aMSE run. This code
    was borrowed and extensively modified from Ian Taylors r4ss. By default,
    this function will look in the directory where PNG files were created
    for a plotFileTable_runname.csv file with the name 'runname' written by
    aMSE. HTML files are written to link to these plots and put in the same
    directory.
}
\author{
Originally Ian Taylor, modified by Malcolm Haddon
}
