% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{makeabpop}
\alias{makeabpop}
\title{makeabpop generates full population structure in an unfished state}
\usage{
makeabpop(popparam, midpts, projLML)
}
\arguments{
\item{popparam}{the vector of biological parameters values that define
the specific properties of this population. Obtained from 'popdefs'}

\item{midpts}{the center values of the size classes from
condDat{parseFile}}

\item{projLML}{the LML expected in the projection years 2 - Nyrs;
matrix read in from datafile}
}
\value{
a list of objects as shown in the description.
}
\description{
makeabpop generates the full population structure in an
   unfished state. The structure is pre-determined: MaxDL L50 L95 MaxSig
   Me Mc R0 A0 B0 steeph ExploitB MatureB MatBCypt HarvestR Catch popdef
   MSY MSYDepl LML bLML popq SaM cpue CatchN deplExB Recruit ExB0 G
   Maturity WtL Emergent Select Nemerg SelWt MatWt. See the AbMSE
   documentation to see the full definition of an abpop and the structure
   of a zone
}
\examples{
\dontrun{
 data(condDat)
 bConst <- condDat$constants
 blkpop <- condDat$blkpop
 numpop <- condDat$numpop
 blockI <- defineBlock(nblock,blkpop,numpop)
 popdefs <- definepops(nblock,blockI,bConst)
 pop1 <- makeabpop(popdefs[1,],condDat$midpts,
                   condDat$ProjLML[popdefs[1,20]])
 str(pop1,max.level=1)
}
}
