% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{makecompareoutput}
\alias{makecompareoutput}
\title{makecompareoutput generates HTML files when comnparing scenarios}
\usage{
makecompareoutput(
  rundir,
  glbc,
  scenes,
  postdir,
  filesused,
  openfile = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{rundir}{the full path to the directory holding the all result files}

\item{glbc}{a list of the globals objects from each scenario}

\item{scenes}{a vector of the names for each scenario}

\item{postdir}{the name of the directory holding the results, also used to
name the internal webpage}

\item{filesused}{a vector of the names of the scenarios that were compared,
this would usually be files[pickfiles]. Added to the home page.}

\item{openfile}{should the website be opened automatically? default=TRUE}

\item{verbose}{should details of producing the HTML files be written to the
console?  default=FALSE}
}
\value{
nothing but it does generate a set of HTML files in rundir. If
    verbose=TRUE it also writes text to the console
}
\description{
makecompareoutput simplifies taking the output from when
    comparing scenarios and producing the HTML files used to display all the
    results in rundir. Note that 'runnotes' is a vector of character strings
    made up using paste0.
}
\examples{
print("wait on internal data-sets")
}
