% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{makefilename}
\alias{makefilename}
\title{makefilename Generates a filename for output files}
\usage{
makefilename(hcrLabel, runlabel, reps, initDepl)
}
\arguments{
\item{hcrLabel}{the name of the harvest strategy beign used}

\item{runlabel}{the optional additional label for each run}

\item{reps}{the number of iterations of the MSE run of the dynamics}

\item{initDepl}{the initial depletion for the zone}
}
\value{
makefilename generates a filename for output files
}
\description{
makefilename Generates a filename for output files based upon
    he name of the HCR (hcrLabel), the runlabel, the number of reps and
    the initial depletion (initDepl), which should be sufficient to
    uniquely identify any different scenarios to be run
    Uses splitDate.
}
\examples{
\dontrun{
 tmp <- splitDate()
 print(tmp)
 print(names(tmp))
 print(as.numeric(tmp[1:3]))
 print("still need a full example for this function")
}
}
