% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{makeLabel}
\alias{makeLabel}
\title{makeLabel converts a vector of numbers or strings into a label}
\usage{
makeLabel(invect, insep = "_")
}
\arguments{
\item{invect}{the input vector of numbers or strings}

\item{insep}{the seperator between each part of the label}
}
\value{
a text string
}
\description{
makeLabel converts a vector of numbers or strings into a
    single label
}
\examples{
\dontrun{
 x <- 1:5
 makeLabel(x,insep="_")
 makeLabel(x,insep="-")
}
}
