% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{makelistobj}
\alias{makelistobj}
\title{makelistobj extracts a list of objects from a bigger list of objects}
\usage{
makelistobj(res, objname)
}
\arguments{
\item{res}{the list object from which individual objects are wanted from each
scenario}

\item{objname}{the name of the object wanted from each scenario}
}
\value{
a list of the named objects from res
}
\description{
makelistobj when making comparisons between scenarios first it
    is necessary to gather information from each scenario into large lists.
    For example, after running d the do_comparison function the output is
    a list which includes an obj called 'ans'. This contains the output from
    do_MSE for each scenario. Inside that are objects such as glb, ctrl,
    zoneDP, sauout, and so on. For further analysis we need to be able to
    extract copies of some of these objects into lists of their own. The
    function makelistobj can do that.
}
\examples{
\dontrun{
  # see the help for do_comparison for code lines prior to the next line
  result <- do_comparison(rundir,postfixdir,outdir,files,pickfiles=c(1,2,3))
  glball <- makelistobj(res=result$ans,objname="glb")
  ctrlall <- makelistobj(res=result$ans,objname="ctrl")
  str(glball)
}
}
\seealso{
{
    \link{do_comparison}, \link{do_MSE}
}
}
