% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{makemove}
\alias{makemove}
\title{makemove produces the movement matrix}
\usage{
makemove(npop, recs, ld, sigmove = 0)
}
\arguments{
\item{npop}{the number of populations}

\item{recs}{the original unfished recruitment levels estimated for
each population}

\item{ld}{the larval dispersal rate for all populations}

\item{sigmove}{currently no variation is assumed to occur. This is
here in case it needs to get implemented, but as there is
little or no data on actual larval dispersal levels currently
this is set to zero and not used.}
}
\value{
a npop x npop matrix describing movement among populations
}
\description{
makemove produces the movement matrix, which assumes
    that only adjacent populations contribute to each other. Thus,
    a central population would lose half its larvae to one side and
    half to the other side, but would receive half of their larval
    dispersal each. As larval dispersal is modelled as a proportion
    those populations that produce more larvae will lose more in
    absolute terms. Those populations at the edges of the zone
    only lose half the larval dispersal into the zone and it is
    assumed that what they lose out of the zone will be matched
    by what will come into the zone.
}
\examples{
recruits <- c(636146,263878,819189,1112513,285025,671573.9)
makemove(npop=6,recs=recruits,ld=0.04)
}
