% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{makeoutput}
\alias{makeoutput}
\title{makeoutput take the output from do_MSE and generates the HTML files}
\usage{
makeoutput(
  out,
  rundir,
  postdir,
  controlfile,
  hsfile = NULL,
  doproject = TRUE,
  openfile = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{out}{the output from do_MSE}

\item{rundir}{the full path to the directory holding the all scenario files}

\item{postdir}{the name of the directory holding the results, also used to
name the internal webpage}

\item{controlfile}{the controlfile used to run the MSE}

\item{hsfile}{the name of the harvest strategy file, default=NULL}

\item{doproject}{have the projections been run? default = TRUE.}

\item{openfile}{should the website be opened automatically? default=TRUE}

\item{verbose}{should details of producing the HTML files be written to the
console?  default=FALSE}
}
\value{
nothing but it does generate a set of HTML files in rundir. If
    verbose=TRUE it also writes text to the console
}
\description{
makeoutput simplifies taking the output from do_MSE and
    producing the HTML files used to display all the results in rundir. Note
    that 'runnotes' is now a vector of character strings made up using paste0.
}
\examples{
print("wait on internal data-sets")
# out=out;rundir=rundir;postdir=postfixdir;controlfile=controlfile;
# hsfile="TasHS Package"; doproject=FALSE;openfile=TRUE;verbose=FALSE
}
