% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{makeregionC}
\alias{makeregionC}
\title{makeregionC makes the constant part of the simulation}
\usage{
makeregionC(region, const)
}
\arguments{
\item{region}{the object derived from the readregionfile function}

\item{const}{the object derived from teh readdatafile function}
}
\value{
a list containing the constant part of the simulated region
}
\description{
makeregionC makes the constant part of the simulated
    region. Once defined this does not change throughout the
    simulation. Once made it still requires makeregion to be run
    to fill in the B0, ExBo, MSY, MSYDepl, and the popq values, and
    to produce regionD, the dynamic part of the new region
}
\examples{
print("Wait on built in data")
}
