% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{makewidedat}
\alias{makewidedat}
\title{makewidedat converts long data to a wide data format}
\usage{
makewidedat(inlong, mids, counts = FALSE)
}
\arguments{
\item{inlong}{the long form data.frame derived for commlf's makelongdat
containing at least 'year', 'length', and 'propcounts'}

\item{mids}{the range of lengths to be found within the input data file.
These are matched with the size distribution for each year of data and
the appropriate matrix cells filled with the propcounts.}

\item{counts}{default = FALSE, which means that the wide format will contain
the proportion of counts for each year. If TRUE, then the actual counts
by size-class will be output by year}
}
\value{
a matrix of proportional counts for each size-class x year
}
\description{
makewidedat takes the output of the commlf function makelongdat
    and converts the columns of year x length x count into a matrix of counts
    with axes of lengths x years, ready for comparisons with predicted length
    composition of catches from the operating model.
}
\examples{
data(lf10)
mids <- seq(138,210,2)
answer <- makewidedat(lf10,mids)
answer[1:20,]
answerC <- makewidedat(lf10,mids,counts=TRUE)
answerC[1:20,]
}
