% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{makezone}
\alias{makezone}
\title{makezone generates the parts of the simulated zone}
\usage{
makezone(glob, zoneC)
}
\arguments{
\item{glob}{the global constants defined for the current simulation.
These include numpop, nblock, midptsd, Nclass, and Nyrs}

\item{zoneC}{the zoneC object from makezoneC}
}
\value{
a list of the dynamics and the constant components of the
    simulation
}
\description{
makezone generates the dynamics components of the
    simulated zone and completes the constant components of the
    simulated zone. The term 'zone' refers to the upper level
    of geographical detail that is used. Thus, for example, in
    Tasmania we might simulate a number of statistical blocks with
    multiple populations. In combination, we refer to the total as
    a zone. The matrices are Nyrs x numpop and the two
    arrays are N x Nyrs x numpop.
}
\examples{
  data(zone1)
  glb <- zone1$globals
  data(constants)
  ans <- makezoneC(zone1,constants)
  zoneC <- ans$zoneC
  glb <- ans$glb
  ans2 <- makezone(glb,zoneC)
  str(ans2,max.level=2)
}
