% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{makezoneDP}
\alias{makezoneDP}
\title{makezoneDP generates the container for the projection dynamics}
\usage{
makezoneDP(allyr, iter, glb)
}
\arguments{
\item{allyr}{number of years of historical catches plus projection years}

\item{iter}{the number of replicates}

\item{glb}{the object containing the global variables}
}
\value{
a large list containing an object ready for the projection dynamics
}
\description{
makezoneDP generates an object designed to hold the outputs
    from each replicate within a set of projections. This is identical to
    zoneD except it contains a repeat for each iteration and now includes
    a cesau, which is the population-catch-weighted cpue for each SAU.
}
\examples{
print("Could add variation to the harvest rates so that when ")
print("prepareprojection was run the range of initial H values would increase ")
}
