% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS-HCR.R
\name{mcdahcr}
\alias{mcdahcr}
\title{mcdahcr conducts the MCDA and returns the TAC multiplier and details}
\usage{
mcdahcr(
  vectce,
  wid = 4,
  targqnt = 0.55,
  pmwts = c(0.65, 0.25, 0.1),
  hcr = c(0.25, 0.75, 0.8, 0.85, 0.9, 1, 1.05, 1.1, 1.15, 1.2),
  only = TRUE
)
}
\arguments{
\item{vectce}{the vector of cpue that forms the basis of the assessment}

\item{wid}{the number of years to use in grad4, default=4}

\item{targqnt}{the quantile for the cpue target level, default=0.55}

\item{pmwts}{performance measure weights, default = 0.65, 0.25, 0.1 for the
target, grad4, and grad1 PMs respectively. So the order matters.}

\item{hcr}{the harvest control rule scales that transform the combined score
into a TAC multiplier. A vector of 1 - 10 where each cell index
represents the upper limit of the the combined score.}

\item{only}{should the function only return the score or all details. TRUE
implies only return the score}
}
\value{
a list of the TAC multiplier, the score, and all the details
}
\description{
mcdahcr conducts the MCDA on the basis of a vector of cpue and
     other details prescribed in the function's arguments. It returns the
     TAC multiplier, the combined score, and all the details of the
     calculation. The hcr is a vector of 1:10 values where the cell index
     is used to allocate each combined scrore to a multiplier. The score is
     rounded up to the nearest integer and that is the index within hcr.
     Thus a score of <=1 points to the first cell, >1 and <=2 points to the
     second cell, and so on, up to a score between >9 and <=10 which points
     to the last cell.
}
\examples{
print("wait on data and time")
}
