% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{modprojC}
\alias{modprojC}
\title{modprojC produces three objects used to condition the zone}
\usage{
modprojC(zoneC, glob, projC)
}
\arguments{
\item{zoneC}{the constant part of the zone}

\item{glob}{the globals for the simulation}

\item{projC}{the project definition object from zone1 from readctrlfile}
}
\value{
an updated projC with projSel and projSelWt completed
}
\description{
modprojC produces an object used to condition the zone when
    projecting it following any conditioning. Once the initial conditions
    for the projection have been attained through an initial depletion of an
    unfished equilibrium, or more particularly by conditioning on historical
    data then the projC object will be used to condition the projections. In
    essence, this puts the selectivity and selectivity x weight-at-size into
    projC
}
\examples{
print("wait on new example data")
}
