% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{modregC}
\alias{modregC}
\title{modregC runs the region 3 x Nyrs to equilibrium}
\usage{
modregC(regC, regD, glob, lowlim = 0, uplim = 0.4, inc = 0.005)
}
\arguments{
\item{regC}{the constant components of the simulated region}

\item{regD}{the dynamic components of the simulated region}

\item{glob}{the general global variables}

\item{lowlim}{the lower limit of harvest rate applied, default=0.0}

\item{uplim}{the upper limit of harvest rate applied, default=0.4}

\item{inc}{the harvest rate increment at each step, default=0.005}
}
\value{
a list containing the updated regionC and product
}
\description{
modregC runs the doproduction function so it can then
    appropriately fill in each population's MSY and MSYDepl values.
    It outputs both the finalized regC and the prodution array.
}
\examples{
\dontrun{  # the doproduction part takes too long to run
  data(region1)
  glb <- region1$globals
  data(constants)
  ans <- makeregionC(region1,constants)
  regionC <- ans$regionC
  ans <- makeregion(glb,regionC)
  regionC <- ans$regionC
  regionD <- ans$regionD
  ans2 <- modregC(regionC,regionD,glb)
  str(ans2,max.level=2)
}   # regC=regionC; regD=regionD; glob=glb; lowlim=0.0;uplim=0.4;inc=0.005
}
