% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{modzoneC}
\alias{modzoneC}
\title{modzoneC runs the zone 3 x Nyrs to equilibrium}
\usage{
modzoneC(zoneC, zoneD, glob, lowlim = 0, uplim = 0.4, inc = 0.005)
}
\arguments{
\item{zoneC}{the constant components of the simulated zone}

\item{zoneD}{the dynamic components of the simulated zone}

\item{glob}{the general global variables}

\item{lowlim}{the lower limit of harvest rate applied, default=0.0}

\item{uplim}{the upper limit of harvest rate applied, default=0.4}

\item{inc}{the harvest rate increment at each step, default=0.005}
}
\value{
a list containing the updated zoneC and product
}
\description{
modzoneC runs the doproduction function so it can then
    appropriately fill in each population's MSY and MSYDepl values.
    It outputs both the finalized zoneC and the prodution array.
    Hence we need to push copies of both zoneC and zoneD into
    modzoneC
}
\examples{
\dontrun{  # the doproduction part takes too long to run
  data(zone1)
  glb <- zone1$globals
  data(constants)
  ans <- makezoneC(zone1,constants)
  zoneC <- ans$zoneC
  ans <- makezone(glb,zoneC)
  zoneC <- ans$zoneC
  zoneD <- ans$zoneD
  ans2 <- modzoneC(zoneC,zoneD,glb)
  str(ans2,max.level=2)
}   # zoneC=zoneC; zoneD=zoneD; glob=glb; lowlim=0.0;uplim=0.4;inc=0.005
}
