% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{modzoneCSel}
\alias{modzoneCSel}
\title{modzoneCSel changes the selectivity characteristics in zoneC}
\usage{
modzoneCSel(zoneC, sel, selwt, glb, yrs)
}
\arguments{
\item{zoneC}{the constant zone object from setupzone}

\item{sel}{the new selectivity as a matrix of selectivity by population}

\item{selwt}{new selectivity x WtL as a matrix of SelWt x Population}

\item{glb}{the globals object}

\item{yrs}{the number of years of projection}
}
\value{
the zoneC object modified ready for projection
}
\description{
modzoneCSel  changes the selectivity characteristics in zoneC
    which is necessary when making a projection under a different LML.
    This function is designed to allow multiple LML during any
    projections.
}
\examples{
print("wait on suitable data")
}
