% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{movav}
\alias{movav}
\title{movav generates a moving average of a vector}
\usage{
movav(x, n = 5, centred = TRUE)
}
\arguments{
\item{x}{a vector of numbers for which a moving average is wanted}

\item{n}{the length of the moving average, default = 5}

\item{centred}{should the moving average be centred across the vector or
start at index = 1, default = centred=TRUE}
}
\value{
a numeric vector containing a moving average of the input vector
    or a printed warning if the length is an even number.
}
\description{
movav generates a moving average of a vector of numbers. It
    allows the length of the moving average to be set using 'n', and
    whether the moving average starts at index 1 or is centred across the
    vector, using 'centred'. Only odd numbered lengths for the moving
    average are permitted to allow centring to be balanced. The output is
    originall a time-series but is converted to a numeric vector.
}
\examples{
x <- runif(25,min=1,max=25)
outav <- movav(x,n=5)
cbind(x,outav)
}
