% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mult}
\alias{mult}
\title{mult multiplies two matrices using RcppArmadillo}
\usage{
mult(A, B)
}
\arguments{
\item{A}{the first matrix}

\item{B}{the second matrix}
}
\description{
mult multiplies two matrices using RcppArmadillo. There are no
   tests for conformity so use at your own risk
}
