% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvC}
\alias{mvC}
\title{mvC matrix vector product using RcppArmadillo}
\usage{
mvC(inmat, invect)
}
\arguments{
\item{inmat}{the matrix, usualy the growth transition matrix}

\item{invect}{the vector to be multiplied. Usually the numbers-at-
size in a length based model}
}
\description{
mvC multiple the vector rhs by the matrix lhs, which
    is used all the time in a length-based stock assessment model.
    mvC appears to be 25 percent faster than R's inmat X invect
    when used in a length-based model context.
}
