% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{numbersatsizeSAU}
\alias{numbersatsizeSAU}
\title{numbersatsizeSAU plots the numbers-at-size for a given SAU}
\usage{
numbersatsizeSAU(
  rundir,
  glb,
  zoneC,
  zoneD,
  sau,
  yr = 1,
  defpar = TRUE,
  exploit = TRUE,
  mature = TRUE,
  filename = "Numbers-at-Size_Year1.png"
)
}
\arguments{
\item{rundir}{the results directory, if set to "" then plot is sent to
the console instead}

\item{glb}{the globals list}

\item{zoneC}{the constant part of the zpne structure}

\item{zoneD}{the dynamic part of the zone, zoneD}

\item{sau}{the SAU name to select for plotting. If multiple populations are
contained in an SAU their numbers-at-size will be combined.}

\item{yr}{which year of numbers-at-size should be plotted?}

\item{defpar}{should the plot parameters be defined. Set to FALSE if
numbersatsizeSAU is to be used to add a plot to a multiple plot.}

\item{exploit}{should exploitable numbers be plotted as well? default=TRUE}

\item{mature}{should mature numbers be plotted as well? default=TRUE}

\item{filename}{default='Numbers-at-Size_Year1.png', but can be changed
to suit whichever year is used}
}
\value{
nothing but it adda a plot to the results directory or console
}
\description{
numbersatsizeSAU plots the initial unfished numbers-
    at-size distribution for a given SAU, omitting the first four size
    classes to avoid the recruitment numbers dominating the plot.
}
\examples{
print("this will be quite long when I get to it")
}
