% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{onesau}
\alias{onesau}
\title{onesau plots the dynamics for a single SAU}
\usage{
onesau(
  prerep,
  postrep,
  glb,
  startyr,
  picksau,
  addCI = FALSE,
  CIprobs = c(0.05, 0.5, 0.95),
  histCE = NULL
)
}
\arguments{
\item{prerep}{the zoneDsau object that represents the SAU data}

\item{postrep}{the zonePsau object that represents the SAU data}

\item{glb}{the global constants object}

\item{startyr}{the year from which to begin the conditioned dynamics}

\item{picksau}{which sau should be plotted}

\item{addCI}{should the 90 percent CI be added, default=FALSE}

\item{CIprobs}{what CI should be fitted, default=c(0.05,0.5,0.95)}

\item{histCE}{should the historical CPUE be added to the cpue plot}
}
\value{
a list of the CI for each variable
}
\description{
onesau plots the details of matureB, exploitB, catch, acatch,
    harvestR, cpue, recruit, deplsB, depleB for a single SAU on one plot
}
\examples{
print("Wait on suitable data sets")
}
