% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{oneyearC}
\alias{oneyearC}
\title{oneyearC conducts one year's dynamics using catch not harvest}
\usage{
oneyearC(regC, regD, Ncl, catchp, year, sigmar, npop, movem)
}
\arguments{
\item{regC}{the constant portion of the region with a list of
properties for each population}

\item{regD}{the dynamics portion of the region, with matrices and
arrays for the dynamic variables of the dynamics of the
operating model}

\item{Ncl}{the number of size classes used to describe size}

\item{catchp}{a vector of catches to be taken in the year from each
population}

\item{year}{the year of the dynamics, would start in year 2 as year
1 is the year of initiation.}

\item{sigmar}{the variation in recruitment dynamics, set to 1e-08
when searching for equilibria.}

\item{npop}{the number of populations, the global numpop}

\item{movem}{the larval dispersal movement matrix}
}
\value{
a list containing a revised dynamics list
}
\description{
oneyearD conducts one year's dynamics in the simulation
    using catches rather than harvest rates. The harvest rates are
    estimated after first estimating the exploitable biomass.
    returning the revised regionD, which will have had a single year
    of activity included in each of its components.
}
\examples{
 data(constants)
 data(region1)
 ans <- makeregionC(region1,constants) # classical equilibrium
 regionC <- ans$regionC
 glb <- ans$glb
 ans <- makeregion(glb,regionC) # now make regionD
 regionC <- ans$regionC  # region constants
 regionD <- ans$regionD
 npop <- glb$numpop
 Nc <- glb$Nclass
 nyrs <- glb$Nyrs
 catch <- 360.0 # larger than total MSY ~ 310t
 B0 <- getvar(regionC,"B0")
 totB0 <- sum(B0)
 prop <- B0/totB0
 catchpop <- catch * prop
 for (yr in 2:nyrs)
   regionD <- oneyearC(regC=regionC,regD=regionD,Ncl=Nc,
                   catchp=catchpop,year=yr,sigmar=1e-08,npop=npop,
                   movem=glb$move)
 str(regionD)
 round(regionD$catchN[60:105,1:5,1],1)
}
