% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{oneyearcat}
\alias{oneyearcat}
\title{oneyearcat one year's catch dynamics for one abpop}
\usage{
oneyearcat(inpopC, inNt, Nclass, incat, yr)
}
\arguments{
\item{inpopC}{a single population from a zoneC, an abpop}

\item{inNt}{the numbers at size for the year previous to the year
of dynamics. These are projected into the active year.}

\item{Nclass}{the number of size classes used to describe growth.
used to define vectors}

\item{incat}{the literal annual catch from the population. Derived
from teh harvest control rule for each SAU, then allocated to
each population with respect to its relative catching
performance.}

\item{yr}{the year in the dynamics being worked on. The first year
is generated when the zone is defined or when it is initially
depleted. All dynamics are appllied from year 2 - Nyrs; scalar}
}
\value{
a list containing ExploitB, MatureB, Catch, Harvest, Nt,
    ce, and CatchN used to update the given pop in yr + 1
}
\description{
oneyear does one year's dynamics for one input
    population. Where the fishing is based on a given catch per
    population, which would be determined by any harvest control
    rule. It is used to step through populations of a zone.
    Its dynamics are such that each population first undergoes
    growth, then half natural mortality. This allows an estimate of
    exploitable biomass before fishing occurs. The remaining
    dynamics involve the removal of the catch, after estimating the
    harvest rate from catch/exploitb, the application of the last
    half of natural mortality and the addition of recruits. This
    allows the exploitable biomass to be estimated after fishing.
    The cpue uses the average of the pre-fishing and post-fishing
    exploitB to smooth over any changes brought about by fishing.
    The recruitment occurs in oneyearC so that larval dispersal can
    be accounted for. oneyearcat is not used independently of
    oneyearC.
}
\examples{
print("need to wait on built in data sets")
}
