% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{oneyearcatF}
\alias{oneyearcatF}
\title{oneyearcatF one year's catch dynamics for one abpop}
\usage{
oneyearcatF(
  MatWt,
  SelWt,
  selyr,
  Me,
  G,
  scalece,
  WtL,
  inNt,
  incat,
  sigce,
  lambda,
  qest
)
}
\arguments{
\item{MatWt}{maturity x Weight-at-size from zoneCC for the population and year}

\item{SelWt}{selectivity x Weight-at-size from zoneCC for the population and year}

\item{selyr}{selectivity from zoneCC for the population and year}

\item{Me}{natural morality from zoneCC for the population}

\item{G}{growth transtion matrix from zoneCC for the population}

\item{scalece}{the zoneC'pop'$scalece used to scale al cpue to the sau
nominal scale}

\item{WtL}{the weight-at-length from zoneCC for the population}

\item{inNt}{the numbers at size for the year previous to the year
of dynamics. These are projected into the active year.}

\item{incat}{the literal annual catch from the population. Derived
from the harvest control rule for each SAU, then allocated to
each population with respect to its relative catching
performance.}

\item{sigce}{the process error variation associated with the
estimates of cpue from the model. Found in the ctrl object
as withsigCE. It is implemented as Log-Normal error on the
exploitable biomass value from the model}

\item{lambda}{the hyper-stability term from zoneC}

\item{qest}{the estimated catchability from sizemod from zoneC}
}
\value{
a list containing a vector of ExploitB, MatureB, Catch, and ce, and
    a matrix NaL containing Nt and CatchN used to update a pop in yr + 1
}
\description{
oneyearcatF does one year's dynamics for one input
    population. Where the fishing is based on a given catch per
    population, which would be determined by any harvest control
    rule. It is used to step through populations of a zone.
    Its dynamics are such that each population first undergoes
    growth, then half natural mortality. This allows an estimate of
    exploitable biomass before fishing occurs. The remaining
    dynamics involve the removal of the catch, after estimating the
    harvest rate from catch/exploitb, the application of the last
    half of natural mortality and the addition of recruits. This
    allows the exploitable biomass to be estimated after fishing.
    The cpue uses the average of the pre-fishing and post-fishing
    exploitB to smooth over any changes brought about by fishing.
    The recruitment occurs in oneyearC so that larval dispersal can
    be accounted for. oneyearcatF is not used independently of
    oneyearC.
}
\examples{
print("need to wait on built in data sets")
print("Uses Instantaneous fishing mortality rates")
# oneyearcatF(MatWt=pop$MatWt,SelWt=pop$SelWt[,year],selyr=pop$Select[,year],
#          Me=pop$Me,G=pop$G,scalece=pop$scalece,WtL=pop$WtL,inNt=inN[,popn],
#          incat=popC[popn],sigce=sigce,lambda=pop$lambda,qest=pop$qest)
}
\seealso{
{
 \link{dohistoricC}, \link{oneyearcat}, \link{oneyearrec}
}
}
