% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{oneyearrec}
\alias{oneyearrec}
\title{oneyearrec calculates the Beverton-Holt recruitment}
\usage{
oneyearrec(steep, R0, B0, Bsp, sigR)
}
\arguments{
\item{steep}{the steepness for the population; scalar}

\item{R0}{the unfished recruitment levels for the population; scalar}

\item{B0}{the unfished spawning biomass; scalar}

\item{Bsp}{the current spawning biomass; scalar}

\item{sigR}{standard deviation of the recruitment residuals;
scalar. set this to 1e-08 to avoid recruitment variability}
}
\value{
an absolute number of recruits from a given spawning biomass
}
\description{
oneyearrec calculates the Beverton-Holt recruitment for the
   input populations in a single year; parameterized with steepness,
   R0, and B0. To drop variation to insignificant levels set sigmar-1-08
}
\examples{
\dontrun{
steep <- 0.707
R0 <- 319971
B0 <- 313
Bsp <- 147
insigmar <- 0.3
oneyearrec(steep,R0,B0,Bsp,insigmar)
}
}
