% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{onezoneplot}
\alias{onezoneplot}
\title{onezoneplot plots out one variable from the zone}
\usage{
onezoneplot(invar, rundir, CIprobs, varname, addfile = TRUE)
}
\arguments{
\item{invar}{which zone variable to plot}

\item{rundir}{the run directory for the results}

\item{CIprobs}{the quantiles to use for the CI.}

\item{varname}{the name of the variable for use in labels}

\item{addfile}{should a png file be added to the results, default = TRUE}
}
\value{
it generates a png file of the plot if addfile remains TRUE
}
\description{
once the projected zone has been summarized by poptozone one
    can add plots to the results using plotZone. This in turn uses
    onezoneplot to plot each variable in turn.
}
\examples{
print("wait on suitable data")
}
