% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docondition.R
\name{optimizeAvRec}
\alias{optimizeAvRec}
\title{optimizeAvRec improves OM fit to CPUE by adjustnig AvRec}
\usage{
optimizeAvRec(
  rundir,
  controlfile,
  datafile,
  calcpopC,
  lowmult = 0.7,
  highmult = 1.3,
  linenumber = 29,
  verbose = TRUE
)
}
\arguments{
\item{rundir}{the rundir for the scenario}

\item{controlfile}{the name of the control file}

\item{datafile}{the name of the datafile from the controlfile}

\item{calcpopC}{the funciton, from the HS file that allocates catches
across each of the populations within each SAU}

\item{lowmult}{the multiplier to place a lower bound when search for the
optimum AvRec for each SAU. default = 0.7, ie param AvRec * 0.7}

\item{highmult}{the multiplier to place an upper bound when search for the
optimum AvRec for each SAU. default = 1.3, ie param AvRec * 1.3}

\item{linenumber}{the linenumber inside the datafile that is to be changed
to the optimum AvRec values. Default = 29, whic suits Tasmania.}

\item{verbose}{should updates on progress be sent to the console.
default=TRUE}
}
\value{
a vector of the final optimized AvRec values. This also modifies
    the datafile - SO BE CAREFUL.
}
\description{
optimizeAvRec is used to improve the fit to the time-series
    of CPUE by adjusting the SAU initial value for AvRec. If one assesses
    an SAU using the sizemod package, that summarizes across a whole SAU.
    After population allocation within aMSE, the productivity of an SAU
    may well have been disturbed away (up or down) from the assessed
    optimum. Hence, it becomes necessary, still, to optimize both the
    AvRec and the recdevs during the conditioning.
}
\examples{
print("wait on suitable data files")
}
