% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{phaseplotpolygons}
\alias{phaseplotpolygons}
\title{phaseplotpolygons sets up the colour blocks for phase plots}
\usage{
phaseplotpolygons(maxy, maxx = 1, targx = 0.4, limx = 0.2, limy = 0.175)
}
\arguments{
\item{maxy}{the upper bound on the y-axis}

\item{maxx}{the upper bound on the x-axis}

\item{targx}{the target reference point depletion or biomass}

\item{limx}{the limit reference point depletion or biomass}

\item{limy}{the limit reference point for fishing mortality or harvest rate}
}
\value{
nothing but it does add to a plot
}
\description{
phaseplotpolygons is used to put together coloured phase plots
    a common example of which would be kobe plots. This function assumes that
    a plot has already been defined as well as the maximum x and y values.
    In a standard kobe plot y would relate to fishing mortality or harvest
    rate, and x would relate to biomass or depletion. If plotting a kobe plot
    then the target and limit depletion are important as is the limit harvest
    rate. This function only generates the coloured blocks it is for use
    within other functions
}
\examples{
\dontrun{
# illustrating a form of use
 medH <- apply(H,1,median)
 meddepl <- apply(depl,1,median)
 plotprep(width=7,height=6,filename=filen,cex=0.9,verbose=FALSE)
 parset(cex.lab=1.25)
 maxH <- getmax(medH)
 maxdepl <- getmax(meddepl,mult=1.25)
 plot(meddepl,medH,type="p",pch=16,xlim=c(0,maxdepl),ylim=c(0,maxH),xaxs="i",
      yaxs="i",xlab=paste0("Mature Biomass Depletion ",glb$saunames[sau]),
      ylab="Harvest Rate")
 phaseplotpolygons(maxH=maxH,maxdepl=maxdepl,targdepl=targdepl,
                   limdepl=limdepl,limH=limH)
}
}
