% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{plotallphaseplots}
\alias{plotallphaseplots}
\title{plotallphaseplots generate a set of phase plots of dynamics variables}
\usage{
plotallphaseplots(
  rundir,
  dyn,
  glb,
  scenes,
  width = 9,
  height = 10,
  fnt = 7,
  pntcex = 1.5,
  zero = FALSE,
  legloc = "bottomright"
)
}
\arguments{
\item{rundir}{the compare scenario's directory}

\item{dyn}{a list of the dynamics objects from each scenbario}

\item{glb}{the globals object. The assumption is that all glb objects are
very similar between scenarios, although different reps can be compared}

\item{scenes}{the names of the scenarios selected for comparison}

\item{width}{width of the plot. default=9}

\item{height}{height of the plot. default=10}

\item{fnt}{which font to use. default=7 = bold times}

\item{pntcex}{the size of the year points along each median line}

\item{zero}{should the phase plots have a zero origin. default=FALSE}

\item{legloc}{location of the legend placed into the last plot.
default='bottomright'}
}
\value{
nothing
}
\description{
plotallphaseplots produced a series of phase plots for a set of
    different scenarios of how mature biomass and its depletion varies with
    aspirational catch along with other combinations of cpue and harvest
    rate against catch.
}
\examples{
print("wait on sutiable data")
}
\seealso{
[\link{plotallphaseplots} which wraps a set of phase plots]
}
