% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{plotbysau}
\alias{plotbysau}
\title{plotbysau plots a collection of projections by SAU}
\usage{
plotbysau(zoneDP, glb, rundir)
}
\arguments{
\item{zoneDP}{the dynamic object produced by the projections}

\item{glb}{the object containing the global variables}

\item{rundir}{the results directory}
}
\value{
adds 6 plots to rundir and returns a list of qauntile for 6
    variables
}
\description{
plotbysau generates and store plots in resir
}
\examples{
print("wait on suitable data-sets")
}
