% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotclosed}
\alias{plotclosed}
\title{plotclosed generates two plots of the closed and open replicates}
\usage{
plotclosed(invar, whichclosed, sau, yrs, reps, label, hline = 0)
}
\arguments{
\item{invar}{which variable from sauout to plot}

\item{whichclosed}{which replicates are closed}

\item{sau}{which sau are we plotting}

\item{yrs}{what years to plot hyrs : (hyrs + pyrs)}

\item{reps}{the total number of replicates from glb}

\item{label}{the ylabel usually invarname + sauname}

\item{hline}{shoudl a horizontal reference line be included. default = 0}
}
\value{
nothing but it does generate two plots
}
\description{
plotclosed is used if any sau experiences clsosures at any time
    in any replicate. For a selected variable from sauout, which includes
    matureB, exploitB, midyexpB, catch, acatch, harvestR, cpue, recruit,
    deplsB, and depleB, it plots replicates that included closures and then
    plots those without closures for the the same sau. The outputs are sent
    to the closure tab. If no closures then no closure tab. Each pair of
    plots have the same y-axis scale to make comparisons simpler.
}
\examples{
# syntax plotclosed(invar=cpue,whichclosed=whichclosed,sau=pickC[i],yrs=yrs,
#                   label=label)
}
