% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotCNt}
\alias{plotCNt}
\title{plotCNt plots the historical conditioning Nt or catchN for all years}
\usage{
plotCNt(Nt, glb, vline = NULL, start = 3)
}
\arguments{
\item{Nt}{the 3D array of numbers-at-size x years x SAU from prepareDDNt}

\item{glb}{the global constants object}

\item{vline}{add a vertical dashed line, perhaps at the LML, default=NULL}

\item{start}{from what size class should the maximum Y value be measured.
The default = 3 (= 6mm)}
}
\value{
nothing but it does plot a graph
}
\description{
plotCNt the historical conditioning on the fisheries data leads
    to a three dimensional array of numbers-at-size x years x numpop. We use
    'prepareDDNt' to convert this to a 3D array by SAU. PlotCNt then plots,
    for each SAU, numbers-at-size for all years of conditioning with the
    first year emphasized in black and the last year in red.
}
\examples{
 print("wait on suitable built in data-sets")
}
