% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docondition.R
\name{plotcondCPUE}
\alias{plotcondCPUE}
\title{plotcondCPUE plots the historical cpue against conditioned cpue}
\usage{
plotcondCPUE(histCE, saucpue, glb, rundir, filen = "")
}
\arguments{
\item{histCE}{the matrix of historical cpue by SAU}

\item{saucpue}{the predicted cpue from the conditioning on historical data}

\item{glb}{the globals object}

\item{rundir}{the rundir for the given scenario}

\item{filen}{the file name of the plot if it is to be saved}
}
\value{
a vector of length nsau containing the ssq for each SAU
}
\description{
plotcondCPUE generates a plot of the historical cpue and compares
    it with the predicted cpue from the conditioning. The predicted is black
    and the observed is green. This is primarily there are an aid to
    conditioning the operating model. It also calculates the simple sum of
    squared differences between the observed and predicted, again to aid in
    the conditioning.
}
\examples{
print("wait on suitable internal data sets")
}
