% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotconditioning}
\alias{plotconditioning}
\title{plotconditioning illustrate the zones dynamics after conditioning}
\usage{
plotconditioning(zoneDD, glb, zoneC, histCE, rundir, recdevs)
}
\arguments{
\item{zoneDD}{the conditioned dynamic zone object}

\item{glb}{the global constants object}

\item{zoneC}{the constant zone object}

\item{histCE}{the historical CPUE matrix}

\item{rundir}{the rundir for the given scenario}

\item{recdevs}{the recruitment deviates from the control file, condC}
}
\value{
the sau scale zone dynamics, invisibly
}
\description{
plotconditioning converts the conditioned dynamic zone object
    from population scale to sau scale and then plots the various components
    to depict their state after conditioning. This is designed to place such
    plots into the rundir to assist in the conditioning process.
}
\examples{
print("wait on suitable data sets")
}
