% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotdeltarec}
\alias{plotdeltarec}
\title{plotdeltarec if timevary then generates a plot of recruitment change}
\usage{
plotdeltarec(glb, console = TRUE)
}
\arguments{
\item{glb}{the globals object}

\item{console}{print to the console or to file, default = TRUE}
}
\value{
nothing but plots a graph is TIMEVARY option is active
}
\description{
plotdeltarec This plots the relationship between year of
    projection and the change in recruitment or settlement success brought
    about by the TIMEVARY inputs
}
\examples{
print("Wait on example data")
}
