% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{plotdevs}
\alias{plotdevs}
\title{plotdevs plots projection deviates for a variable across scenerios}
\usage{
plotdevs(invar, scenes, saunames, filen = "", resol = 300)
}
\arguments{
\item{invar}{the sub-object from each scenario as a list, see examples}

\item{scenes}{a vector of names fr each scenario being compared}

\item{saunames}{a vector of the names of each sau in the projections}

\item{filen}{the filename and path if the plot is to be saved, default = "".
if a filename is given it should be a .png file}

\item{resol}{default = 300, the dpi for the png plots}
}
\value{
a list f the mean deviates per sau and scenario, and a list of their
    respective standard deviations. It also plots a graph
}
\description{
plotdevs provides a plot of a scenario Performance Measure. In
    this case the PM are the deviations from a loess fit to  whatever input
    variable has been selected from the dyn objects in do_comparison(). One
    needs first generate a list of the selected variable from the the input
    dyn objects (one from each scenario). A standard comparison would be on
    the variation of catches through the years of the projection. But this
    could be applied to any of the variables within the dyn object, which
    includes matureB, exploitB, midyexpB, catch, acatch, harvestR, cpue,
    recruit, deplsB, and depleB.
}
\examples{
\dontrun{
# example syntax, assumes dynamics have been extracted from all scenarios
   glb <- glbc[[1]]
   saunames <- glb$saunames
   nsau <- glb$nSAU
   nscen <- length(scenes)
   x <- makelist(scenes)
   for (scen in 1:length(scenes)) x[[scen]] <- dyn[[scen]]$catch[59:88,,]
   devout <- plotdevs(x,scenes,saunames,filen="")
}
}
