% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{plotdynphase}
\alias{plotdynphase}
\title{plotdynphase plots the medians of 2 dynamic variables as a phase diagram}
\usage{
plotdynphase(
  xlist,
  ylist,
  scenes,
  glb,
  rundir = "",
  xlab = "xlabel",
  ylab = "ylabel",
  legloc = "bottomright",
  console = TRUE,
  width = 9,
  height = 10,
  fnt = 7,
  pntcex = 1.25,
  zero = FALSE,
  yline = NULL
)
}
\arguments{
\item{xlist}{x-axis list of each scenes replicates for the selected variable,
generated from result using getcompout - 'get component out'}

\item{ylist}{y-axis list of each scenes replicates for the selected variable}

\item{scenes}{names of the difference scenes being compared. result$scenes}

\item{glb}{the aMSE globals object, inside one of the result$ans lists}

\item{rundir}{the directory into which all files and results are placed}

\item{xlab}{label for the x-axis}

\item{ylab}{label for the y-axis}

\item{legloc}{location of the legend placed into the last plot.
default='bottomright'}

\item{console}{should the plot go to the console or a png file. default=TRUE}

\item{width}{width of the plot. default=9}

\item{height}{height of the plot. default=10}

\item{fnt}{which font to use. default=7 = bold times}

\item{pntcex}{the size of the year points along each median line}

\item{zero}{should the phase plots have a zero origin. default=FALSE}

\item{yline}{default = NULL, used to add horizontal lines to each plot}
}
\value{
an invisible caption to the plot.
}
\description{
plotdynphase is used when comparing how the population dynamics
     are affected by multiple scenarios. The median of the replicate
     trajectories for each selected variable are plotted against each other
     in a classical phase plot. Thus, a stable solution should display a
     single point or an oscillation around a point of convergence. Each plot
     can have an origin at zero or be scaled to display maximum detail using
     the 'zero' argument. As with getcompout there are only ten valid variable
     names that can be combined: matureB, exploitB, midyexpB, catch, acatch,
     harvestR, cpue, recruit, deplsB, and depleB. The results for each SAU
     are plotted together.
}
\examples{
print("wait on suitable data sets")
# result <- do_comparison(rundir=rundir,postfixdir=postfixdir,outdir="",
#                         files=fullfiles,pickfiles=c(1,4),verbose=TRUE,
#                         intensity=100)
# matureB <- getcompout(result,pickvar="matureB")
# acatch <- getcompout(result,pickvar="acatch")
# glb <- result$ans[[1]]$glb
# scenes <- result$scenes
# plotdynphase(xlist=acatch,ylist=matureB,scenes=scenes,glb=glb,rundir="",
#              xlab="Aspirational Catch (t)",ylab="Mature Biomass (t)",
#              legloc="bottomright",console=TRUE,width=9,height=10,fnt=7,
#              pntcex=1.5,zero=FALSE)
}
\seealso{
[\link{getcompout} for how to generate the xlist and ylist]
}
