% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{plotdynvarinyear}
\alias{plotdynvarinyear}
\title{plotdynvarinyear  histograms of a dyn variable by scenario and sau}
\usage{
plotdynvarinyear(
  rundir,
  dyn,
  whichvar,
  whichyr,
  glb,
  bins = 15,
  console = TRUE,
  whichfun = mean,
  resol = 300
)
}
\arguments{
\item{rundir}{the directory in which comparisons are being made. It is best
to be a separate directory form any particular scenario.}

\item{dyn}{the dynamics objects from each scenario in a single dyn object
from the output of do_comparison}

\item{whichvar}{which variable within the dyn object to characterize}

\item{whichyr}{which year number to use}

\item{glb}{the globals object needed for the number and names of the sau,
and the years and their names.}

\item{bins}{how make breaks should be used in the histograms, default=15}

\item{console}{should each plot go to the console or be saved to rundir.
default = TRUE ie go to console}

\item{whichfun}{each plot has by default the mean of the distribution added
if one wanted the median or geomean then change to that function name}

\item{resol}{default = 300, the dpi for the png plots}
}
\value{
the filename (no path) used if any, and it plots a graph
}
\description{
plotdynvarinyear generates a set of histograms of a user
    select variable within the dyn object from do_comparison, all with the
    same x-axis, for a given year by scenario and SAU. The number of breaks
    can be modified but defaults to 15. Only a single year at a time is
    characterized.
}
\examples{
# assumes that do_comparison function has been run. syntax used:
# filen <- plotdynvarinyear(rundir=rundir,dyn=dyn,whichvar="catch",
#                           whichyr=c(63),glb,console=TRUE)
}
