% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{plotfinalscores}
\alias{plotfinalscores}
\title{plotfinalscores plots projected catches, cpue, PMs, and final score}
\usage{
plotfinalscores(
  outscores,
  minprojC = 0,
  minprojCE = 60,
  mintargCE = 110,
  filen = ""
)
}
\arguments{
\item{outscores}{each sau's results from getcpueHS, which extracts the HS
scores and other results from the projection dynamics}

\item{minprojC}{the minimum y-axis value for the projected catches}

\item{minprojCE}{the minimum y-axis value for the projected cpue}

\item{mintargCE}{the minimum y-axis value for the projected target cpue}

\item{filen}{a filename for use if saving the output to said file,
default = '', which implies the plot goes to the console.}
}
\value{
invisibly a list of the median catches, cpue, targetce, finalscore,
    grad1 score, grad4 score, and target cpue score.
}
\description{
plotfinalscores takes the inidivual sau output from getcpueHS
    and generates a plot made up of six sub-plots of the replicates for each
    sau. In sequence these plots are the projected catches, the projected
    cpue, the gradient 1 scores, the gradient 4 scores, the target cpue, and
    the final total score. This enables the relationships between the scores
    to be examined. Only used in do_MSE
}
\examples{
print("wait on data sets")
}
\seealso{
{
   \link{do_MSE}
}
}
