% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotfisfit}
\alias{plotfisfit}
\title{plotfisfit compares the observed fis with the predicted for each sau}
\usage{
plotfisfit(condC, predfis, forfis, rundir = "", console = TRUE)
}
\arguments{
\item{condC}{object holding observed fisheries data including the fisindex}

\item{predfis}{the predicted fis values during conditioning}

\item{forfis}{the sauWts, selfis, yearfis, and fissau added to hsargs within
do_MSE}

\item{rundir}{the output directory for a scenario, default = ''}

\item{console}{should plot go to the console or to rundir, default = TRUE}
}
\value{
a martix of observed and predicted FIS values for each SAU, it
    also generates a plot.
}
\description{
plotfisfit plots the observed and predicted FIS for each SAU
}
\examples{
# syntax plotfisfit(condC=condC,predfis=zoneDD$predfis,forfis=hsargs$forfis,
#                   rundir=rundir,console=FALSE)
# condC=condC;predfis=zoneDD$predfis;forfis=hsargs$forfis;rundir="";console=TRUE
}
