% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{plothistcatch}
\alias{plothistcatch}
\title{plothistcatch plots the historical catches used for conditioning}
\usage{
plothistcatch(zone1, pops = NULL, rundir = "", defpar = TRUE)
}
\arguments{
\item{zone1}{The zonewide properties as in readzonefile(indir,ctrl$zonefile)}

\item{pops}{default = NULL. The plot can be restricted to a specific set
of populations by providing the indices of the populations to be
included. Thus if there were 8 populations the pops=c(1,2,3,8), would
plot the first three and the last population.}

\item{rundir}{the results directory used by makehtml to store plots and
tables. If set to "", the default, then it plots to the console}

\item{defpar}{define the plot parameters. Set to FALSE if using
plothistcatch to add a plot to a multiple plot}
}
\value{
nothing, it adds a plot into rundir and modifies resultTable.csv
}
\description{
plothistcatch provides a visual representation of the catches
    taken in each SAU/population used in the conditioning of the
    Operating Model.
}
\examples{
print("wait until I have altered the internals data sets")
}
