% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plothistce}
\alias{plothistce}
\title{plothistce generates a plot of the historical cpue by SAU}
\usage{
plothistce(rundir, condC, glb, proportion = FALSE, filen = "")
}
\arguments{
\item{rundir}{the direcotry where all results are stored}

\item{condC}{the R object contiaining the historical fishery data generated
by the readctrlfile function}

\item{glb}{the globol variables object}

\item{proportion}{should the plot be of proportion relative to the maximum
catch through each time-series or as raw tonnes x year, default=FALSE}

\item{filen}{the name of the file used to save the plot, default="", which
sends the plot to a separate window}
}
\value{
nothing but it does generate a plot either to a window or a file
}
\description{
plothistce generates a plot of historical cpue by SAU. It can
    plot as cpue or proportion of maximum cpue.
}
\examples{
print("wait on suitable data sets")
}
