% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plothsstats}
\alias{plothsstats}
\title{plothsstats plots some HS performance statistics}
\usage{
plothsstats(rundir, hsstats, glb)
}
\arguments{
\item{rundir}{the scenario's results directory}

\item{hsstats}{the statistics about the harvest strategies performance}

\item{glb}{the global constants object}
}
\value{
nothing but it does generate two plots into rundir
}
\description{
plothstats generates plots for the sum of catches for the
    first 5 and 10 years of the projections. This is really only
    interesting when compared with the same statistics from a different
    operating model definition or different harvest strategy. The input
    data is generated during the running of do_MSE.
}
\examples{
print("wait on suitable in ternal data sets")
}
