% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotindices}
\alias{plotindices}
\title{plotindices combines CPUE and FIS plots for each SAU with a FIS}
\usage{
plotindices(zoneDD, glb, condC, hsargs, rundir = "", console = TRUE)
}
\arguments{
\item{zoneDD}{The zoneDD object contains the dynamics up to but not including
the projections.}

\item{glb}{the globals object, contains useFIS as a boolean}

\item{condC}{contains the historical catches and CPUE and any fisindex data}

\item{hsargs}{the flexible input arguments to do_MSE, that includes fispar
which has the FIS selectivity parameters L50, and the delta parameter, as
well as the LML during the FIS operations.}

\item{rundir}{the output directory for a scenario, default = ''}

\item{console}{should plot go to the console or to rundir, default = TRUE}
}
\value{
a list of the ratio of the observed fis and predicted fis relative
    to their respective means, its also generates a plot.
}
\description{
plotindices if any SAU has FIS data then this generates a plot
    that combines both COPUE and FIS data in each SAU and puts them on the
    same scale. It does this by dividing each time-series through by the
    arithmetic mean of each time-series.
}
\examples{
print("wait on data sets")
}
