% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotNt}
\alias{plotNt}
\title{plotNt plots the size-composition for each SAU}
\usage{
plotNt(Nt, year, glb, start = 3, medcol = 0)
}
\arguments{
\item{Nt}{the 4 dimensional array of numbers-at-size Nclass,years,nsau,reps.}

\item{year}{which year from dimension 2 should be plotted?}

\item{glb}{the global constants object}

\item{start}{from what size class should the maximum Y value be measured.
The default = 3 (= 6mm)}

\item{medcol}{colour of the median line; default=0 = not plotted}
}
\value{
invisibly the matrix of median values, Nclass x nsau
}
\description{
plotNt accepts size-composition data (either the population Nt,
    of the catchN), and plots the replicate size-distributions for each SAU.
    It then puts the first year's size-distribution on top and if the medcol
    that defines the colour of the median of the replicates is different
    from 0 (not plotted) then it plots the median as well. A contrasting
    colour choice is 4 = Blue.
}
\examples{
print("wait on suitable built in data-sets")
}
