% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotpopprops}
\alias{plotpopprops}
\title{plotpopprops plots out some properties of input populations}
\usage{
plotpopprops(x, rundir, glb, varnames, startyr, console = TRUE, bins = 25)
}
\arguments{
\item{x}{the array containing columns to plot}

\item{rundir}{the run directory for the results, can be set to "" if plotting
to the console}

\item{glb}{the global constants object}

\item{varnames}{the names of each variable to select the columns and to use
in labels}

\item{startyr}{in what year number should the plot refer to?}

\item{console}{should plot go to the console or a png be saved? Default=TRUE}

\item{bins}{the number of breaks to plot in each histogram, default=25}
}
\value{
it generates a png file of the plot if console = FALSE, otherwise it
    is drawn to the console
}
\description{
once the depleted zone (zoneCP and zoneDD) has been generated
    the component populations have their properties summarized in propD. This
    is saved as propertyDD.csv and output to the zoneDD tab. To provide a
    visual consideration of these properties this function generated
    histograms for an array of input array column names.
}
\examples{
print("wait on suitable data")
}
