% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotpredfis}
\alias{plotpredfis}
\title{plotfisfit compares the observed fis with the predicted for each sau}
\usage{
plotpredfis(predfis, glb, forfis, rundir = "", console = TRUE)
}
\arguments{
\item{predfis}{the predicted fis values during conditioning}

\item{glb}{the globals object}

\item{forfis}{the sauWts, selfis, yearfis, and fissau added to hsargs within
do_MSE}

\item{rundir}{the output directory for a scenario, default = ''}

\item{console}{should plot go to the console or to rundir, default = TRUE}
}
\value{
Nothing but it does generate a plot for each SAU with FIS data
}
\description{
plotfisfit plots the observed and predicted FIS for each SAU
    combioned with a residual plot in each case (sau)
}
\examples{
# syntax plotpredfis(predfis=zoneDD$predfis,glb=glb,forfis=hsargs$forfis,
#                   rundir=rundir,console=FALSE)
# predfis=zoneDD$predfis;glb=glb;forfis=hsargs$forfis;rundir="";console=TRUE
}
