% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotprod}
\alias{plotprod}
\title{plotprod graphs a selected productivity variable}
\usage{
plotprod(
  product,
  xname = "MatB",
  yname = "Catch",
  xlimit = NA,
  xlab = "Mature Biomass t",
  ylab = "Production t"
)
}
\arguments{
\item{product}{the output from doproduction}

\item{xname}{the name of the productivity variable for the x-axis,
defaults to MatB}

\item{yname}{the name of the y-xis variable, default=Catch=Yield}

\item{xlimit}{default=NA, enables the range of the x-axis to be
constrained, for example using xlimit=c(0.15,0.4)}

\item{xlab}{the x-axis label, default=""}

\item{ylab}{the y-axis label, default=""}
}
\value{
invisibly a list of the x and y matrices plotted
}
\description{
plotprod graphs a selected productivity variable from
    the choice of ExB exploitable biomass, MatB mature of spawning
    biomass = Bmsy, AnnH the actual annual harvest rate Hmsy, Catch
    the yield at Bmsy and Hmsy = MSY, Deplet the mature biomass
    depletion level, and RelCE the relative cpue at MSY
}
\examples{
data(product)
plotprep(width=7, height=5)
plotprod(product)
msy <- getlistvar(regionC,"MSY")
effB0 <- getlistvar(regionC,"effB0")
abline(h=msy,col=1:6,lwd=2)
abline(v=effB0,col=1:6,lwd=2)
}
