% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{plotproductivity}
\alias{plotproductivity}
\title{plotproductivity characterizes each population's yield curve}
\usage{
plotproductivity(rundir, product, glb)
}
\arguments{
\item{rundir}{the results directory}

\item{product}{the productivity 3-D array}

\item{glb}{the globals list}
}
\value{
It produces five png files into rundir and returns sauprod, a matrix
    of B0, Bmsy, MSY, Depletionmsy, and CEmsy for each sau
}
\description{
plotproductivity characterizes each population's yield curve, it
  also describes the total productivity of the zone.
}
\examples{
print("this will be quite long when I get to it")
}
