% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotproj}
\alias{plotproj}
\title{plotproj aids the polotting of a projected variable}
\usage{
plotproj(invar, varlabel, plotconst, vline = NULL, iters = 0, addqnts = FALSE)
}
\arguments{
\item{invar}{the array containing the year x SAU x reps values for a selected
variable out of sauzoneDP}

\item{varlabel}{the label to place along the outer Y-axis}

\item{plotconst}{a list containing nsau, saunames, reps, projyrs, and plts
a vector of two describing the layout of plots, default=c(4,2)}

\item{vline}{default=NULL, which means nothing extra is plotted. If given a
year then a vertical line in red will be added to the plot}

\item{iters}{default=0, which means all iterations will be plotted. If iters
has a value then only that many trajectories will be plotted}

\item{addqnts}{will calculate and add the median and 90 percent quantiles}
}
\value{
nothing but it does generate a plot
}
\description{
plotproj plots out the projections from teh mse for a selected
    variable. iters is included so that details of a few trajectories can
    be viewed as well as seeing all replicates.
}
\examples{
print("wait on data files")
}
