% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{plotrecruitment}
\alias{plotrecruitment}
\title{plotrecruitment literally plots stock-recruitment relationships}
\usage{
plotrecruitment(rundir, resultpop, glb)
}
\arguments{
\item{rundir}{the scenarios results directory}

\item{resultpop}{a summary of the biological properties of each population
generated in the biology_plots function.}

\item{glb}{the globals object}
}
\value{
nothing but it does save two figures and a table into rundir
}
\description{
plotrecruitment plots an approximation to the summarized
    stock recruitment relationship for each sau together on the same plot.
    It also plots the stock-recruitment relationship for each population
    within each sau, which illustrates the distribution of the recruitment
    total at the sau level distributed relative to yield from the areas of
    persistent production results. Finally, it tabulates the sau results.
}
\examples{
print("wait on data sets")
# rundir=rundir; resultpop=resultpop; glb=glb
}
