% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{plotsau}
\alias{plotsau}
\title{plotsau generates nSAU plots in one figure for a given variable}
\usage{
plotsau(
  invar,
  glb,
  plots,
  ylab = "Catch",
  xlab = "Year",
  col = "grey",
  medcol = "red",
  addCI = FALSE,
  CIcol = "blue",
  CIprobs = c(0.05, 0.5, 0.95)
)
}
\arguments{
\item{invar}{the three dimensional array containing the variable to plot}

\item{glb}{the global object}

\item{plots}{the number of plots.eg for 8 plots one might put c(4,2)}

\item{ylab}{the prefix for the Y labels, which will have the SAU number
added to it. Default="Catch"}

\item{xlab}{The single X label for the bottom of the plot, default="Year"}

\item{col}{the colour for the reps replicate projection lines, default="grey"}

\item{medcol}{the colour for the median line, default="red", set to 0 or NA
to avoid adding to the plots.}

\item{addCI}{should quantile confidence bounds be included on the plots,
default=FALSE}

\item{CIcol}{if CI are to be included let their colour be this, default=blue}

\item{CIprobs}{the CI quantiles. defaults are c(0.05,0.95), the inner 90 perc}
}
\value{
invisibly the 5, 50, and 95 percent quantiles by year for each SAU
}
\description{
plotsau summarizes a single variable from the dynamic object
    produced by the replicate projections by generating nSAU plots in a
    single figure. These are all lineplots in grey (change using col), with
    the median in red (change using medcol).
}
\examples{
print("wait on appropriate internal data sets")
}
