% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{plotscene}
\alias{plotscene}
\title{plotscene literally plots up the output from comparevar}
\usage{
plotscene(
  scenquant,
  glbc,
  var = "cpue",
  ymin = 0,
  filen = "",
  legloc = "topleft",
  legplot = 1,
  qnt = 90
)
}
\arguments{
\item{scenquant}{a list of the quantiles for the given variable from each
scenario}

\item{glbc}{a list of the global objects from each scenario being compared}

\item{var}{what variable from the dynamics to summarize, valid names include
catch, acatch, cpue, harvestR, desplsB, depleB, recruit, matureB, and
exploitB, default = 'cpue'}

\item{ymin}{allows one to set the lower limit to the yaxis, default = 0}

\item{filen}{a filename for use if saving the output to said file,
default = '', which implies the plot goes to the console.}

\item{legloc}{legend location, default='topleft'}

\item{legplot}{which plot should contain the legend, default=1}

\item{qnt}{which quantile to plot? defauilt = 90 percent}
}
\value{
nothing but it does generate a plot with nsau panels
}
\description{
plotscene takes the output from comparevar and plots the
    quantiles relative to each other.
}
\examples{
 print("wait on internal datasets")
 #
}
