% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{plotsceneproj}
\alias{plotsceneproj}
\title{plotscenproj plots a 3D array of data by sau in a single plot}
\usage{
plotsceneproj(
  rundir,
  inarr,
  glb,
  scene,
  filen = "",
  label = "",
  maxy = 0,
  Q = 90,
  hline = NA
)
}
\arguments{
\item{rundir}{the directory in which all results are held for a scenario or
comparison of scenarios}

\item{inarr}{the 3D array of projections for a given scenario derived from
catch or cpue or whatever}

\item{glb}{the global object relating to the particular acenario}

\item{scene}{the scenario name}

\item{filen}{the filename in which to store the plot, default="" which draws
the plot to the console}

\item{label}{what is the name of the variable being plotted, default=""}

\item{maxy}{what constant maximum y-axis to use? default=0 which uss the
maximum for each sau}

\item{Q}{which quantile to use, default = 90. If set to 0 no median or
quantiles are plotted, 95 will lead to the 95 percent quantiles being
plotted}

\item{hline}{should a horizontal dashed line be included. default=NA, which
means that no line is added. Otherwise whichever value is given this
argument will lead to a dashed horizontal black line of width 1.}
}
\value{
invisibly a list of the quantiles for each sau
}
\description{
plotsceneproj uses a 3D array of years x sau x replicates to
    produce a plot by sau of those projections all in a single plot. Options
    exist to include a horizontal dashed line and to include a median line
    and either 90 or 95 quantiles across the plot.
}
\examples{
print("wait on data sets")
#  rundir=rundir; inarr=cdivmsy[[1]];glb=glbc[[1]];scenes=scenes[1];
#  filen="";label="Catch / MSY";maxy=0
}
