% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotsizecomp}
\alias{plotsizecomp}
\title{plotsizecomp generates a plot of available size composition data}
\usage{
plotsizecomp(
  rundir,
  incomp,
  SAU = "",
  lml = NULL,
  catchN = NULL,
  start = NA,
  proportion = TRUE,
  console = TRUE,
  width = 10,
  height = 9,
  fnt = 7,
  tabcategory = "predictedcatchN"
)
}
\arguments{
\item{rundir}{the directory for all ctrl, data, and output files.}

\item{incomp}{filtered size-composition data from the fishery}

\item{SAU}{the name of the SAU used to label the y-axis}

\item{lml}{a vector of the lml in each year represented by incomp}

\item{catchN}{the predicted size-composition of the catch}

\item{start}{which size-class to start from, default=NA, which means all
size-classes in the samples will be used.}

\item{proportion}{should the plots be as proportions or counts,default=TRUE}

\item{console}{should the plot be sent ot the console or saved to rundir.
default=TRUE}

\item{width}{the width of the plot, default = 10}

\item{height}{the height of the plot, default = 9}

\item{fnt}{what font to use, default font = 7, bold times}

\item{tabcategory}{what name to give to the webage tab when the plots are
saved. Default="predictedcatchN"}
}
\value{
nothing but it does generate a plot
}
\description{
plotsizecomp generates a plot of available size composition
    data with the option of plotting the predicted size-composition of
    catch on top of it; this latter option is only possible when plotting
    the distributions as proportions.
}
\examples{
print("wait on data sets")
}
