% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotTAC}
\alias{plotTAC}
\title{plotTAC plots the projected TAC across all replicates}
\usage{
plotTAC(zoneDP, rundir, CIprobs = c(0.05, 0.5, 0.95))
}
\arguments{
\item{zoneDP}{the dynamic zone projection object}

\item{rundir}{the rundir for the scenario}

\item{CIprobs}{the quantiles used to generate the confidence intervals. The
default = c(0.05,0.5,0.95)}
}
\value{
nothing but it adds a plot to the zonescale tab in the rundir
}
\description{
plotTAC plots out the projected TAC for all replicates. The
    first value is the sum of the final catches in the conditioning period.
}
\examples{
print("wait on suitable datasets")
}
