% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotZone}
\alias{plotZone}
\title{plotZone plots the projected TAC across all replicates}
\usage{
plotZone(inzone, rundir, CIprobs = c(0.05, 0.5, 0.95), addfile = TRUE)
}
\arguments{
\item{inzone}{the dynamic zone projection object}

\item{rundir}{the rundir for the scenario}

\item{CIprobs}{the quantiles used to generate the confidence intervals. The
default = c(0.05,0.5,0.95)}

\item{addfile}{should the plots be added to the rundir =TRUE or only sent to
the console = FALSE. Default = TRUE}
}
\value{
nothing but it adds a plot to the zonescale tab in the rundir
}
\description{
plotZone plots out the projected TAC for all replicates. The
    first value is the sum of the final catches in the conditioning period.
}
\examples{
print("wait on suitable datasets")
}
