% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{plotzonedyn}
\alias{plotzonedyn}
\title{plotzonedyn ribbonplot of zone catch, mature biomass, harvestR and CE}
\usage{
plotzonedyn(
  rundir,
  scenes,
  zone,
  glbc,
  console = TRUE,
  q90 = TRUE,
  polys = TRUE,
  intens = 127,
  hlines = NULL,
  scencol = NULL
)
}
\arguments{
\item{rundir}{the directory in which all results are held for a scenario or
comparison of scenarios}

\item{scenes}{the names of the different scenarios being compared}

\item{zone}{a list of the out$outzone objects from each chosen scenario}

\item{glbc}{all of the global objects from out. They should all have the
same reps and years but this will be checked.}

\item{console}{should the plot be sent to the console or saved for use in the
web-page output? default=TRUE, set to FALSE to save it}

\item{q90}{should the 90th quantiles be plotted? default=TRUE. If FALSE then
the 95th quantiles are used.}

\item{polys}{should transparent polygons be plotted = TRUE, or lines = FALSE}

\item{intens}{if polys=TRUE then intens signifies the intensity of colour on
a scale of 0 - 255. 127 is about 50 percent dense.}

\item{hlines}{should reference lines be drawn on each plot. default=NULL
which = no lines if lines are required then hlines needs to be set to a
list of h values for each scenario, for each plotted variable, see
examples for syntax}

\item{scencol}{what sequence of colours should each scenario have. The default=NULL
which implies the colour will reflect the sequence in which they are
read in.}
}
\value{
nothing but it does generate a plot and saves it if console=FALSE
}
\description{
plotzonedyn generates ribbonplots of the zone-wide total catches,
    the mature biomass, the harvest rate, and the CPUE across the years of
    projection.
}
\examples{
print("wait on data sets")
# syntax plotzonedyn(rundir,scenes,zone,glbc,console=FALSE,q90=TRUE,
#                    polyd=TRUE,intens=100,hlines=list(catch=outprod[,"MSY"],
#                    spawnB=outprod[,"Bmsy"],harvestR=0,
#                    cpue=outprod[,"CEmsy"]))
}
\seealso{
{
  \link{poptozone}, \link{plotZone}
}
}
