% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{plotzoneyrtovar}
\alias{plotzoneyrtovar}
\title{plotzoneyrtovar plots the time to invar at a zone scale}
\usage{
plotzoneyrtovar(rundir, invar, varname, glbc, category, console = TRUE)
}
\arguments{
\item{rundir}{the directory in which all results are held for a scenario or
comparison of scenarios}

\item{invar}{a list of the zone-scale replicate runs for the varname from
each scenario. the zone object is generated inside do_comparison}

\item{varname}{just the name of the variable being plotted, to make sure the
figures all have the correct labelling}

\item{glbc}{the list of the global objects generated within do_comparison}

\item{category}{if a file is saved this argument identifies the webpage tab
into which the plot will be placed.}

\item{console}{should the plot be sent to the console or saved for use in the
web-page output? default=TRUE, set to FALSE to save it}
}
\value{
nothing but it does generate a plot of nscen histograms
}
\description{
plotzoneyrtovar it is common to want to know how long each
    replicate in a scenario run takes to be >= to given value, eg when is the
    catch/msy ratio first = 1. The zonecatchvsmsy function generates that
    catch/msy information, this plots the histograms. Other varnames such
    as matureB would be of interest when compared to Bmsy.
}
\examples{
print("wait on example data")
# syntax  plotzoneyrtovar(rundir=rundir,invar=zyrtomsy,varname="MSY",
#                         glbc=glbc,category="C_vs_MSY",console=TRUE)
}
\seealso{
{
   \link{whenfirst},  \link{zonecatchvsmsy}, \link{do_comparison}
}
}
