% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{popgrowth}
\alias{popgrowth}
\title{popgrowth plots implied size-at-age for each population for each sau}
\usage{
popgrowth(rundir, zoneC, glb, console = TRUE, maxage = 30, startsize = 2)
}
\arguments{
\item{rundir}{the full path to the directory in which all files relating to a
particular run are to be held.}

\item{zoneC}{the zoneC object from makezoneC}

\item{glb}{the globals object}

\item{console}{should the plot be sent to the console or saved as a file.
default = TRUE = to the console}

\item{maxage}{What maximum age to use, default = 30}

\item{startsize}{what starting size to use, default = 2.0}
}
\value{
a list of the implied size-at-age in a matrix for each sau,
    containing the details for each population in each sau.
}
\description{
popgrowth uses the defined growth parameters and a starting size
    of 2mm and increments the shell length for 30 years to describe the
    implied size-at-age for each population. Both plots and tables are
    produced.
}
\examples{
print("wait on data")
# rundir=rundir;zoneC=zoneC;glb=glb; console=TRUE;maxage=30;startsize=2
}
