% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{poplevelplot}
\alias{poplevelplot}
\title{poplevelplot generates population x replicate trajectory plots}
\usage{
poplevelplot(rundir, sau, popvar, glb, label = "", console = TRUE)
}
\arguments{
\item{rundir}{the scenario's directory for datafiles and results}

\item{sau}{which sau is to be worked on. In Tas = 1:8 in the west}

\item{popvar}{which 3D array population variable is to be used. This will be
generally derived from zoneDP, eg zoneDP$catch or zoneDP$depleB}

\item{glb}{the globals object used for sauindex, saunames, hyrs, and reps}

\item{label}{the sau name will be added to this label for the y-axis, this
should be simple, such as 'catch' or have words connected by hyphens or
underscores as it will also be added to the filename.}

\item{console}{should the plot go to the console or be saved into rundir.
default=TRUE = to the console}
}
\value{
invisibly the population medians
}
\description{
poplevelplot is used to produce population level plots of an
    input variable derived from zoneDP. It produces a single plot for a given
    sau which depicts each population's replicate trajectories, all in grey.
    However, medians of the variable through each year are calculated and
    plotted in distinct colours for each population up to a maximum of 12
    populations per sau. If more than 12 then the population index 1:npop is
    used. These plots should be considered as a partial diagnostic for how
    well the implied fleet dynamics is distributing catch among an sau's
    populations. The trajectories are plotted from the last year of
    conditioning to the last year of projections.
}
\examples{
print("poplevelplot(rundir=rundir,sau=7,popvar=zoneDP$catch,glb=glb,")
print("label='Catch',console=TRUE")
}
