% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{popNAStosau}
\alias{popNAStosau}
\title{popNAStosau converts NAS data by population to NAS x sau}
\usage{
popNAStosau(popNAS, glb)
}
\arguments{
\item{popNAS}{either the catchN or the Nt from zoneDD, that is the zone
after the equilibrium zone (zoneD) has had the historical catches
applied.}

\item{glb}{the globals object. Needed for the sauindex and various array
dimnesion labels.}
}
\value{
An Nclass x nyrs x nsau array of numbers-at-size data
}
\description{
popNAStosau converts numbers-at-size data x population into
    the same by SAU. This is designed to compare ny observed NAS data from
    a zone with those predicted by the conditioned model, it could also be
    used to improve the conditioning code to avoid fitting only to CPUE
}
\examples{
 #e.g.  popNAStosau(out$zoneDD$catchN,out$glb)
}
