% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{popsdefine}
\alias{popsdefine}
\title{popsdefine makes a parameter vector, popdef, for each population}
\usage{
popsdefine(const, glob)
}
\arguments{
\item{const}{a matrix with rows containing the constants defined for
each of the biological parameters. The columns of the matrix
reflect values for each populations from readpopdatafile.}

\item{glob}{the globals object from readctrlfile}
}
\value{
a matrix with a row for each population and whose columns
    are parameters defining the biological parameters for each population.
}
\description{
popsdefine makes a parameter vector, popdef, for each
    population. However, instead of imputing random variation the input
    matrix of constants is already defined by population. This contrasts with
    the function definepops which is used if saudata is read instead of
    population level data.
}
\examples{
\dontrun{
  data(zone1)
  glb <- zone1$globals
}
}
\seealso{
{
  \link{definepops}, \link{readpopdatafile}, \link{readsaudatafile}
}
}
