% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{poptosau}
\alias{poptosau}
\title{poptosau combines population cpue into sau as catch weighted sums}
\usage{
poptosau(catvect, cpuevect, sauindex)
}
\arguments{
\item{catvect}{the vector of catches x population for a given year}

\item{cpuevect}{the vector of cpue x population for a given year}

\item{sauindex}{the sau indices of each population}
}
\value{
a list of saucpue and saucatch
}
\description{
poptosau combines the cpue from separate populations into their
    respective sau using a catch-weighted strategy. The sauindex is used to
    identify which populations to apply the sau total catches to.
}
\examples{
print("wait on appropriate built-in data files")
}
