% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{poptosauCE}
\alias{poptosauCE}
\title{poptosauCE combines population cpue into sau as catch weighted sums}
\usage{
poptosauCE(catvect, cpuevect, sauindex)
}
\arguments{
\item{catvect}{the vector of catches x population for a given year}

\item{cpuevect}{the vector of cpue x population for a given year}

\item{sauindex}{the sau indices of each population}
}
\value{
a list of saucpue and saucatch
}
\description{
poptosauCE combines cpue from separate populations into their
    respective sau using a catch-weighted strategy. The sauindex is used to
    identify which populations to apply the sau total catches to.
}
\examples{
print("wait on appropriate built-in data files")
# catvect=zoneDD$catch[1:finalyr,]; cpuevect=zoneDD$cpue[1:finalyr,]
}
