% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{predsizecomp}
\alias{predsizecomp}
\title{predsizecomp plots size-composition for each replicate in a subset of years}
\usage{
predsizecomp(
  sau,
  NSC,
  glb,
  minSL = 10,
  interval = 5,
  prop = TRUE,
  console = TRUE,
  rundir = "",
  omit = 5
)
}
\arguments{
\item{sau}{which sau to plot, the index number not its name, in TAS western
zone this would be between 1 : 8}

\item{NSC}{the numbers-at-size matrix from sauout, either Nt or catchN}

\item{glb}{the globals object}

\item{minSL}{the minimum shell length to be plotted. This constrains each
plot so that details can be clearer. default=10mm which implies all
sizes above the post-larval sizes.}

\item{interval}{default=5 which selects the year just before projections and
then the data for every interval years along the projection period. The
years selected always includes the last year of projection and the year
before projections started.}

\item{prop}{default=TRUE. Should proportions be plotted or actual numbers}

\item{console}{should the plot go to the console or will it be saved}

\item{rundir}{the directory into which all results are placed. default="" but
generally this would = rundir}

\item{omit}{default = 5. how many size classes should be omitted from the
start of each compositions plot. The default = 5 which in Tasmania,
which has 2mm size classes starting at 2mm, this omits all animals less
than 10mm.}
}
\value{
Nothing but it does plot a graph.
}
\description{
predsizecomp generates plots the size composition data for a
    subset of the projection years (determined by the interval argument).
    All plots omit the first five size classes so that the post-settlement
    spikes do not obscure the rest of the plots. The minSL (minimum shell
    length argument) determines the left-most size in the plots to allow for
    details to be made clearer.
}
\examples{
print("wait on suitable data")
}
