% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{prepareDDNt}
\alias{prepareDDNt}
\title{prepareDDNt converts the population based historical Nt to SAU-based}
\usage{
prepareDDNt(inNt, incatchN, glb)
}
\arguments{
\item{inNt}{the 3 dimensional array of population numbers-at-size}

\item{incatchN}{the 3 dimensional array of catch numbers-at-size}

\item{glb}{the global constants object}
}
\value{
a list of Nt and catchN x SAU
}
\description{
prepareDDNt processes the zoneDD, following the conditioning on
    the historical fishery data, so that the population numbers-at-size are
    converted to SAU-based numbers-at-size. It does this for both Nt and
    catchN. This is to allow analysis, tabulation, and plotting at the SAU
    scale. Input data is 3D  Nclass x years x numpop
}
\examples{
print("wait on suitable internal data-sets")
}
