% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{prepareforfis}
\alias{prepareforfis}
\title{prepareforfis generates WtL and Sel for a FIS if one occurs}
\usage{
prepareforfis(fisindexdata, constants, glb, fispar)
}
\arguments{
\item{fisindexdata}{the FIS index of relative abundance obtained from the
condC object}

\item{constants}{the constants object, which is used to determine the
weight-at-length parameters of the SAu from teh populations.}

\item{glb}{the globals object}

\item{fispar}{the fispar containing selectivity parameters fisL50, delta,
and the fisLML for the historical years}
}
\value{
a list of WtL by sau with fis data, selectivity by year, yearfis
    the range of years with fis, and fissau a vector of sau in fisindexdata.
}
\description{
prepareforfis in the unlikely event that one has a FIS for at
    least some of the SAU in the abalone zone, then it will need special
    treatment. Whether it should be included in the MSE and more importantly
    in the HCR is a matter for a jurisdictions fishery policy and each
    scientist's conscience. Currently, the implementation assumes that the
    FIS index represents the exploitable biomass (using the FIS Selectivity)
    in a linear fashion. So, spatial issues and hyperstability in the
    sampling are omitted (did I mention the scientist's conscience?).
}
\examples{
print("wait on suitable internal data sets")
# fisindexdata=condC$fisindexdata; fispar=hsargs$fispar
}
