% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{prepareprojection}
\alias{prepareprojection}
\title{prepareprojection high level function that sets up a projection}
\usage{
prepareprojection(
  projC = projC,
  condC = condC,
  zoneC = zoneC,
  glb = glb,
  zoneDD = zoneDD,
  ctrl = ctrl,
  varyrs = varyrs,
  calcpopC = calcpopC,
  lastsigR = 0.3
)
}
\arguments{
\item{projC}{the projection object from readctrlfile}

\item{condC}{historical conditioning data}

\item{zoneC}{the constant part of the zone}

\item{glb}{the global variables}

\item{zoneDD}{the zone after initial depletion through conditioning on the
fishery}

\item{ctrl}{the ctrl object for the scenario run}

\item{varyrs}{how many years at the end to add recruitment variation}

\item{calcpopC}{a function that takes the output from hcrfun and gernerates
the actual catch per population expected in the current year.}

\item{lastsigR}{recruitment variation for when it is applied for varyrs}
}
\value{
a list of the dynamic zone object as a list of arrays of projyrs x
    populations x replicates, plus the revised projC and revised zoneC
}
\description{
prepareprojection is a high level function that restructures
    zoneC by including the projected selectivity (in case the LML is set to
    change during the projections), and also selectivity x weight-at-size
    (for computational speed). It then generates a new zoneD with room for
    all replicates as well as the aspirational catches from each HS. It does
    this by converting the arrays of year x pop, to year x pop x replicate.
    It then uses the conditioned data in zoneDep to predict the first
    aspirational catches for the projections and conducts the initial
    replicate, thus starting the application of the HS.
    Finally, it adds recruitment variation
    to each of the replicates and keeps the last year of each iteration of
    the addrecvar function as the start of each replicate projection, with
    zeros in the catch, cpue, and cesau arrays
}
\examples{
print("wait on data files")
}
