% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{print.zone}
\alias{print.zone}
\title{print.zone S3 method for printing a summary of a given zone}
\usage{
\method{print}{zone}(x, ...)
}
\arguments{
\item{x}{a zone object}

\item{...}{in case there are extra parameters}
}
\value{
nothing, but does print out a summary of the zone
}
\description{
print.zone S3 method for printing a summary of a given zone
}
\examples{
\dontrun{
txt1 <- 'all example code should be able to run'
}
}
