% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docondition.R
\name{printline}
\alias{printline}
\title{printline literally prints a selected line from a text file}
\usage{
printline(rundir, datafile, linenumber = 29)
}
\arguments{
\item{rundir}{the scenario directory holding all files for a scenario}

\item{datafile}{the name of the datafile for the scenario, from the ctrl
object}

\item{linenumber}{which linenumber to print, default =29}
}
\value{
nothing but it does print a line to the console
}
\description{
printline is used to check that one has selected the
    correct line for modification inside a datafile. This is done when
    conditioning the MSE operating model on the parameter AvRec
}
\examples{
print("wait on suitable example")
}
