% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docondition.R
\name{prodforpop}
\alias{prodforpop}
\title{prodforpop plots the productivity characteristics for a population}
\usage{
prodforpop(rundir, inprod, pop, console = TRUE)
}
\arguments{
\item{rundir}{the full path to the scenario directory}

\item{inprod}{the production matrix for a single population, for example,
out$production[,,2] is the second population's production properties.}

\item{pop}{which population is being examined? A single integer}

\item{console}{should the plot be saved to a png file or go to the console.
the default = TRUE plotting to theconsole}
}
\value{
nothing but it does generate a plot
}
\description{
prodforpop plots the productivity characteristics for a
    population. This is especially useful for a detailed consideration of
    an SAU's dynamics. This also allows one to check that the harvest rate
    range used to estimate production runs past the MSY level.
}
\examples{
# syntax prodforpop(rundir=rundir,inprod=produciton[,,2],pop=2,console=FALSE)
# rundir=""; inprod=out$production[,,2]; pop=2; console=TRUE
}
