% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{projectiononly}
\alias{projectiononly}
\title{projectiononly converts hyrs+pyrs_sau_reps to pyrs_sau_reps array}
\usage{
projectiononly(x, glb)
}
\arguments{
\item{x}{a 2D, 3D or 4D array of dimension (hyrs + pyrs) x reps,
nyrs x nsau x reps, or Nclass x nyrs x reps}

\item{glb}{the globals object. Only hyrs and pyrs are used}
}
\value{
an array truncated appropriately to include only the projection years
}
\description{
projectiononly is a utility function that takes an output 3D
    array from aMSE with dimensions allyears x nsau x reps and truncates it
    so that only the projection years remain, pyrs x nsau x reps, to simplify
    plotting and tabulating results. This can be used within both aMSE and in
    do_comparison of scenarios.
}
\examples{
print("wait on example data sets")
# x <- catch[[1]];glb <- glbc[[1]]
}
