% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{putNA}
\alias{putNA}
\title{putNA can add NAs to the start and end of a vector}
\usage{
putNA(x, pre, post)
}
\arguments{
\item{x}{the vector to which NAs are to be added}

\item{pre}{how many NA to be added the front}

\item{post}{how many to be added to the end}
}
\value{
a vector of length(x) + pre + post
}
\description{
putNA fulfils a common requirement to expand a vector with NAs
    to assist a plot by ensuring the x and y vectors are the same length.
    It can accept both character and numeric vectors. This used to be called
    addNA, but there is a function in base R with that name that does
    something completely different.
}
\examples{
vect <- rnorm(10,mean=5,sd=1)
putNA(vect,3,5)
}
