% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{read_conddata}
\alias{read_conddata}
\title{read_conddata reads datasets used to condition the operating model}
\usage{
read_conddata(filename)
}
\arguments{
\item{filename}{the filename of the text file containing the data to be used
in the conditioning.}
}
\value{
a matrix of reported catches for years vs SAU
}
\description{
read_conddata facilitates the reading of dataset used to
    condition the operating model. It is used in all aspects of conditioning,
    including attempting to model the productivity of each SAU using either
    production models or size-structured models. Once conditioned the catches
    can also be used to generate the initial depletion of the operating
    model, for scenarios which attempt to closely simulate the dynamics of an
    actual fishery. Currently, the function only reads in time-series of
    catches for each SAU across the years of the fishery. The current format
    for the input csv file should have the fishery name in the first line,
    then a heading NYRS with the number of years following that, separated by
    a comma, then another heading YEARS, sau1, sau2, sau3,...,saulast,
    followed by nyrs lines containing the year and catches for each SAU
}
\examples{
print("set up a csv file with the format given in the description")
print("then call condat <- read_conddata('yourname.csv')) to read the data")
}
