% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{readctrlfile}
\alias{readctrlfile}
\title{readctrlfile reads the rundir scenario's control file}
\usage{
readctrlfile(rundir, infile = "control.csv", verbose = TRUE, fisreadfun = NULL)
}
\arguments{
\item{rundir}{the directory in which all files relating to a
particular run are to be held.}

\item{infile}{default="control.csv", the filename of the control
file present in rundir containing information regarding the
run.}

\item{verbose}{Should progress comments be printed to console, default=TRUE}

\item{fisreadfun}{a function to read in the fissettings.csv file if a FIS
is being used.}
}
\value{
the control list for the run
}
\description{
readctrlfile checks rundir contains the required csv
    files including the named control file, which then contains
    the names of the region data file, and the population data
    file. The run stops if any are not present or are misnamed.
}
\examples{
\dontrun{
# this has sinoce been modified and needs updating
rundir <- tempdir()
ctrlzonetemplate(rundir)
datafiletemplate(6,rundir,filename="zone1sau2pop6.csv")
ctrl <- readctrlfile(rundir)
print(ctrl)
}
}
