% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{readdatafile}
\alias{readdatafile}
\title{readdatafile reads in a matrix of data defining each population}
\usage{
readdatafile(numpop, indir, infile)
}
\arguments{
\item{numpop}{the total number of populations across the zone}

\item{indir}{directory in which to find the date file}

\item{infile}{character string with filename of the data file}
}
\value{
a matrix of values defining the PDFs used to define the
    properties of each population. The contents of popdefs
}
\description{
readdatafile expects a matrix of probability density
    function definitions that are used to define the populations
    used in the simulation. These constitute the definition of
    popdefs.
}
\examples{
\dontrun{
data(zone1)
glb <- zone1$globals
glb
data(constants)
constants
ctrlfile <- "control.csv"
ctrl <- readctrlfile(glb$numpop,rundir,ctrlfile)
reg1 <- readzonefile(rundir,ctrl$zonefile)
popdefs <- readdatafile(reg1$globals,rundir,ctrl$datafile)
print(popdefs)
}
}
