% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{readdatafileold}
\alias{readdatafileold}
\title{readdatafile reads in the constants and matrices of data required}
\usage{
readdatafileold(infile)
}
\arguments{
\item{infile}{a character string containing the filename of the data file}
}
\value{
Twenty different objects, some matrices, others single numbers
\itemize{
  \item constants the array of variables as means and stdev used to
      define the variation within each of the populations used to define
      a zone
  \item projLML the LML expected in each block in each projection year;
      a matrix
  \item histLML LML imposed in each block in each historical year used
      in the conditioning
  \item randomseed literally the randomseed used to ensure that each
      simulation run starts in the same place
  \item catches catches in each block in each year during the
      conditioning
      period
  \item geomCE a matrix of the geometric mean CPUE for each block in each
      year of the conditioning period
  \item standCE a matrix of the standardized mean CPUE for each block in
      each year of the conditioning period
  \item effort a matrix of the effort imposed for each block in each year
      of the conditioning period
  \item nblock a scaler defining the number of blocks simulated
  \item blockNames a vector of character with the names of each block
  \item blkpop defines the number of populations within each block
  \item numpop scaler defining the total number of populations in zone
  \item minc scaler defining the mid point of the smallest size class
  \item cw scaler defining the class width
  \item Nclass scaler defiing the number of size classes
  \item midpts vector of midpoints calculated from minc, cw, and Nclass
  \item prodfile filename for zone production data; used in zoneStart
  \item projCatch a matrix of catches expected in each block in each
     projection year
}
}
\description{
given a filename readdatafile will expect a given format.
    Each required section starts with the section names in capitals. The
    required sections are: SIZECLASS, BLOCKNAMES, PDFs, CONDITIONING,
    PROJLML, HISTLML, YEARS, RANDOMSEED, PRODUCTIVITY, CATCHES,
    PROJCATCH, CPUE, STANDARDIZED, EFFORT. An example data file is
    included in the package to illustrate the required formatting for the
    data
}
\examples{
\dontrun{
#need to have an example data file before I can run an example
filename <- datafiletemplate(numblock=3,filename="block3.csv")
condDat <- readdatafile(filename)
str(condDat,max.level=1)
}
}
