% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{readhcrfile}
\alias{readhcrfile}
\title{readhcrfile literally reads a csv file for controling the AbMSE}
\usage{
readhcrfile(infile)
}
\arguments{
\item{infile}{the filename of the control file}
}
\value{
a list object containing the contro variables
}
\description{
readhcrfile complements the readdatafile. The MSE requires
    a data file to condition the operating model but it also needs a
    control file to setup the details of the simulation test being
    conducted. See ctrlfileTemplate./
}
\examples{
\dontrun{
  print("Still to be developed.")
}
}
