% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{readpopdatafile}
\alias{readpopdatafile}
\title{readpopdatafile reads in a matrix of data defining each population}
\usage{
readpopdatafile(indir, infile)
}
\arguments{
\item{indir}{directory in which to find the date file = rundir}

\item{infile}{character string with filename of the data file}
}
\value{
a matrix of values defining the PDFs used to define the
    properties of each population. The contents of popdefs
}
\description{
readpopdatafile expects a matrix of constants defining the
    biological properties used to define the populations
    used in the simulation. These constitute the definition of
    popdefs.
}
\examples{
print("wait on data sets")
}
