% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{readsaudatafile}
\alias{readsaudatafile}
\title{readsaudatafile generates the constants matrix from sau data}
\usage{
readsaudatafile(rundir, infile)
}
\arguments{
\item{rundir}{the directory in which the data file is to be found. This will
usually be the rundir for the scenario run}

\item{infile}{the name of the specific datafile used.}
}
\value{
a list of the constants matrix with values for each population and
    the original matrix of sau values from readsaudatafile
}
\description{
readsaudatafile uses data described at the SAU level to make
    the constants file, which is then used to generate the popdefs matrix
    containing the specific values. This change from explicitly defining
    each population has been implemented to simplify the conditioning of
    each operating model.
}
\examples{
print("wait on suitable data sets")
}
