% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{readzonefile}
\alias{readzonefile}
\title{readzonefile reads in the constants for the zone}
\usage{
readzonefile(indir, infile)
}
\arguments{
\item{indir}{directory in which to find the zone file}

\item{infile}{character string with filename of the zone file}
}
\value{
15 objects relating to constnats for the zone
\itemize{
  \item SAUnames the labels given to each SAU
  \item SAUpop the number of populations in each SAU, in sequence
  \item minc scaler defining the mid point of the smallest size class
  \item cw scaler defining the class width
  \item randomseed used to ensure that each simulation run starts
      in the same place. Could use getseed() to produce this.
  \item projyrs, he number of years of projection
  \item outyear a vector of three, with projyrs, fixyear, and firstyear
  \item projLML the LML expected in each projection year, a vector
  \item HS the name of the harvest strategy to be used in the projection
  \item condition the number of years of historical catch data
  \item histCatch the years if historical catch and historical LML
  \item histyr the actual years for which SAU catch data available
  \item histCE the historical cpue data for each SAU
  \item yearCE the actual years fo which cpue data available
  \item globals a list containing numpop, nSAU, midpts, Nclass, and
      Nyrs
}
}
\description{
with the zone filename from the control file the
    readzonefile will read the data from a csv file arranged with
    a standard layout. Once again this uses the utility functions
    for reading and parsing lines of text. This is illustrated by
    the function makezonefile, which produces an example *.csv
    file that can then be customized to suit your own simulations.
    Each required section contains a series of constants which are
    read in individually,so their labels are equally important. This
    important function reads in the details of any conditioning and also
    of any projections.
}
\examples{
\dontrun{
datadir <- "./../../rcode2/aMSE/data-raw/"
ctrlfile <- "control.csv"
ctrl <- readctrlfile(datadir,ctrlfile)
reg1 <- readzonefile(datadir,ctrl$zonefile)
str(reg1)
}
}
