% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{region1}
\alias{region1}
\title{region1 the constants common to a region}
\format{A list of constants that are uniform across all populations
    in a region
\describe{
  \item{SMUnames}{the names of each SMU}
  \item{SMUpop}{a vector of how many populations in each SMU}
  \item{minc}{the mid-point of the minimum size class}
  \item{cw}{the size-class width in mm}
  \item{larvdisp}{the rate of larval dispersal as a proportion}
  \item{randomseed}{used if results need repeating}
  \item{outyear}{a vector on Nyrs, fix year, and start year}
  \item{projLML}{the time series of LML used in the MSE projections}
  \item{histLML}{time series of LML if conditioning the model prior
      to projections}
  \item{globals}{a list of global constants, containing numpop,
      nSMU, midpts, Nclass, Nyrs, and larvdisp}
}}
\description{
region1 contains the constants relating to the whole
    region rather than th epopulations. See th elisting below.
}
\examples{
 data(region1)
 print(region1)
}
