% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{regionfiletemplate}
\alias{regionfiletemplate}
\title{regionfiletemplate generates a template region file}
\usage{
regionfiletemplate(filename = "region1.csv")
}
\arguments{
\item{filename}{the name for the generated region file, a character
string that defaults to region1.csv. It is best to include the
complete path}
}
\value{
a standard definition region file
}
\description{
regionfiletemplate generates a standardized region file
    template containing constants tha toperate at a regional level.
    Generate this and then modify the contents to suit
    the system you are attempting to simulate.
}
\examples{
 yourdir <- tempdir()
 regionfile <- filenametopath(yourdir,"region1.csv")
 regionfiletemplate(filename=regionfile)   #
 region1 <- readregionfile(yourdir,"region1.csv")
 str(region1,max.level=1)
}
