% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{replaceVar}
\alias{replaceVar}
\title{replaceVar replaces values of a variable in the input datafile}
\usage{
replaceVar(infile, invar, newval)
}
\arguments{
\item{infile}{the name, and path,  of the data file to be altered}

\item{invar}{the text name of the variable to be changed}

\item{newval}{the new value with which to replace the current values}
}
\value{
The function over-writes the original file but saves the original
    by adding an '_old' to the end of the original filename.
}
\description{
replaceVar replaces the values of a variable in the input
    datafile. TAKE CARE, it overwrites the original file! However, it
    also saves the original file by adding an '_old' to the filename.
    Obviously if the function is used multiple times the original
    'original' file will be over-written on the second use.
    Alternatively, one could readin the datafile using readdateFile and
    then directly alter the condDat, e.g., in the case of a four block
    zone in which the AvRec vale is changed: condDat$constants["AvRec",]
    <- c(12.5,12.2,12.4,12.1)
}
\examples{
\dontrun{
filename <- datafiletemplate(numblock=1,filename="oneblock.csv")
replace(filename,"AvRec",15.75)
condDat <- readdatafile(filename)
print(round(condDat$constants,4))
}
}
