% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{resetexB0}
\alias{resetexB0}
\title{resetexB0 resets the unfished exploitable biomass at time zero}
\usage{
resetexB0(zoneC, zoneD)
}
\arguments{
\item{zoneC}{the constant components of the simulated zone}

\item{zoneD}{the dynamic components of the simulated zone}
}
\value{
a refreshed zoneC with updated ExB0 values and zoneD$depleB=1
}
\description{
resetexB0 sets the unfished exploitable biomass at time
    zero to the average of the exploitable biomass levels before and
    after any fishing (when there are zero catches). Hence growth and
    natural mortality are included.
}
\examples{
print("wait on data")
}
