% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{resetLML}
\alias{resetLML}
\title{resetLML sets the LML for a given zone from a given year}
\usage{
resetLML(inzone, inLML, inyear, glob)
}
\arguments{
\item{inzone}{a zone object}

\item{inLML}{a vector of LML for each block to be imposed from inyear on.}

\item{inyear}{the start year for the new LML}

\item{glob}{the global variables object}
}
\value{
A zone object with the LML changed from inyear on
}
\description{
resetLML sets the LML for a given zone from a given year
    out to the final year. It changes the selectivity, the SelWt values,
    and the LML value per block.
}
\examples{
\dontrun{
txt1 <- 'all example code should be able to run'
}
}
