% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{resetSel}
\alias{resetSel}
\title{resetSel alters the selectivity of all popualtions}
\usage{
resetSel(inzone, inLML, glob)
}
\arguments{
\item{inzone}{the simulated zone}

\item{inLML}{the new LML}

\item{glob}{the global variables object}
}
\value{
a revised zone
}
\description{
resetSel alters the selectivity of all populations before
    proceeding with further dynamics
}
\examples{
\dontrun{
  print("develop an example.")

}
}
