% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{restart}
\alias{restart}
\title{restart transfers final year values of zoneD into the first year}
\usage{
restart(oldzoneD, nyrs, npop, N, zero = TRUE)
}
\arguments{
\item{oldzoneD}{the old zoneD containing the dynamics as run for
Nyrs.}

\item{nyrs}{The number of years of dynamics, the global Nyrs}

\item{npop}{The number of populations, the global numpop}

\item{N}{the number of size classes, teh global Nclass}

\item{zero}{should the arrays be otherwise filled with zeros? The
default = TRUE}
}
\value{
a list containing a revised dynamics list
}
\description{
restart transfers the final year values from the
    dynamics part of the zone (zoneD), into the first year.
    This is used, for example, when searching for an equilibrium
    state if there is larval dispersal > 0.0. Of if one sets the
    initial depletion to anything other than 1.0. Contains the
    option of setting every other cell to zero, which is the
    default.
}
\examples{
print("wait for built in data sets")
}
