% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{rewritecompdata}
\alias{rewritecompdata}
\title{rewritecompdasta recreates the size composition data file}
\usage{
rewritecompdata(indir, zone1)
}
\arguments{
\item{indir}{the directory into which the new file should be written. This
would usually be the same as rundir, the scenario directory}

\item{zone1}{a large object inside the output object from makeequilzone that
contains SAUnames, SAUpop, minc, cw, larvdisp, randomseed, initLML,
condC, projC, globals, ctrl, catches, and projyrs. Obviously there is
some redundency.}
}
\value{
nothing but it does write a file called 'lffilename.csv'
    into rundir, which will overwrite the same if already there.
}
\description{
rewritecompdata recreates the size composition data file
}
\examples{
\dontrun{
  data(zone1)
  tmpdir <- tempdir()
  rewritecompdata(tmpdir,zone1)
  dir(tmpdir)
}
}
