% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{rewritedatafile}
\alias{rewritedatafile}
\title{rewrwitedatafile generates a revised saudatafile}
\usage{
rewritedatafile(indir, glb, zone1, saudat)
}
\arguments{
\item{indir}{the directory into which the new file should be written. This
would usually be the same as rundir, the scenario directory}

\item{glb}{the globals object}

\item{zone1}{a large object inside the output object from makeequilzone that
contains SAUnames, SAUpop, minc, cw, larvdisp, randomseed, initLML,
condC, projC, globals, ctrl, catches, and projyrs. Obviously there is
some redundency.}

\item{saudat}{the main contents of the saudata file}
}
\value{
nothing but it does write a file called 'new_saudatafile.csv' into
    the indir
}
\description{
rewritedatafile is used to generate a new saudata file after
    the option of reading in parameters from sizemod has been used. The
    rewritten file contains the revised parameter values.
}
\examples{
print("wait on internal data sets")
}
