% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{rlnormz}
\alias{rlnormz}
\title{rlnormz when sd=0 it returns rep(mean,n) but also uses n random numbers}
\usage{
rlnormz(n, meanlog = 0, sdlog = 1)
}
\arguments{
\item{n}{number of random numbers needed}

\item{meanlog}{the mean of the distribution, on the log scale, default=0}

\item{sdlog}{the standard deviation of the distribution, on the log scale,
default = 1}
}
\value{
a vector of length n of Log-Normal random numbers, if sd=0 it returns
    the meanlog, but uses n random numbers.
}
\description{
when rlnorm is used with sd=0 it returns the meanlog but it does
    not use up any of the random numbers produced by the pseudo-random number
    generator. Hence, if one just uses rlnorm with sd=0 sometimes, then the
    results would differ if it was repeated with sd > 0 because a slightly
    different pseudo-random number sequence would eventuate. rlnormz is a
    wrapper function that returns the meanlog if sdlog = 0, but also uses up
    the required number of random values.
}
\examples{
set.seed(12345)
rlnorm(5, meanlog=5, sdlog=1)
rlnorm(1,meanlog=5,sdlog=1)
set.seed(12345)
rlnorm(5,meanlog=5,sdlog=0)
rlnorm(1,meanlog=5,sdlog=1)
set.seed(12345)
rlnormz(5,meanlog=5,sdlog=0)
rlnorm(1,meanlog=5,sdlog=1)
}
\seealso{
{
  \link{rnormz}
}
}
