% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{rnormz}
\alias{rnormz}
\title{rnormz when sd=0 it returns rep(mean,n) but also uses n random numbers}
\usage{
rnormz(n, mean = 0, sd = 1)
}
\arguments{
\item{n}{number of random numbers needed}

\item{mean}{the mean of the distribution, could be a vector, default=0}

\item{sd}{the standard deviation of the distribution, could be a vector,
default = 1}
}
\value{
a vector of length n or Normal random numbers of mu = mean and stdev
    = sd, if sd=0 it returns the mean, but uses n random numbers.
}
\description{
when rnorm is used with sd=0 it returns the mean but it does not
    use up any of the random numbers produced by the pseudo-random number
    generator. Hence, if one just uses rnorm with sd=0 sometimes, then the
    results would differ if it was repeated with sd > 0 because a slightly
    different pseudo-random number sequence would eventuate. rnormz is a
    wrapper function that returns the mean if sd = 0, but also uses up the
    required number of random values.
}
\examples{
set.seed(12345)
rnorm(5, mean=5, sd=1)
rnorm(1,mean=5,sd=1)
set.seed(12345)
rnorm(5,mean=5,sd=0)
rnorm(1,mean=5,sd=1)
set.seed(12345)
rnormz(5,mean=5,sd=0)
rnorm(1,mean=5,sd=1)
}
\seealso{
{
  \link{rlnormz}
}
}
