% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docondition.R
\name{sauavrecssq}
\alias{sauavrecssq}
\title{sauavrecssq applies the AvRec and returns the ssq from the cpue}
\usage{
sauavrecssq(
  param,
  rundir,
  controlfile,
  datafile,
  linenum,
  calcpopC,
  extra,
  picksau,
  nsau,
  outplot = FALSE
)
}
\arguments{
\item{param}{is the AvRec for the selected SAU}

\item{rundir}{the rundir for the scenario being considered}

\item{controlfile}{the name of the control file, no path}

\item{datafile}{the name of the data file, no path}

\item{linenum}{the linenumber containing the AvRec vector}

\item{calcpopC}{the HS function that calculates the aspirational catches
for each SAU}

\item{extra}{the vector of SAU AvRecs minus the selected SAU value}

\item{picksau}{which SAU is to be worked on as in 1:nsau}

\item{nsau}{the total number of SAU}

\item{outplot}{shoudla plot be generated for output to the webpage.
default=FALSE}
}
\value{
the SSQ for the selected SAU in a comparison of predicted and
    observed CPUE
}
\description{
sauavrecssq applies the AvRec vector while adjusting just
    one for the picksau, and then compares the predicted CPUE with the
    observed and returns a sum-of-squared differences. This is used by
    optim while using the Brent option to find an optimum for a single
    parameter.
}
\examples{
print("Wait on suitable internal data files")
#  param=param;rundir=rundir;controlfile=controlfile;datafile=datafile;linenum=29,
#  calcpopC=calcexpectpopC;extra=extra;picksau=sau;nsau=nsau
}
