% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{saudat}
\alias{saudat}
\title{saudat is a matrix of constants read by readsaudatafile}
\format{
A data.frame of biological and fishery parameters
\describe{
  \item{DLMax}{maximum growth increment}
  \item{L50}{size at half the maximum growth increment}
  \item{L50inc}{L95 - L50, size where growth increment only 5percent of max}
  \item{Wtb}{exponent of weight-at-size relationship}
  \item{Wta}{intercept of weight-at-size relationship}
}
}
\description{
saudat is a 32 x 2 matrix of values for biological variables
    relating to growth, weight-at-size, maturity-at-size, average unfished
    recruitment, selectivity, and other details, including the variation
    expected for each main variable. This information is used to define the
    populations within each SAU.
}
\examples{
 data(saudat)
 round(saudat,5)
}
