% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{sauplots}
\alias{sauplots}
\title{sauplots generates the dosauplots for the dynamic variables}
\usage{
sauplots(
  zoneDP,
  zoneDDR,
  glb,
  rundir,
  B0,
  ExB0,
  startyr,
  addCI = TRUE,
  histCE = NULL,
  tabcat = "projSAU"
)
}
\arguments{
\item{zoneDP}{the dynamic object produced by the projections}

\item{zoneDDR}{the dynamic conditioning object after preparing for replicates
and the addition of recruitment variation in the final years.}

\item{glb}{the object containing the global variables}

\item{rundir}{the results directory}

\item{B0}{the B0 values by population use getvar(zoneC,"B0")}

\item{ExB0}{the ExB0 values by population use getvar(zoneC,"ExB0")}

\item{startyr}{the index for the first year of the conditioning dynamics to
include in the trajectory, to give startyr:indexoflastyear,eg startyr=40}

\item{addCI}{should confidence intervals be added to envelope plots,
default=TRUE}

\item{histCE}{historical CPUE data used in CPUE plots, default=NULL}

\item{tabcat}{the name of the results website tab for the plots}
}
\value{
a list of lists of CI for each SAU and variable as well as the
    zoneDsau and zonePsau
}
\description{
sauplots generates the dosauplots for the dynamic variables
   including cpue, catch, acatch (aspirational catch), matureB, exploitB,
   recruit, and harvestR
}
\examples{
print("wait on suitable internal data-sets")
}
