% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{sauquantbyscene}
\alias{sauquantbyscene}
\title{sauquantbyscene calcualtes the quantiles for all sau and all scenarios}
\usage{
sauquantbyscene(invar, glb)
}
\arguments{
\item{invar}{the output from the scenebyvar function which should be an array
of nyrs x nsau x replicates for the variable selected, be it catch, cpue,
matureB, etc.}

\item{glb}{the globals object. The comparisons invovled imply that all
scenarios considered should have the same number of years and replicates.}
}
\value{
a list of length nsau containing the quantiles for each scenario in
    each sau, for the input variable. The quantiles include 0.025, 0.05, 0.5,
    0.95, and 0.975.
}
\description{
sauquantbyscene takes the output from the scenebyvar function
    and for each sau extracts the quantiles for the input variable for
    each scenario. The output is a list nsau long each containing the five
    quantiles for the variable input for each scenario ready for use by
    doquantplot, or whatever other analysis is required.
}
\examples{
print("wait on data sets")
# invar=cpue; glb=out$glbc[[1]]
}
\seealso{
{
    \link{scenebyvar}, \link{doquantplot}, \link{do_comp_outputs}
}
}
