% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{saurecdevs}
\alias{saurecdevs}
\title{saurecdevs plots the recruitment deviates for each SAU}
\usage{
saurecdevs(recdevs, glb, rundir, filen = "")
}
\arguments{
\item{recdevs}{the matrix of recdevs (out$condC$recdevs)}

\item{glb}{the globals object}

\item{rundir}{the scenario directory for results storage}

\item{filen}{the filename (no path needed) to be used in the rundir}
}
\value{
nothing, but it does generate a plot and save a png file if filen
    is populated
}
\description{
saurecdevs enables a visual comparison of the 'fitted'
    recruitment deviates for each SAU.
}
\examples{
print("wait on suitable data sets")
# recdevs=condC$recdevs;glb=glb;rundir=rundir; filen=""
}
