% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{sauribbon}
\alias{sauribbon}
\title{sauribbon a ribbon plot the quantiles x scenes for the input variable}
\usage{
sauribbon(
  rundir,
  scenes,
  varqnts,
  glb,
  varname,
  console = TRUE,
  q90 = TRUE,
  intens = 127,
  addleg = "bottomright",
  addmedian = 3,
  defpar = TRUE
)
}
\arguments{
\item{rundir}{the directory in which all results are held for a scenario or
comparison of scenarios}

\item{scenes}{the names of the different scenarios being compared}

\item{varqnts}{the quantiles for each scenario as output by sauquantbyscene}

\item{glb}{one of the global objects from out. They should all have the
same reps and years.}

\item{varname}{just the name of the variable being plotted, to make sure the
figures all have the correct labelling}

\item{console}{should the plot be sent to the console or saved for use in the
web-page output? default=TRUE, set to FALSE to save it}

\item{q90}{should the 90th quantiles be plotted? default=TRUE. If FALSE then
the 95th quantiles are used.}

\item{intens}{if polys=TRUE then intens signifies the intensity of colour on
a scale of 0 - 255. 127 is about 50 percent dense.}

\item{addleg}{add a legend or not. If not use "", if yes use any of topleft,
topright, bottomleft, or bottomright, default = bottomright.}

\item{addmedian}{should each polygon also have its median plotted as a line.}

\item{defpar}{should the formatting of the plot use the internal version
(the default setting) or, if set = FALSE use an external definition.}
}
\value{
nothing but it does generate a plot and saves it if console=FALSE
}
\description{
sauribbon is used to produce a ribbon plot for a given sau across
    a set of scenarios for a given variable (cpue, catch, harvest rate, etc)
    out of the dynamics object. After running do_comparison
}
\examples{
\dontrun{
 # a possible example run where scenarios 1, 13, and 14 are compared.
   result <- do_comparison(rundir=rundir,postfixdir=postfixdir,outdir=outdir,
                           files=files,pickfiles=c(1,2,3),verbose=TRUE)
   out <- do_comp_outputs(result,projonly=TRUE)
   catch <- scenebyvar(dyn=out$dyn,byvar="catch",glb=out$glbc[[1]])
   sauribbon("",scenes=out$scenes,sau=8,varqnts=catqnts,glb=out$glbc[[1]],
   varname="Catch",console=TRUE,q90=TRUE,intens=100,addleg="bottomright")
}
}
\seealso{
{
  \link{sauquantbyscene}, \link{doquantribbon}, \link{do_comp_outputs}
}
}
