% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{sautopop}
\alias{sautopop}
\title{sautopop translates a vector of SAU properties into population properties}
\usage{
sautopop(x, sauindex)
}
\arguments{
\item{x}{the vector of length nsau, of properties}

\item{sauindex}{the index of which populations are in which SAU}
}
\value{
a vector of length numpop with each population having the property
    of the respective SAU
}
\description{
sautopop uses teh sauindex object to distribute a set of SAU
    properties (for example recruitment deviates) into a set of population
    properties.
}
\examples{
sauprop <- c(1,2,3,4,5)  # 5 SAU each with 3 populations
sauind <- c(1,1,1,2,2,2,3,3,3,4,4,4,5,5,5) # the sauindex
sautopop(sauprop,sauind)
}
