% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{save_hsargs}
\alias{save_hsargs}
\title{save_hsargs sends a copy of the hs arguments to rundir}
\usage{
save_hsargs(rundir, hsargs)
}
\arguments{
\item{rundir}{the data and results direcotry for the scenario}

\item{hsargs}{the harvest strategy arguments object.}
}
\value{
it saves a text file into the rundir sub-directory and modifies
    the resultTable.csv file for the web-page summary of scenario results
}
\description{
save_hsargs saves a copy of the HS arguments (held in hsargs)
    to the rundir directory ready to be printed as text into the HSperf
    tab of the output webpage. They are stored in a file called hsargs.txt.
    The function expects hsargs to be a list, which it prints component
    by component to a txt file.
}
\examples{
print("wait on suitable internal data sets")
}
