% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{saveobject}
\alias{saveobject}
\title{saveobject is used to save RData files of particular objects}
\usage{
saveobject(obname, object, postfix = "", rundir)
}
\arguments{
\item{obname}{the name of the object to be saved, as a character string}

\item{object}{the parent object from which the object$obname is to extracted}

\item{postfix}{any postfix addition you want for the name default=""}

\item{rundir}{the run directory for the scenario}
}
\value{
nothing but it does save a file to the rundir
}
\description{
saveobject is used to save RData files of particular objects.
    Currently, after projecting 56 populations for 100 replicates the final
    'out' object is over 500Mb, and zoneDP, the projection object, makes up
    430Mb of that. Save that if you wish, but if, say, one wanted only to
    save the outzone object (which summarizes outcomes at the zone level),
    that is only 22kb. This function facilitates the saving process.
}
\examples{
print("wait on tempdir use")
# obname="outzone"; postfix="test"; object=out; rundir=rundir
}
