% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{scaleto1}
\alias{scaleto1}
\title{scaleto1 scales an input vector of CPUE to a mean of one x avCE}
\usage{
scaleto1(invect)
}
\arguments{
\item{invect}{a vector of linear scale CPUE}
}
\value{
a vector of CPUE re-scaled to a mean of one
}
\description{
scaleto1 scales a vector of CPUE to a mean of
    one or avCE. The use of a mean of one means that visual comparisons
    between different time-series becomes visually simplified. The
    avCE option could be used to scale the CPUE to the average
    geometric mean - so as to put it on the nominal scale
}
\examples{
 ce <- c(0.4667187,1.2628564,0.8442146,0.9813531, 0.5554076,0.7426321)
 scaleto1(ce)
}
