% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{scaletoOne}
\alias{scaletoOne}
\title{scaletoOne divides a vector by its means to scale it to mean = 1.0}
\usage{
scaletoOne(invect)
}
\arguments{
\item{invect}{te vector in need of re-scaling}
}
\value{
a rescaled vector
}
\description{
scaletoOne divides a vector by its mean which scales it
    to have a mean of 1.0
}
\examples{
\dontrun{
  x <- 1:9
  y <- scaletoOne(x)
  cbind(x,y)
}
}
