% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{scenebyvar}
\alias{scenebyvar}
\title{scenebyvar extracts as sau x variable x scenario from do_comparison output}
\usage{
scenebyvar(dyn, byvar, glb, projonly = TRUE)
}
\arguments{
\item{dyn}{a list of the dynamics at an sau scale for each scenario in the
comparison made within do_comparison}

\item{byvar}{the variable to be extracted from the dyn object. Valid values
are matureB, exploitB, midyexpB, catch, acatch, harvestR, cpue, recruit,
deplsB, and depleB. Anything will stop the function and shed an error
message.}

\item{glb}{the globals object, which shuold be the same across scenarios}

\item{projonly}{should only the projection years be output, default=TRUE}
}
\value{
a list of the nscene projyrs x nsau x reps for the byvar values
}
\description{
scenebyvar facilitates making comparisons between scenarios by
    at the sau level by extracting individual variables from the 'dyn'
    object that contains each scenarios dynamics by sau and is part of the
    output from the do_comparison function.
}
\examples{
print("wait on example data sets")
}
\seealso{
{
   \link{do_comparison}
}
}
