% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{scenebyzone}
\alias{scenebyzone}
\title{scenebyzone extracts as variable x scenario from do_comparison output}
\usage{
scenebyzone(zone, byvar, glb, projonly = TRUE)
}
\arguments{
\item{zone}{a list of the dynamics at a zone scale for each scenario in the
comparison made within do_comparison}

\item{byvar}{the variable to be extracted from the dyn object. Valid values
are matureB, exploitB, midyexpB, catch, acatch, harvestR, cpue, recruit,
deplsB, and depleB. Anything will stop the function and shed an error
message.}

\item{glb}{the globals object, which should be the same across scenarios}

\item{projonly}{should only the projection years be output, default=TRUE}
}
\value{
a list of the nscene projyrs x reps for the byvar values
}
\description{
scenebyzone facilitates making comparisons between scenarios
    at the zone level by extracting individual variables from the 'zone'
    object that contains each scenarios dynamics at the zone scale, where
    each variable has been amalgamated across populations and sau in an
    appropriate manner. cpue, for example, is catch weighted, whereas matureB
    is simply summed. The zone object is part of the
    output from the do_comparison function.
}
\examples{
print("wait on example data sets")
}
\seealso{
{
   \link{do_comparison}
}
}
