% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{setupdirs}
\alias{setupdirs}
\title{setupdirs sets up and checks the directories used in a run}
\usage{
setupdirs(rundir, verbose = TRUE)
}
\arguments{
\item{rundir}{the principle directory in which all all files
relating to a particular run are to be held}

\item{verbose}{default=TRUE, prints directory status to the console
if FALSE no status reports will be made to the console}
}
\value{
nothing, it allocates directly to the global environment
}
\description{
setupdirs sets up and checks the directories used in
    a run, if they do not exist to start with they will be created.
    Note the use of the <<- funciton, which assigns the directory
    names in teh global environment rather than only within the
    function environment. This saves returning them and extracting
    them.
}
\examples{
rund <- tempdir()
setupdirs(rund)
}
