% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{setuphtml}
\alias{setuphtml}
\title{setuphtml initiates csv files lsiting results to be included}
\usage{
setuphtml(pldir, runname)
}
\arguments{
\item{runname}{the name of the particular run being summarized.}

\item{plotdir}{full path to the directory to contain the plots}
}
\value{
nothing but it does generate a csv file stored in the plots
    directory, named directly into the global environment.
}
\description{
setuphtml initiates the csv file used to contain the
    filenames, captions, and categories of the plots and tables to
    be included in the html results. The format of the csv file
    is to have column names of file, caption, category, and
    timestamp. Then, each plot and table is included with an entry
    for each column.
}
\examples{
indir <- tempdir()
plotdir <- filenametopath(indir,"plots")
dirExists(plotdir,verbose=FALSE)
setuphtml(plotdir,"example_only")
dir(plotdir)
}
