% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{setupregion}
\alias{setupregion}
\title{setupregion makes region's constant, dynamic, and productivity parts}
\usage{
setupregion(constants, region1)
}
\arguments{
\item{constants}{the population constants derived from readdatafile}

\item{region1}{the regional constants}
}
\value{
a list of regionC, regionD, product, and glb the main
    components of the region
}
\description{
setupregion makes the region's constant, dynamic, and
    productivity parts returning them, along with glb, in a list.
}
\examples{
data(constants)
data(region1)
out <- setupregion(constants,region1)
regionC <- out$regionC
glb <- out$glb
str(regionC[[1]])
str(glb)
}
