% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{setupzone}
\alias{setupzone}
\title{setupzone makes zone's constant, dynamic, and productivity parts}
\usage{
setupzone(constants, zone1)
}
\arguments{
\item{constants}{the population constants derived from readdatafile}

\item{zone1}{the zoneal constants}
}
\value{
a list of zoneC, zoneD, product, and glb the main
    components of the zone
}
\description{
setupzone makes the zone's constant, dynamic, and
    productivity parts returning them, along with glb, in a list.
}
\examples{
data(constants)
data(zone1)
out <- setupzone(constants,zone1)
zoneC <- out$zoneC
glb <- out$glb
str(zoneC[[1]])
str(glb)
}
