% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{simplephaseplot}
\alias{simplephaseplot}
\title{simplephaseplot plots phase plots for multiple scenarios}
\usage{
simplephaseplot(
  rundir,
  xvar,
  yvar,
  xlabel,
  ylabel,
  console = TRUE,
  width = 7,
  height = 7,
  font = 7,
  hline = NULL,
  vline = NULL,
  legpos = "bottomright",
  scenes = c("A", "B")
)
}
\arguments{
\item{rundir}{the scenario directory for results storage}

\item{xvar}{a vector of values defining the dynamics of the x variable}

\item{yvar}{a vector of values defining the dynamics of the y variable}

\item{xlabel}{the labelfor the x-axis, also used in the file name if saved}

\item{ylabel}{the labelfor the y-axis, also used in the file name if saved}

\item{console}{should the plot go to the console or be saved as a png file}

\item{width}{the width of the plot, default = 7}

\item{height}{the height of the plot, default = 7}

\item{font}{which font to use, default = 7 = bold Serif}

\item{hline}{adds horizontal reference line to the plot. default = NULL,
which means no line is added, perhaps use hline = 1.0}

\item{vline}{adds vertical reference line to the plot. default = NULL,
which means no line is added, perhaps use vline = 1.0}

\item{legpos}{where to put the legend on the plot, default = 'bottomright'}

\item{scenes}{the names of scenarios to be compared.}
}
\value{
nothing but it does generate a plot
}
\description{
simplephaseplot commonly one wants to plot a dynamic variable
    agsint another in a phase plot to illistrate how they change relative to
    one another. This is done for multiple scenarios for comparison. This
    function attempts to provide a simplified wrapper to do that.
}
\examples{
x = c()
require(codeutils)
require(hplot)
x <- c(0.8,1.1,1.5,1.7, 1.8,1.7,1.6, 1.3,1.2, 1.1,1.1,1.35,1.425,1.35)
y <- c(0.8,0.5,0.5,0.65,0.9,1.2,1.25,1.3,1.27,1.1,1.0,0.9, 0.95,1.05)
simplephaseplot(xvar=x,yvar=y,xlabel="X",ylabel="Y",console=TRUE,scenes="A")
x1 <- matrix(c(0.8,1.1,1.5,1.7, 1.8,1.7,1.6, 1.3,1.2, 1.1,1.1,1.35,1.425,
               1.35,0.8,1.15,1.55,1.75,1.85,1.75,1.65,1.35,1.25,1.15,1.15,
               1.40,1.475,1.4),nrow=14,ncol=2)
y1 <- matrix(c(0.8,0.5,0.5,0.65,0.9,1.2,1.25,1.3,1.27,1.1,1.0,0.9,0.95,1.05,
               0.8,0.5,0.5,0.65,0.9,1.2,1.25,1.3,1.27,1.1,1.0,0.9,0.95,1.05),
               nrow=14,ncol=2)
simplephaseplot(xvar=x1,yvar=y1,xlabel="X1",ylabel="Y1",console=TRUE,
                scenes=c("A","B"))
}
