% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{state_by_year}
\alias{state_by_year}
\title{state_by_year applies a selected function to replicates in dynamics}
\usage{
state_by_year(indyn, whichvar, whichyrs, whichfun, glb)
}
\arguments{
\item{indyn}{usually the dynamics object from do_comparison}

\item{whichvar}{which dynamics variable to use as a character string}

\item{whichyrs}{which year numbers to use}

\item{whichfun}{any R function applicable to vectors or numbers or a custom
user-defined function}

\item{glb}{the globals object needed for the number and names of the sau,
and the years and their names.}
}
\value{
array of function values for each of the years, scenarios, and sau
}
\description{
state_by_year applies a user selected or generated function to
    each replicate in each SAU within one of the variables inside the
    dynamics array. This can be used to determine the mean, median, stdev,
    or any other custom function one cares to define. In the dyn object from
    do_comparison the valid whichvar values are: matureB, exploitB,
    midyexpB, catch, acatch, harvestR, cpue, recruit, deplsB, and depleB.
    The year numbers refer to the total number of years in the simulation,
    both the historic catches and the projection years.
}
\examples{
\dontrun{
# assumes that do_comparison function has been run.
  res <- state_by_year(indyn=dyn,whichvar="catch",whichyrs=c(63,68),
                       whichfun=mean,glb=glb)
# Using the first 4 examples from aMSEGuide should give something like
  round(res[1,,],3)
           sau6   sau7   sau8   sau9  sau10  sau11   sau12  sau13
  EG      7.208 17.537  8.811 41.957 37.559  89.514 48.686 33.714
  EGMR12  7.015 18.477  9.357 43.826 40.893  90.421 52.538 34.061
  EGMR3   8.754 18.774 11.639 47.301 35.875  94.545 51.320 36.831
  EGMRall 7.962 19.887 12.278 53.922 41.175 101.453 57.521 35.796
}
}
