% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{STM}
\alias{STM}
\title{STM Generates the Size Transition Matrix for Inverse Logistic}
\usage{
STM(p, mids)
}
\arguments{
\item{p}{a vector of four parameters in the following order
MaxDL the maximum growth increment of the inverse logistic,
L50 the initial length at which the inflection of the growth
increment curve occurs, L95 - the initial length that defines the
95th percentile of the growth increments, SigMax - the maximum
standard deviation of the normal distribution used to describe the
spread of each distribution of growth increments}

\item{mids}{a vector of initial lengths which also define the width of
each size class thus, mids from 2 - 210 would imply 2mm size
classes 1 - 3 = 2, 3 - 5 = 4, etc}
}
\value{
A square matrix with dimension =the length of the mids vector
}
\description{
STM With the input of the four parameters inside a vector,
    and a vector of initial lengths or mid-points of size classes STM
    generates a square transition matrix with the probabilities of
    growing from each initial size into the same or larger sizes.
    Negative growth is disallowed. All columns in the matrix sum to one.
}
\examples{
 param <- c(25.0,120.0,170.0,4.0)
 midpts <- seq(2,210,2)
 G <- STM(param,midpts)
 print(round(G[1:30,1:8],4))
}
\references{
Haddon, M., Mundy, C., and D. Tarbath (2008) Using an
    inverse-logistic model to describe growth increments of blacklip
    abalone (Haliotis rubra) in Tasmania. Fisheries Bulletin 106: 58-71
}
