% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{summarizeprod}
\alias{summarizeprod}
\title{summarizeprod generates a summary of the productivity properties}
\usage{
summarizeprod(product, saunames)
}
\arguments{
\item{product}{The productivity array from doproduction containing the
range of imposed harvest rates, and the resulting outputs for each
population}

\item{saunames}{the names of the different SAU}
}
\value{
a matrix containing the approximate productivity matrix for the zone
}
\description{
summarizeprod generates a summary of the productivity properties
    by examining the productivity matrix for each SAU/population and
    extracting the Bmsy, annualH, MSY, Depletion, and RelCE at the maximum
    catch level (which approximates the MSY). It summarizes the total zone
    by summing all the productivity matrices and search for the largest
    catch again. It generates estimates of the annualH, depletion and RelCE
    by using a weighted average of those values from teh separate SAU or
    populations, where the weighting is the proportion of the sum of the
    MSYs taken in each sau or population. This latter is only an
    approximation but provides at least an indication.
}
\examples{
\dontrun{
data(zone)
product <- zone$product
zoneprod <- summarizeprod(product,saunames=zone$zone1$SAUnames)
round(zoneprod,3)
}
}
