% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{sumpop2sau}
\alias{sumpop2sau}
\title{sumpop2sau gathers population data into sau data using sauindex}
\usage{
sumpop2sau(invect, sauindex)
}
\arguments{
\item{invect}{a vector of population values for a given variable}

\item{sauindex}{the indices of each sau for each population}
}
\value{
a vector of length nsau containing the sum of population values
    for each sau
}
\description{
sumpop2sau gathers population data into sau data using sauindex
}
\examples{
vect <- c(5.8,6.2,13.2,23.8,3.3,3.7,29.7,26.3,38.9,9.1)
sauind <- c(1,1,2,2,3,3,4,4,5,5)
sumpop2sau(vect,sauind)   # should be 12 37 7 56 48
}
