% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{sumpops}
\alias{sumpops}
\title{sumpops takes the zoneDP and sums the populations within each SAU}
\usage{
sumpops(invar, sauindex, saunames)
}
\arguments{
\item{invar}{either zoneDP$matureB or exploitB, or catch, or recruit}

\item{sauindex}{the SAU index of each population}

\item{saunames}{the names of each SAU}
}
\value{
an array of projyrs x nSAU x reps
}
\description{
sumpops summarizes the matureB, exploitB, catch, recruit
    variables within the depleted zone that has undergone the replicate
    application of an HS, and sums the values from each population within
    each SAU.
}
\examples{
print("wait on new data")
}
