% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svvC}
\alias{svvC}
\title{svvC sum of two multiplied vectors using RcppArmadillo}
\usage{
svvC(lhs, rhs)
}
\arguments{
\item{lhs}{the first vector, possibly numbers at size}

\item{rhs}{the second vector, possibly maturity x weight-at-length}
}
\description{
svvC multiple the vector rhs by the vector lhs and
    sums the resulting vector.
}
