% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{tabulatefinalHSscores}
\alias{tabulatefinalHSscores}
\title{tabulatefinalHSscores saves and write the medina final scores to rundir}
\usage{
tabulatefinalHSscores(rundir, meds, scenes, category = "Tables")
}
\arguments{
\item{rundir}{the location for all results of the comparisons}

\item{meds}{a list of each scenarios median final HS scores from
comparefinalscores}

\item{scenes}{the scenario names}

\item{category}{in what tab should the tables be stored. default="Tables"}
}
\value{
nothing but it does save two tables and generates the HTML for the
    output website
}
\description{
tabulatefinalHSscores saves csv files for each scenario and
    includes a table of each of the HS final scores in the website output.
    The 'meds' object is obtained from the 'comparefinalscores' function.
}
\examples{
print("wait on data")
}
