% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS-HCR.R
\name{targblockHCR}
\alias{targblockHCR}
\title{targblockHCR - calculates the block scores for the targetCE}
\usage{
targblockHCR(incpueBlock, targetCE, modifyTarg = deltaCE)
}
\arguments{
\item{incpueBlock}{the matrix of cpue by Block by year}

\item{targetCE}{- a vector of the targetCE for each of the blocks}

\item{modifyTarg}{- the constant that sets the range of CPUE in the
scoring function}
}
\value{
a vector of scores relating to the targetCE PM; used in
   blockMCDA
}
\description{
targblockHCR - calculates the block scores for the
    targetCE
}
\examples{
print("Need to use a dataset for an example")
}
