% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{tasphaseplot}
\alias{tasphaseplot}
\title{tasphaseplot generates a Tasmanian proxy phase (kobe) plot}
\usage{
tasphaseplot(
  proxyB,
  proxyH,
  glb,
  sau,
  rundir = "",
  console = TRUE,
  fnt = 7,
  setpar = TRUE
)
}
\arguments{
\item{proxyB}{the target CPUE score from outhcr. This is a 3D matrix so the
median across replicates is used.}

\item{proxyH}{the gradient 4 score from outhcr. This is a 3D matrix so the
median across replicates is used. The plot is of 10 - median(proxyH)}

\item{glb}{the globals object - just for the saunames}

\item{sau}{which sau is being plotted. For the western zone this is 1 - 8}

\item{rundir}{the rundir for the scenario where the plot will be stored if
console = FALSE}

\item{console}{plot to the console or save the plot. default=TRUE}

\item{fnt}{what font to use, default = 7 bold times roman. This option is here
because, for some reason, many people prefer to use fnt=2, which is some
pale sans-serif difficult to read horror. But we must recognize that
tastes differ, though really we should encourage people to be more
considerate.}

\item{setpar}{if this is a stand-alone plot setting this TRUE will define a
plot size and set the par values. If the plot is to be part of a
multiple plot then setpar should be FALSE. default=TRUE}
}
\value{
a list of the median mature biomass depletion proxy and median
    harvest rate proxy by year, invisibly, it also plots these vectors.
}
\description{
tasphaseplot generates a Tasmanian proxy phase (kobe) plot. This
    means it uses the gradient 4 scores as a proxy for the harvest rate and
    the target CPUE score as a proxy for the biomass depletion level. The
    scores rank from 0 - 10. The target CPUE score behaves appropriately in
    that high scores = things are good low scores things are bad, but the
    gradient 4 score is the reverse. So, the scores are reversed 10 - score
    and the axis labels adjusted accordingly. Thus, original high g4 scores
    indicate lower harvest rates while low g4 scores indicate high harvest
    rates. 10 - g4s reverses this trend and to indicate the change we use
    a scale from 5 - minus 5 with a limit refernece point at 0.
}
\examples{
\dontrun{
# eample syntax
tasphaseplot(proxyB=outhcr$g4s,proxyH=outhcr$targsc,glb=glb,sau=6,
             rundir="",console=TRUE,fnt=7)
}
}
