% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{testequil}
\alias{testequil}
\title{testequil runs a region Nyrs and determines stability}
\usage{
testequil(regC, regD, glob, inH, verbose = TRUE)
}
\arguments{
\item{regC}{the constants components of the simulated region}

\item{regD}{the dynamic components of the simulated region}

\item{glob}{the general global variables}

\item{inH}{a vector of numpop harvest rates}

\item{verbose}{should results go to the console, default=TRUE}
}
\value{
the dynamics component with Nyrs of dynamics
}
\description{
testequil runs a given region for Nyrs at the given
    harvest rate, and then tests that the last values of matureB,
    exploitB, recruitment, and spawning biomass depletion are the
    same as the first (to three decimal places). It reports this
    to the console if verbose=TRUE
}
\examples{
print("wait on built in data sets")  #regC=regionC; regD=regionD; glob=glb; inH=
}
