% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{testregC}
\alias{testregC}
\title{testregC is a region list made up of 6 equilibrium populations}
\description{
testregC is a region list made up of 6 equilibrium
    populations. These have been run with a laral dispersal rate of
    0.03 so the change from B0 to effB0 is not great, but still
    required for an initial equilibrium. This is here to simplify
    the internal testing of funcitons that require a completed
    region starting at equilibrium. Its name is to avoid conflict
    with any actual use of regionC. use str(testregC, max.level=1)
    to see its format. It can be expected to be used with testregD
}
\section{Subjects}{

 \itemize{
   \item testing of functions that require a full region
   \item initial equilibrium
 }
}

\examples{
 data(testregC)
 data(testregD)
 data(region1)
 glb <- region1$globals
 ans <- testequil(testregC, testregD, glb)
}
