% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{testregD}
\alias{testregD}
\title{testregC is a list of 8 matrices and 2 arrays defining the dynamics of a region}
\description{
testregC is a list of 8 matrices and 2 arrays defining
    the dynamics of a region. These have been run with a larval
    dispersal rate of 0.03 to achieve an initial equilibrium. This
    is here to simplify the internal testing of functions that
    require a completed region starting at equilibrium. Its name is
    to avoid conflict with any actual use of regionD. use
    str(testregD, max.level=1) to see its format. It can be
    expected to be used with testregC.
}
\section{Subjects}{

 \itemize{
   \item testing of functions that require a full region
   \item initial equilibrium
 }
}

\examples{
 data(testregC)
 data(testregD)
 data(region1)
 glb <- region1$globals
 testequil(testregC, testregD, glb)
}
