% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{testzoneC}
\alias{testzoneC}
\title{testzoneC is a zone list made up of 6 equilibrium populations}
\description{
testzoneC is a zone list made up of 6 equilibrium
    populations. These have been run with a laral dispersal rate of
    0.03 so the change from B0 to effB0 is not great, but still
    required for an initial equilibrium. This is here to simplify
    the internal testing of funcitons that require a completed
    zone starting at equilibrium. Its name is to avoid conflict
    with any actual use of zoneC. use str(testzoneC, max.level=1)
    to see its format. It can be expected to be used with testzoneD
}
\section{Subjects}{

 \itemize{
   \item testing of functions that require a full zone
   \item initial equilibrium
 }
 @export
}

\examples{
 data(testzoneC)
 data(testzoneD)
 data(zone1)
 glb <- zone1$globals
 r0 <- getvar(testzoneC,"R0")
 move <- makemove(glb$numpop,r0,glb$larvdisp)
 glb$move <- move
 ans <- testequil(testzoneC, testzoneD, glb)
 str(testzoneC[[1]])
}
