% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{testzoneD}
\alias{testzoneD}
\title{testzoneD is a list of 8 matrices and 2 arrays defining the dynamics of a zone}
\description{
testzoneD is a list of 8 matrices and 2 arrays defining
    the dynamics of a zone. These have been run with a larval
    dispersal rate of 0.03 to achieve an initial equilibrium. This
    is here to simplify the internal testing of functions that
    require a completed zone starting at equilibrium. Its name is
    to avoid conflict with any actual use of zoneD. use
    str(testzoneD, max.level=1) to see its format. It can be
    expected to be used with testzoneC.
}
\section{Subjects}{

 \itemize{
   \item testing of functions that require a full zone
   \item initial equilibrium
 }
 @export
}

\examples{
 data(testzoneC)
 data(testzoneD)
 data(zone1)
 glb <- zone1$globals
 r0 <- getvar(testzoneC,"R0")
 move <- makemove(glb$numpop,r0,glb$larvdisp)
 glb$move <- move
 ans <- testequil(testzoneC, testzoneD, glb)
 str(testzoneD)
}
