% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{twophaseplots}
\alias{twophaseplots}
\title{twophaseplots generates a formal kobe plot plus a Tasmanian proxy plot}
\usage{
twophaseplots(
  dyn,
  glb,
  outhcr,
  sau,
  kobeRP,
  rundir = "",
  startyr = 1992,
  console = TRUE,
  fnt = 7
)
}
\arguments{
\item{dyn}{a list of the dynamics for every sau. This function only requires
the mature biomass depletion and the harvest rates}

\item{glb}{the globals object}

\item{outhcr}{the scores object from the hcr containing the targsc and g4s}

\item{sau}{which sau is to be plotted}

\item{kobeRP}{reference points for the eHS pseudo kobe plot. A vector of the
targdepl, default=0.4, limdepl, default=0.2, and the limH, default=0.15}

\item{rundir}{the directory for all files and results, default=""}

\item{startyr}{default = 1992. Otherwise could use condC$yearCE[1]}

\item{console}{should the plot go to the console or be saved? Default=TRUE}

\item{fnt}{what font to use, default = 7 bold times roman. This option is here
because, for some reason, many people prefer to use fnt=2, which is some
pale sans-serif difficult to read horror. But we must recognize that
tastes differ, though really we should encourage people to be more
considerate.}
}
\value{
a vector of median depletion and H, and median Grad4 and TargCE, it
    also plots a graph containing two kobe plots if in Tas.
}
\description{
twophaseplot generates a combination of the formal kobe plot of
    the predicted biomass depletion level against the predicted harvest rate,
    along with a Tasmanian proxy plot of the target CPUE score (proxy for
    biomass depletion) against the gradient 4 CPUE score (proxy for harvest
    rate).
}
\examples{
print("wait on data")
}
