% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{updatedeltarec}
\alias{updatedeltarec}
\title{updatedeltarec if recruitment timevarys extend across the varyrs}
\usage{
updatedeltarec(glb, varyrs)
}
\arguments{
\item{glb}{the globals object generated by readctrlfile and makezone}

\item{varyrs}{how many years at the end of conditioning on the fishery
data into zoneDD, to which to add recruitment variation, default = 7,
which suits the Tasmanian west coast. Used in prepareprojection and in
updatedeltarec is timevarying recruitment option turned on.}
}
\value{
a revised copy of glb the globals object
}
\description{
updatedeltarec if a linear change in recruitment is modelled
    using the TIMEVARY option in the control file, then update the sequence
    to include the varyrs prior to projections in that linear trend. This
    is to spread the annual decreases (or increases) across the earlier
    years so that any effects will become apparent sooner in the projections.
    Without this the assumption would be that mean recruitment has been
    stable and only started to change once projections occurred.
}
\examples{
print("Wait on date sets")
}
