% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{whenfirst}
\alias{whenfirst}
\title{whenfirst returns the first occurrence of a value in a vector}
\usage{
whenfirst(x, value)
}
\arguments{
\item{x}{a vector of yearly values within a single replicate}

\item{value}{the value to be searched for within x as in x >= value}
}
\value{
a vector replicates long of the first year in which the value occurs
}
\description{
whenfirst is a utility function to be used primarily within an
    apply function call. It could then search across all replicates for the
    first occurrence of a value >= a given target value.
}
\examples{
x <- seq(1,100,1)
whenfirst(x,75)
}
