% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{WtatLen}
\alias{WtatLen}
\title{WtatLen Power function to describe weight at length relationship}
\usage{
WtatLen(ina, inb, lens)
}
\arguments{
\item{ina}{is the intercept of the power function}

\item{inb}{is the gradient (or the explonent) of the power function}

\item{lens}{a vector of lengths for which the weight will be calculated}
}
\value{
A vector of length(lens) containing the predicted values
}
\description{
WtatLen This can be used to generate any power function.
}
\examples{
\dontrun{
a <- 0.0000542856
b <- 3.161415
lens <- seq(2,210,2)
wtL <- WtatLen(a,b,lens)
}
}
