% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{zone}
\alias{zone}
\title{zone a list of 8 major components from makeequilzone function}
\description{
zone is the output from the function makeequilzone and contains
    8 R objects including zoneC, zoneD, glb, constants, saudat, product,
    ctrl, and zone1. This is for the scenario of 2 SAU each with 1 population
    each, so it is not a large object. It is mainly used in examples for
    various functions.
}
\examples{
 data(zone)
 str(zone,max.level=1)
}
