% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{zonecatchvsmsy}
\alias{zonecatchvsmsy}
\title{zonecatchvsmsy calc catch/msy and year-to-msy by zone by scenario}
\usage{
zonecatchvsmsy(prods, zone, glbc)
}
\arguments{
\item{prods}{a list of the production statistics matrix for each scenario,
this is the out$sauprod matrix stored for each scenario.}

\item{zone}{the list of zone outputs for each scenario}

\item{glbc}{the globals object for each scenario (needed in case the hyrs
differ between scenarios)}
}
\value{
an invisible list consisting of a list of the catch/msy by zone for
    each scenario and a matrix of the year to first meeting the MSY in each
    replicate zone catch projection in each scenario
}
\description{
zonecatchvsmsy uses the prods, zone, and glbc lists to
    estimate the ratio of catch/MSY for each replicate of the zone catches,
    plus it determines the first year in each replicate in which MSY is
    achieved, If MSY is not met in a replicate an NA is returned. To compare
    scenarios they must all have the same number of replicate projections
    and must all have the same number of projection years
}
\examples{
print("wait on data sets")
# syntax:  zonecatchvsmsy(prods,zone,glbc)
}
