% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{zonefiletemplate}
\alias{zonefiletemplate}
\title{zonefiletemplate generates a template zone file}
\usage{
zonefiletemplate(indir, filename = "zone1.csv")
}
\arguments{
\item{indir}{directory in which to place the zone file}

\item{filename}{the name for the generated zone file, a character
string that defaults to zone1.csv. It is best to include the
complete path}
}
\value{
nothing, but it creates a template zone file within indir
}
\description{
zonefiletemplate generates a standardized zone file
    template containing constants tha toperate at a zoneal level.
    Generate this and then modify the contents to suit
    the system you are attempting to simulate.
}
\examples{
 yourdir <- tempdir()
 zonefiletemplate(yourdir,filename="zone1.csv")   #
 zone1 <- readzonefile(yourdir,"zone1.csv")
 str(zone1,max.level=1)
}
