% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{zonetosau}
\alias{zonetosau}
\title{zonetosau translates the zonexpop objects to zonexsau objects}
\usage{
zonetosau(inzone, NAS = NULL, glb, B0, ExB0)
}
\arguments{
\item{inzone}{one of the zone dynamics objects made up of populations}

\item{NAS}{the numbers-at-size 4D arrays from doprojection; default=NULL so
it can be ignored during conditioning}

\item{glb}{the object containing the global constants}

\item{B0}{the estimate B0 for each population use getvar(zoneC,"B0")}

\item{ExB0}{the estimate ExB0 for each population use getvar(zoneC,"ExB0")}
}
\value{
a list of dynamics variables by SAU
}
\description{
zonetosau combines the dynamic results for each variable so
    that results by population become results by SAU. matureB, exploitB,
    catch, recruit, catchN, and Nt are simple summations of the totals for
    each population into their respective SAU. The harvest rate would be
    end of year or beginning of year estimates derived from dividing the
    catchxsau by the exploitable biomass x sau. Similarly the deplsB and
    depleB are the end of year matureB and exploitB divided by their
    respective unfished estimated by SAU obtained using getvar(zoneC,"B0").
}
\examples{
print("wait on suitable internal data sets ")
}
