% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{addfilename}
\alias{addfilename}
\title{addfilename adds a filename to the autoresult csv file}
\usage{
addfilename(filename, caption, category, filen = plottabfile)
}
\arguments{
\item{filename}{the full path and filename for the file being added}

\item{caption}{the caption for the figure or table}

\item{category}{what HTML tab should it be displayed on?}

\item{filen}{the file to be added to. defaults to plottabfile, which
is defined into the0 global environment from inside setuphtml.R,
found in aMSE_utils}
}
\value{
nothing but it does add a line to the plottabfile
}
\description{
addfilename is used to facilitate the produciton of
    the HTML results summary for a particular run. This depends
    upon a csv file containing the names of each file to be
    plotted or tabulated. This function adds a filename and the
    supporting caption and category, without one needing to
    remember the syntax.
}
\examples{
indir <- tempdir()
plotdir <- filenametopath(indir,"plots")
dirExists(plotdir,verbose=FALSE)
setuphtml(plotdir,"example_only")
filename <- filenametopath(plotdir,"example.png")
plotprep(filename=filename,verbose=FALSE)
plot(runif(100),runif(100),type="p")
graphics.off()  # could use dev.off()
dir(plotdir)
}
