% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{addrecvar}
\alias{addrecvar}
\title{addrecvar adds recruitment variation to end of conditioning step}
\usage{
addrecvar(
  zoneDD,
  zoneDP,
  zoneC,
  glob,
  condC,
  ctrl,
  varyrs,
  multTAC,
  calcpopC,
  sigR = 1e-08,
  sigB = 1e-08,
  lastsigR = 0.3
)
}
\arguments{
\item{zoneDD}{the dynamic zone after historical fishery data has been used
to finalize the conditioning}

\item{zoneDP}{the empty dynamic zone object ready to receive its first year}

\item{zoneC}{the zone's constants object for each population}

\item{glob}{the object containing the global constants}

\item{condC}{the object containing the historical fishery data}

\item{ctrl}{the control object}

\item{varyrs}{the number of years at the end of the historical period to
which recruitment variation is to be added}

\item{multTAC}{the TAC multiplication matrix from the HCR}

\item{calcpopC}{a function that takes the output from hcrfun and gernerates
the actual catch per population expected in the current year.}

\item{sigR}{the initial recruitment variation default=1e-08}

\item{sigB}{the initial biomass cpuie variation default = 1e-08}

\item{lastsigR}{the recruitment variation to be added to the final varyrs}
}
\value{
an initialized dynamics zone object for the projecitons with the
    first year populated
}
\description{
addrecvar replicates the historical depletion step in zoneDD
    reps times and copies the first Nyrs - varyrs of values into each
    replicate. Then it steps through each replicate adding recruitment
    variation to the dynamics for the last varyrs years. This is so that
    when the projections begin each replicate will already have recruitment
    variability fully developed and there will be no time lag on the
    recruitment and subsequent dynamics.
}
\examples{
print("wait on suitable data-sets")
}
