% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{addrecvar}
\alias{addrecvar}
\title{addrecvar adds recruitment variation to start of generic simulation}
\usage{
addrecvar(zoneC, zoneDR, inHt, glb, ctrl, nyrs = 25, reset = TRUE)
}
\arguments{
\item{zoneC}{the constant zone object. Its selectivity patterns should be
altered to reflect the projection conditions using modzoneCSel}

\item{zoneDR}{the dynamic zone object expanded to include all replicates,
using makezoneDR}

\item{inHt}{the initial harvest strategy that was used to deplete the zone}

\item{glb}{the globals object}

\item{ctrl}{the ctrl object}

\item{nyrs}{defaults to 25, which appears to be enough to produce the
required variation}

\item{reset}{should the first ten rows of the each output be set to the last
10 rows of the initial variation run. default = TRUE. If FALSE the
reps trajectories when exposed to recruitment variation are passed out}
}
\value{
a revised zoneDR object with the first row filled with the last row
    from addrecvar
}
\description{
addrecvar conducts projyrs of simulation at the same constant
    harvest rate used to deplete each population to the desired level, only
    it does this with recruitment variability turned on, which sets up a
    series of replicates with different starting positions. The final step
    is to take the final 10 yearly values from each population and replicate
    and place them in the first ten years of the projection matrices within
    the zoneDR object
}
\examples{
print("wait on suitable data")
}
