% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{asSAU}
\alias{asSAU}
\title{asSAU generates an output list of SAU values from the in put zoneDP}
\usage{
asSAU(zoneDP, sauindex, saunames, b0, exb0)
}
\arguments{
\item{zoneDP}{the population based results list zoneDP}

\item{sauindex}{the SAU index of each population}

\item{saunames}{the names of each SAU}

\item{b0}{a vector of the B0 for each SAU}

\item{exb0}{a vector of the ExB0 for each SAU}
}
\value{
a list of results based around SAU
}
\description{
asSAU takes the population based results in zoneDP and converts
    them into a list based around SAU
}
\examples{
print("wait on new data")
}
