% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS-HCR.R
\name{calibrateMCDA}
\alias{calibrateMCDA}
\title{calibrateMCDA uses historical CPUE to calibrate the MCDA}
\usage{
calibrateMCDA(histCE, saunames)
}
\arguments{
\item{histCE}{this is a matrix of nominal scale CPUE from the fishery of
interest. Usually standardized. Obtainable from the projC object but
also the condC object}

\item{saunames}{the names of each SAU (from zone1$SAUnames)}
}
\value{
a list of three matrices of performance measure scores used to
    condition the MCDA
}
\description{
calibrateMCDA uses the projC$histCE to apply the getgrad1,
    getgrad4, and targscore functions so as to provide vewctors of
    permance measure scores to enhance the calibration of the MCDA before
    any projections are done. This is not absolutely required but improves
    the reality of the process. It also means that the productivity of each
    SAU needs to be scaled to the current cpue (see Conditioning the MSE).
}
\examples{
print("wait on example data being available")
}
