% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{checksizecompdata}
\alias{checksizecompdata}
\title{checksizecompdata is a utility to allow a check of their sizecomp data}
\usage{
checksizecompdata(rundir, controlfile, verbose = TRUE)
}
\arguments{
\item{rundir}{the directory in which all files relating to a
particular run are to be held.}

\item{controlfile}{default="control.csv", the filename of the control
file present in rundir containing information regarding the run.}

\item{verbose}{Should progress comments be printed to console, default=TRUE#'}
}
\value{
nothing but it does plot a graph
}
\description{
checksizecompdata is a utility which allows a user to check the
    coverage and level of sampling of size-composition data prior to
    analysis in the MSE (or sizemod) so that a selection can be made to
    potentially excluded inadequate samples. This is not a required function
    for every analysis it is designed simply as a tool for quality control of
    the input data.
}
\examples{
print("wait on data sets")
}
