% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{comparevar}
\alias{comparevar}
\title{comparevar generates the quantiles for each of a set of input scenarios}
\usage{
comparevar(dyn, glbc, scenes, var = "cpue")
}
\arguments{
\item{dyn}{this is a list of the out$sauout$zonePsau produced by each
scenario. This has to be be generated from the saved RData files from
each scenario}

\item{glbc}{a list of the global objects from each scenario being compared}

\item{scenes}{a  list of the out$ctrl$runlabel from each scenario}

\item{var}{what variable from the dynamics to summarize, valid names include
catch, acatch, cpue, harvestR, desplsB, depleB, recruit, matureB, and
exploitB, default = 'cpue'}
}
\value{
a list of the quantiles of the input var for all sau, with each
    scenario being a different component of the quantscen list, and a list of
    three dimensional arrays of the actual values of var in the varc list.
}
\description{
comparevar generates quantiles for each of a set of input
    scenarios. It takes the full timeline of dynamics and outputs just the
    projections and the quantiles of those projections. Used when comparing
    scenarios after multiple runs. Used when post-processing results.
}
\examples{
print("wait on internal datasets")
}
\seealso{
{
  \link{plotscene}
}
}
