% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS-HCR.R
\name{constCatch}
\alias{constCatch}
\title{constCatch implements a constant TAC harvest strategy}
\usage{
constCatch(inTAC, zoneDP, glob, ctrl, projyrs, inityrs = 10)
}
\arguments{
\item{inTAC}{the constant catch for the zone, from projC$HSdetail}

\item{zoneDP}{the results list zoneDP after it has been generated by
makezoneDR and then had variation added using addrepvar}

\item{glob}{the global object}

\item{ctrl}{the control object}

\item{projyrs}{the number of years of projection from projC$projyrs}

\item{inityrs}{the initial number of years used to express starting variation}
}
\value{
a revised zoneDP now containing the dynamics of each replicate run
}
\description{
constCatch implements a constant catch harvest strategy. It
    distributes the constant TAC across the SAU with respect to the relative
    amounts of exploitable biomass present, including some log-normal error
    that uses withsigB. It still needs error to be added to the distribution
    of each SAUs catch among its component populations
}
\examples{
print("wait on new data")
}
