% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{copyto}
\alias{copyto}
\title{copyto copies the control.csv file from a scenario to a new directory}
\usage{
copyto(
  fromdir,
  todir,
  filename = "control.csv",
  makenew = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{fromdir}{the full path of the current rundir}

\item{todir}{the name of the new destination rundir}

\item{filename}{the filename of the file to be copied}

\item{makenew}{if the 'todir' does not exist should it be created using
dir.create? default = TRUE}

\item{verbose}{should details be printed to the console, default=TRUE}
}
\value{
a vector of 1 or -1 denoting which files are transferred
}
\description{
copyto copies the control.csv file from one scenario's
    directory to another. 'copyto' includes the option of copying to a
    completely different path and will create the 'todir' if it
    does not already exist.
}
\examples{
\dontrun{
# When constructing a new scenario, one can copy a control.csv file from
# a different scenario's rundir into a new one prior to editing it to match
# any new requirements.
copyto(fromdir=rundir,todir=destdir,filename="control.csv")
}
}
