% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{ctrl}
\alias{ctrl}
\title{ctrl the control file for a particular MSE run}
\description{
crtl contains the information required to conduct a
    particular MSE run. It identifies directories, filenames, the
    number of replicates,m and other variables
}
\section{contents}{

\itemize{
  \item runlabel  the identifying name for the run
  \item regionfile  filename containing the region data, see region1
  \item datafile  filename containing the population defintions, see constants
  \item hcrfile  filename continaing the details of the HCR used
  \item outdir  the output directory, containing a plots subdir
  \item reps  how many replicates in this instance, usually 1000
  \item initdepl  the initial depletion level for the region
  \item assessinterval  how often should the region be assessed
  \item withsigR  the level of recruitment variability
  \item withsigB  the level of noise on biomass estimates
  \item withsigce  the level of noise on cpue estimates
}
}

\examples{
 data(ctrl)
 print(ctrl)
}
