% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{ctrlfileTemplate}
\alias{ctrlfileTemplate}
\title{ctrlfileTemplate generates a template input control file}
\usage{
ctrlfileTemplate(filename = "tmpctl.csv")
}
\arguments{
\item{filename}{the name for the generated ctrlfile, character string,
defaults to tmpctl.csv}
}
\value{
a standard definition control file
}
\description{
ctrlfileTemplate generates a standardized input control
    file template. Generate this and then modify the contents to suit
    the system you are attempting to simulate. Defaults to 4 replicates,
    starting at 100% depletion,i.e. at B0, with an assessInterval = 1
    year, a very small probability of an unusual recruitment event,
    and using the mcdaHCR harvest strategy.
}
\examples{
\dontrun{
 ctrlfile <- paste0(yourdir,"/newctrl.csv")
 ctrlfileTemplate(filename=ctrlfile)   #
 control <- readcrtlFile(ctrlfile)
 str(control,max.level=1)
}
}
