% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{ctrlfiletemplate}
\alias{ctrlfiletemplate}
\title{ctrlfiletemplate generates a template input control file}
\usage{
ctrlfiletemplate(filename = "control.csv")
}
\arguments{
\item{filename}{the name for the generated ctrlfile, a character
string that defaults to control.csv. It is best to include the
complete path}
}
\value{
a standard definition control file
}
\description{
ctrlfiletemplate generates a standardized control file
    template. Generate this and then modify the contents to suit
    the system you are attempting to simulate. Defaults to 10
    replicates, starting at 100% depletion,i.e. at B0 and with
    effectively no variation in recruitment, mature biomass or cpue
    calculations.
}
\examples{
 yourdir <- tempdir()
 ctrlfile <- filenametopath(yourdir,"testctrl.csv")
 ctrlfiletemplate(filename=ctrlfile)   #
 control <- readctrlfile(yourdir,"testctrl.csv")
 str(control,max.level=1)
}
