% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{ctrlfiletemplate}
\alias{ctrlfiletemplate}
\title{ctrlfiletemplate generates a template input control file}
\usage{
ctrlfiletemplate(indir, filename = "control.csv")
}
\arguments{
\item{indir}{directory in which to place the control.csv file}

\item{filename}{the name for the generated ctrlfile, a character
string that defaults to control.csv.}
}
\value{
invisibly the fill path and name of the control file. More
    importantly, it write a control file template to that directory.
}
\description{
ctrlfiletemplate generates a standardized control file
    template. Generate this and then modify the contents to suit
    the system you are attempting to simulate. Defaults to 10
    replicates, and with
    effectively no variation in recruitment, mature biomass or cpue
    calculations.
}
\examples{
 yourdir <- tempdir()
 ctrlfiletemplate(yourdir,filename="testctrl.csv")   #
 control <- readctrlfile(yourdir,"testctrl.csv")
 str(control,max.level=1)
}
