% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{datafiletemplate}
\alias{datafiletemplate}
\title{datafiletemplate generates a template input datafile}
\usage{
datafiletemplate(numpop, indir, filename = "tmpdat.csv")
}
\arguments{
\item{numpop}{the totla number of populations in the region}

\item{indir}{the directory into which to place the data template}

\item{filename}{the name for the generated datafile, a character
string, defaults to tmpdat.csv}
}
\value{
a standard definition data file, to be read by readdatafile
    whose name and path is returned invisibly
}
\description{
datafiletemplate generates a standard input datafile
    to use as a template, go in and edit it appropriately to suit
    your own needs. It contains the probability distributions that
    are sampled to provide the necessary biological constants for
    each population.
}
\examples{
\dontrun{
 yourdir <- tempdir()
 datafiletemplate(numpop=6,yourdir,"tmpdat.csv")
 data(region1)
 glb <- region1$globals
 constants <- readdatafile(numpop,yourdir,"tmpdat.csv")
 str(constants,max.level=1)
}
}
