% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{datafiletemplate}
\alias{datafiletemplate}
\title{datafiletemplate generates a template input datafile}
\usage{
datafiletemplate(
  numblock = 3,
  Nyrs = NA,
  filename = "tmpdat.csv",
  FisheryData = FALSE
)
}
\arguments{
\item{numblock}{literally the number of blocks in the simulated zone. The
number of populations per block is pre-set but can easily be altered by
editing the generated data file; a scalar, defaults to 3}

\item{Nyrs}{the number of years of projection from year 2 to Nyrs; year 1
is the initial state, either depleted to a predetermined state or in an
unfished state}

\item{filename}{the name for the generated datafile, character string,
defaults to tmpdat.csv}

\item{FisheryData}{a boolean determining whether or not fishery data,
meaning CPUE, historical catches, effort, and historical LML are to be
included in the data file. These would be used in any stock reduction
that becomes part of the conditioning; defaults to FALSE}
}
\value{
a standard definition data file ready to be read by readdatafile
}
\description{
datafiletemplate generates a standard input datafile to use
    as a template. It is possible to define the number of blocks and
    then, once the data file is created, go in an edit it appropriately
    to suit exactly your own needs.
}
\examples{
\dontrun{
 filename <- dataTemplate()
 condDat <- readdatafile(filename)
 str(condDat,max.level=1)
}
}
