% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{definepops}
\alias{definepops}
\title{definepops makes a parameter vector, popdef, for each population}
\usage{
definepops(inSAU, inSAUindex, const, glob)
}
\arguments{
\item{inSAU}{the number of SAUs defined in the zone, a scalar}

\item{inSAUindex}{a vector of the block index for each population}

\item{const}{a matrix with rows containing the PDF parameters for
each of the biological parameters. The columns of the matrix
reflect values for each populations from readdatafile}

\item{glob}{the globals object from readzonefile}
}
\value{
a matrix with a row for each population and whose columns
    are parameters defining the biological parameters for each population.
}
\description{
definepops makes a parameter vector, popdef, for each
    population. Each population will have slightly different
    biological properties sampled from probability density
    distributions defined in const=condDat$constants. The input can
    be either the whole of condDat or just condDat$constants, which
    contains the definitions of the probability density functions
    used to define each population.
}
\examples{
\dontrun{
  data(zone1)
  glb <- zone1$globals
  data(constants)
  ans <- makezoneC(zone1,constants)
  zoneC <- ans$zoneC
  popdefs <- ans$popdefs
  print(popdefs)
 }
}
