% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{definepops}
\alias{definepops}
\title{definepops makes a parameter vector, popdef, for each population}
\usage{
definepops(innblock, inblockI, condDat)
}
\arguments{
\item{innblock}{the number of blocks defined in the zone; scalar}

\item{inblockI}{a vector containng the block index for each population}

\item{condDat}{a matrix with rows containing the PDF parameters for each
of the biological parameters to be sampled and columns reflecting
values for each block of populations}
}
\value{
a matrix with a row for each population  and whose columns are
    the parameters defiing the biological parameters for each population.
}
\description{
definepops makes a parameter vector, popdef, for each
    population. Each population will have slightly different
    biological properties sampled from probability density
    distributions defined in bConst=condDat$constants. The input can
    be either the whole of condDat or just condDat$constants, which
    contains the definitions of the probability density functions
    used to define each population.
}
\examples{
 data(condDat)
 bconst <- condDat$constants
 glb <- condDat$globals
 nblock <- glb$nblock
 blkdef <- condDat$blkpop
 numpop <- condDat$globals$numpop
 blockI <- defineBlock(nblock,blkpop,numpop)
 popdefs <- definepops(nblock,blockI,bConst)
 print(popdefs)    # innblock=nblock; inblockI=blockI; condDat=condDat
}
