% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{depleteSAU}
\alias{depleteSAU}
\title{depleteSAU resets zoneD, approximately to an input depletion level}
\usage{
depleteSAU(zoneC, zoneD, glob, initdepl, product, len = 15)
}
\arguments{
\item{zoneC}{the constants components of the simulated zone}

\item{zoneD}{the dynamic components of the simulated zone}

\item{glob}{the general global variables}

\item{initdepl}{a vector of target depletion levels for each SAU. This
is found in zone1$condC, but could obviously be modified in each run.}

\item{product}{the production curve matrix from doproduction}

\item{len}{the number of intervals in the trial harvest rates}
}
\value{
a revised zoneD object
}
\description{
depleteSAU resets the depletion level of the whole
    zone and does this by searching for the harvest rate that
    leads to the mature biomass in each population is as close as
    possible to the desired depletion level. This means
    the individual populations will likely vary around the target
    depletion, so the depletion across the zone will only be
    approximately at the target depletion. The
    depletion is measured relative to the effective B0 as that
    takes account of any larval dispersal. This function uses the
    production curve array to search for harvest rates that bound
    the target depletion and then re-searches across those bounds
    using len intervals. Within depleteSAU, the depletion within each sau
    will take place using the LML for year 1 for each SAU. If something
    other than this is required it will need re-coding.
}
\examples{
\dontrun{
  data(zone)
  glb <- zone$glb
  Bzero <- zone$zoneD$matureB[1,]
  depl <- rep(0.3,glb$nSAU) # set initial depletion to ~ 30 percent
  zoneDD <- depleteSAU(zone$zoneC,zone$zoneD,glb,initdepl=depl,zone$product)
  zoneDD$matureB[1,]/Bzero  # calculate depletion
  mean(zoneDD$deplsB[1,])
  zoneDD$deplsB[1,]
}  # zoneC=zone$zoneC;zoneD=zone$zoneD;glob=zone$glb;initdepl=origdepl;product=zone$product;len=12
}
