% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{diagnostics}
\alias{diagnostics}
\title{diagnostics generates an array of diagnostic plots}
\usage{
diagnostics(regC, regD, glob, plot = TRUE)
}
\arguments{
\item{regC}{the constants for the region simulated}

\item{regD}{the dynamic aspects of the region simulated}

\item{glob}{the general globals}

\item{plot}{plot the output or not; default = TRUE}
}
\value{
invisibly returns the unique values of LML
}
\description{
diagnostics can generates an array of diagnostic plots,
    however, currently it only plots the selectivity
}
\examples{
\dontrun{
  print("still to devise an example.")
}
}
