% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{do_comparison}
\alias{do_comparison}
\title{do_comparison is a wrapper function that compares scenarios}
\usage{
do_comparison(
  rundir,
  postfixdir,
  outdir,
  files,
  pickfiles,
  verbose = TRUE,
  intensity = 100,
  zero = FALSE
)
}
\arguments{
\item{rundir}{the complete path to the directory into which all results and
plots from the comparisons will be placed.}

\item{postfixdir}{the name of the final sub-directory into which the results
will be placed. This is used as the name of the website generated to
display the results}

\item{outdir}{the full path to the directory holding all the required .RData
files. If set to '' then files is assumed to include the full path}

\item{files}{the vector of RData file names to be used. If outdir set to ''
then files is assumed to include the full path as well as the file name}

\item{pickfiles}{a vector of indices selecting the files to be used.}

\item{verbose}{should progress updates be made to the console, default=TRUE}

\item{intensity}{is the density of the rgb colours used in the ribbon plots.
the default = 100}

\item{zero}{should the phase plots have an origin at zero, default=FALSE}
}
\value{
nothing but it does conduct a comparison of at least two scenarios
    and places tables and plots into a given sub-directory
}
\description{
do_comparison provides a simplified interface for when making
    comparisons between multiple scenarios. It uses the saved .RData files
    from each scenario and it is up to the user to put those filenames into
    a vector of characters.
}
\examples{
\dontrun{
  suppressPackageStartupMessages({
  library(aMSE)
  library(TasHS)
  library(codeutils)
  library(hplot)
  library(makehtml)
  library(knitr)
  })
  dropdir <- getDBdir()
  prefixdir <- paste0(dropdir,"A_codeUse/aMSEUse/scenarios/")
  postfixdir <- "BC_compare"
  rundir <- filenametopath(prefixdir,postfixdir)
  # normally one would use code to select the files, outdir is used if all
  # RData files ar ein one directory, otherwise set outdir='' and files
  # shgould contain the full paths as well as the filenames
  files=c("BC.RData","BC433.RData","BC541.RData")
  do_comparison(rundir,postfixdir,outdir,files,pickfiles=c(1,2,3))
}
}
\seealso{
{
   \link{scenebyvar}, \link{scenebyzone}. \link{RGB}
}
}
