% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docondition.R
\name{do_condition}
\alias{do_condition}
\title{do_condition is a function to assist with conditioning the model}
\usage{
do_condition(
  rundir,
  controlfile,
  datadir,
  calcpopC,
  cleanslate = FALSE,
  verbose = FALSE,
  doproduct = FALSE
)
}
\arguments{
\item{rundir}{the full path to the directory in which all files relating to a
particular run are to be held. If datadir != rundir then the main data
files are kept in datadir. The juridictionHS.R can be held in either.}

\item{controlfile}{the filename of the control file present in rundir
containing information regarding the particular run.}

\item{datadir}{the directory in which the SAU data file is to be found. This
will usually be the rundir for the scenario run but if the data file is
to be shared among a set of scenarios it will be more efficient to define
a separate datadir}

\item{calcpopC}{a function that takes the output from hcrfun (either the
aspirational catch x SAU or TAC x zone) and generates the actual catch
per population in each SAU expected in the current year.}

\item{cleanslate}{should the directory be emptied of all files first? All
html, png, RData, and css files in the directory will be deleted.
default = FALSE. This is obviously a very powerful and potentially
dangerous argument, hence it needs to be set =TRUE explicitly. It does
not delete any .csv files so if the rundir is used to store the data for
the run then 'cleanslate' will not affect the data or any .R files.}

\item{verbose}{Should progress comments be printed to console, default=FALSE}

\item{doproduct}{should production estimates be made. default=FALSE}
}
\value{
a large list containing tottime, projtime, starttime, glb, ctrl,
    zoneDD, zoneDP, projC, condC, sauout, and outzone
}
\description{
do_condition is a utility function that can be used to speed the
    conditioning of the operating model on available fishery data. During the
    generation of the simulated zone the operation that takes the longest is
    the estimation of the productivity of each population. The productivity
    is estimated at equilibrium and getting there is what tales the time.
    When conditioning the model the objective is to match the predicted
    fishery response to the historical catches to the observed responses.
    'do_condition' generates the simulated zone much more rapidly, which
    simplifies any searches for optimal values of AvRec (average unfished
    recruitment), and especially for suitable values of recruitment deviates.
}
\examples{
print("wait on suitable data sets in data")
}
\seealso{
{
 \link{makeequilzone}, \link{dohistoricC}, \link{prepareprojection},
 \link{doprojections}
}
}
