% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domse.R
\name{do_MSE}
\alias{do_MSE}
\title{do_MSE an encapsulating function to hold the details of a single run}
\usage{
do_MSE(
  rundir,
  controlfile,
  hsargs,
  hcrfun,
  sampleCE,
  sampleFIS,
  sampleNaS,
  getdata,
  calcpopC,
  makeouthcr,
  HSPMs,
  varyrs = 7,
  startyr = 42,
  verbose = FALSE,
  ndiagprojs = 3,
  cutcatchN = 56,
  matureL = c(70, 200),
  wtatL = c(80, 200),
  mincount = 100,
  includeNAS = FALSE,
  depensate = 0
)
}
\arguments{
\item{rundir}{the full path to the directory in which all files relating to a
particular run are to be held.}

\item{controlfile}{the filename of the control file present in rundir
containing information regarding the particular run.}

\item{hsargs}{the constants used to define the workings of the hcr}

\item{hcrfun}{the name of the harvest control rule that is used to
calculate the aspirational catches by SAU or TAC by zone for the
following year}

\item{sampleCE}{a function from jurisdictionHS.R that generates the CPUE
statistics from the projected data}

\item{sampleFIS}{a function from jurisdictionHS.R that generates the FIS
statistics}

\item{sampleNaS}{a function from jurisdictionHS.R that generates the
Numbers-at-size samples}

\item{getdata}{a function that gathers all the data required by the hcrfun
and combines it into an hcrdata object ready for the hcrfun}

\item{calcpopC}{a function that takes the output from hcrfun (either the
aspirational catch x SAU or TAC x zone) and generates the actual catch
per population in each SAU expected in the current year.}

\item{makeouthcr}{Another JurisdictionHS.R function that generates (or not)
an object that is continually updated by the hcrfun. To avoid very
inefficient code this object (at least for Tasmania) is cycled through
the iterations.}

\item{HSPMs}{Another function defined in the HS package that reconstructs
the cpue HSPM scores, and other measures ready for plotting}

\item{varyrs}{how many years at the end of the conditioning on the fishery,
data into zoneDD, to which to add recruitment variation, default = 7,
which suits the Tasmanian west coast. Used in prepareprojection}

\item{startyr}{the index for the first year of the conditioning dynamics to
include in the plotted cpue trajectories, to give startyr:indexoflastyear.
used in sauplots.}

\item{verbose}{Should progress comments be printed to console, default=FALSE}

\item{ndiagprojs}{the number of replicate trajectories to plot in the
diagnostics tab to illustrate ndiagprojs trajectories to ensure that
such projections appear realistic; default=3}

\item{cutcatchN}{to reduce the size of the final array of numbers-at-size
in the catch one can remove all the empty cells below a given size
class. In the default there are 105 2mm size classes and setting
cutcatchN = 56 removes size classes 1:55 2 - 110 mm; only from catchN}

\item{matureL}{is a vector of 2, default = c(70,200), that define the x-axis
bounds on the biology maturity-at-Length plots}

\item{wtatL}{is a vector of 2, default = c(80,200), that define the x-axis
bounds on the biology weight-at-Length plots}

\item{mincount}{given size-composition data what minimum sample size will
be deemed acceptable for inclusion in the plots and conditioning.
default=100.}

\item{includeNAS}{should the NAS projections be included in the output. If
they are then the final output size is greatly increased. default=FALSE}

\item{depensate}{should depensation occur. If > 0, eg depensate=0.2, then
depensation will occur when depletion is < whatever value one puts in for
depensate. Implemented inside the function 'oneyearrec'.}
}
\value{
a large list containing tottime, projtime, starttime, glb, ctrl,
    zoneDD, zoneDP, projC, condC, sauout, and outzone
}
\description{
do_MSE is a form of meta-function that contains the code and
    function calls that constitute a single scenario run for aMSE. This is
    the reason the argument list is as long as it is. The code has been
    encapsulated like this to simplify development, maintenance, and
    distribution. Already with the number of variant conditioning examples
    produced as tests of the code-base, errors have arisen through a failure
    to propagate all changes to the code now in this function to all the
    separate MSE run files used to manage each scenario. A potential flaw
    lies with the need to apply the jurisdictionHS functions to any
    historical fishery data used to condition the operating model, ready for
    the first year of the projections. Currently, the code used is unique to
    the Tasmanian case and this obviously still requires generalization.
    This may require a new function to be produced in each jurisdictionHS.R
    file, but that is still under consideration. Part of the development of
    this meta-function will be to further articulate the generation of
    results, and the inclusion of error traps on the argument list entries.
    To aid in the conditioning the option of not calculating the productivity
    has been added. By default it will occur and when projecting it will also
    always occur. HSstats is a list that is always saved and contains
    performance statistics for the HS used (currently contains sum5 and sum10
    the cumulative catches to 5 and 10 years into the projections).
}
\examples{
print("wait on suitable data sets in data")
}
\seealso{
{
 \link{makeequilzone}, \link{dohistoricC}, \link{prepareprojection},
 \link{doprojections}, \link{getprojyrC}, \link{oneyearrec}
}
}
