% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{dohistoricC}
\alias{dohistoricC}
\title{dohistoricC imposes the historical catches on an unfished zone}
\usage{
dohistoricC(zoneDD, zoneC, glob, condC, sigR = 1e-08, sigF = 1e-08)
}
\arguments{
\item{zoneDD}{The input unfished dynamic zone, zoneD, object}

\item{zoneC}{the zone constants object, zoneC}

\item{glob}{the globals object}

\item{condC}{from the zone1 object, contains historical fisheries data plus}

\item{sigR}{the recruitment variation included}

\item{sigF}{the variability introduced to the catches by population by
fishers not knowing the distribution of exploitable biomass exactly. What
this value should be is unknown, the default=1e-08, is arbitrary but
avoids any effective fisher allocation error between populations.}
}
\value{
a zoneD object
}
\description{
dohistoricC is used during the conditioning of the zone/region
    and imposes the historical catches, held in the zone data object
    obtained using readzonefile, onto an unfished initial zone, and it does
    this by imposing the time-series of catches to each SAU/block. The
    operation is through the use of oneyearC, which imposes one year's
    catches, which are a vector of SAU catches for each year of the series.
}
\examples{
print("wait on some data sets")
}
