% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{doproduction}
\alias{doproduction}
\title{doproduction estimates a production curve for each population}
\usage{
doproduction(zoneC, zoneD, glob, lowlim = 0, uplim = 0.35, inc = 0.005)
}
\arguments{
\item{zoneC}{the constants components of the simulated zone}

\item{zoneD}{the dynamic components of the simulated zone}

\item{glob}{the general global variables}

\item{lowlim}{the lower limit of harvest rate applied, default=0.0}

\item{uplim}{the upper limit of harvest rate applied, default=0.35}

\item{inc}{the harvest rate increment at each step, default=0.005}
}
\value{
an array of the six productivity variables by numpop by
    the number of harvest rates applied
}
\description{
doproduction estimates a production curve for each
    population in the simulated zone. It does this by sequentially
    applying a series of increasing harvest rates and running the
    populations to equilibrium to discover, empirically, the yield
    and other details, such as exploitable biomass, mature biomass,
    the actual harvest rate, the catch or yield, the mature
    depletion, and the relative cpue. Keep in mind that the time
    taken to run this function depends on the number of populations
    but mainly on the length of the harvest rate sequence, which is
    determined by lowlim, uplim, and inc. The lonjger the sequence
    the slower the run. However, the estimates of MSY and related
    statistics are expected to have better accuracy (resolution)
    the finer the harvest rate increments. So when initiating the
    zone it is best to have a long sequence of finely incremented
    harvest rates that take a long time.
}
\examples{
print("wait")
}
