% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{doprojection}
\alias{doprojection}
\title{doprojection applies the Tasmanian MCDA to a simulated zone}
\usage{
doprojection(
  zoneCP,
  zoneDP,
  glob,
  ctrl,
  projyrs,
  inityrs = 10,
  wid = 4,
  targqnt = 0.55,
  pmwts = c(0.65, 0.25, 0.1),
  hcr = c(0.25, 0.75, 0.8, 0.85, 0.9, 1, 1.05, 1.1, 1.15, 1.2),
  recvar = TRUE
)
}
\arguments{
\item{zoneCP}{the modified zoneC object ready for projections}

\item{zoneDP}{the modified zoneD object ready for projections}

\item{glob}{the global variables object}

\item{ctrl}{the ctrl object}

\item{projyrs}{the number of years of projection}

\item{inityrs}{the number of years kept from the period of conditioning,
defult=10}

\item{wid}{the number of years for the gradient PM, default=4}

\item{targqnt}{the target quantile for the cpue target, default=0.55}

\item{pmwts}{the relative weights given to the PMs when generating the score,
default = c(0.65, 0.25, 0.1) targce, grad4, grad1}

\item{hcr}{the TAC multiplier for each score default=c(0.25,0.75,0.8,0.85,
0.9,1,1.05,1.1,1.15,1.2)}

\item{recvar}{should recruitment variation be included}
}
\value{
the filled in zoneDP list
}
\description{
doprojection applies the Tasmanian MCDA to a simulated zone. It
    is still in need of optimization as 1000 iterations would currently take
    about 20 minutes.
}
\examples{
print("wait on data files")
# zoneCP=zoneCP;zoneDP=zoneDR;glob=glb;ctrl=ctrl;projyrs=projC$projyrs;inityrs=projC$inityrs
# wid=4;targqnt=0.55;pmwts=c(0.65, 0.25,0.1);hcr = c(0.25,0.75,0.8,0.85,0.9,1,1.05,1.1,1.15,1.2)
}
