% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{doprojections}
\alias{doprojections}
\title{doprojections conducts the replicate model runs for Tasmania}
\usage{
doprojections(
  ctrl,
  zoneDP,
  zoneCP,
  otherdata,
  glb,
  hcrfun,
  hsargs,
  sampleCE,
  sampleFIS,
  sampleNaS,
  ...
)
}
\arguments{
\item{ctrl}{the ctrl object from readctrlfile}

\item{zoneDP}{the object used to contain the dynamics from the replicate
model runs}

\item{zoneCP}{the object used to contain the constants for each population
used in model dynamics}

\item{otherdata}{an object containing any other data used by the different
functions used n the projections. For example, for Tasmania this would
be pointed at the histCE data set which is appended to each year's new
predicted cpue data from each replicate to calibrate the HCR within the
current HS. Other jurisdictions may use other data.}

\item{glb}{the object containing the global constants for the given run}

\item{hcrfun}{the name of the harvest control rule that is used to
calculate the multiplier for the previous aspirational catches (possibly
for each SAU but possibly the TAC for the wholse zone) so as to
estimate the aspirational catches or TAC or the following year}

\item{hsargs}{the constants used to define the workings of the hcr}

\item{sampleCE}{a function that generates the CPUE statistics}

\item{sampleFIS}{a function that generates the FIS statistics}

\item{sampleNaS}{a function that generates the Numbers-at-size samples}

\item{...}{the ellipsis used in case any of the functions hcrfun, sampleCE,
sampleFIS, or sampleNas require extra arguments not included in the
default collection}
}
\value{
a replacement for zoneDP containing the dynamics of all replicates
    for all projection years
}
\description{
doprojections conducts the replicate model runs for
    Tasmania using the mcdahcr and the input hsargs
}
\examples{
print("wait on suitable internal data sets")
}
