% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{finalscoreoutputs}
\alias{finalscoreoutputs}
\title{finalscoreoutputs generates plots of the PM scores and interactions}
\usage{
finalscoreoutputs(
  rundir,
  cpueHSPMs,
  cpue,
  catches,
  glb,
  yearCE,
  hsargs,
  verbose = TRUE
)
}
\arguments{
\item{rundir}{the directory for a given scenario}

\item{cpueHSPMs}{is a function defined in the HS package that reconstructs
the cpue HSPM scores, and other measures ready for plotting}

\item{cpue}{the projected cpue dynamics from out$sauout$zonePsau$cpue}

\item{catches}{the project catches from out$sauout$zonePsau$catch}

\item{glb}{the globals object for the scenario}

\item{yearCE}{the year names for the observed CPUE used to condition the
model, found in condC$yearCE.}

\item{hsargs}{the list of arguments required by the HS}

\item{verbose}{should updates on progress be sent to console? default=TRUE}
}
\value{
a list of two lists. The first being outHS containing all the
    reconstructed HS outputs, and the second being the medians of the
    various scores and targetcpue.
}
\description{
finalscoreoutputs is used to reconstruct the harvest strategy
    responses to the dynamics so they can be examined and understood. This
    function generates plots, for each sau, of the projected catches, cpue,
    gard1 scores, grad4 scores, target cpue scores, the cpue target and the
    final scores. Uses getcpueHS from TasHS.
}
\examples{
print("wait on data sets, example soon")
# rundir=rundir;cpueHSPMs=getcpueHS;cpue=out$sauout$zonePsau$cpue;
# catches=out$sauout$zonePsau$catches;glb=out$glb;yearCE=out$condC$yearCE
# hsargs=out$hsargs
}
