% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{findmsy}
\alias{findmsy}
\title{findmsy identifies the closest productivity value to MSY}
\usage{
findmsy(product)
}
\arguments{
\item{product}{array of productivity values output from doproduction}
}
\value{
a matrix of numpop rows containing the approximate MSY
    related productivity values and the index of the approx MSY
}
\description{
findmsy for each population in the zone, identifies
    the closest productivity value to MSY and returns the vector
    of productivity values for the selected harvest rate. The Catch
    in each populaiton = MSY and the other variables relate to the
    value at MSY
}
\examples{
\dontrun{  # takes far too long
  data(zone1)
  glb <- zone1$globals
  data(constants)
  ans <- makezoneC(zone1,constants)
  zoneC <- ans$zoneC
  glb <- ans$glb
  ans <- makezone(glb,zoneC)
  zoneC <- ans$zoneC
  zoneD <- ans$zoneD
  ans2 <- modzoneC(zoneC,zoneD,glb)
  zoneC <- ans2$zoneC  # zone constants
  product <- ans2$product
  approxMSY <- findmsy(product)
  print(approxMSY)
}
}
