% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{findunfished}
\alias{findunfished}
\title{findunfished runs the region 3 x Nyrs to equilibrium}
\usage{
findunfished(regC, regD, glob, lowlim = 0, uplim = 0.35, inc = 0.005)
}
\arguments{
\item{regC}{the constants components of the simulated region}

\item{regD}{the dynamic components of the simulated region}

\item{glob}{the general global variables}

\item{lowlim}{the lower limit of harvest rate applied, default=0.0}

\item{uplim}{the upper limit of harvest rate applied, default=0.35}

\item{inc}{the harvest rate increment at each step, default=0.005}
}
\value{
a list containing the updated regionC and regionD
}
\description{
findunfished runs the region 3 x Nyrs so as to find
    the equilibrium. This is necessary if there is any larval
    dispersal set in the region data. It populated the values of
    effB0 and effExB0, the popq, the initial cpue is set to NA,
    and the initial depletions are set to 1.0
}
\examples{
\dontrun{  # takes too long to run
  data(region1)
  glb <- region1$globals
  data(constants)
  ans <- makeregionC(region1,constants)
  regionC <- ans$regionC
  ans <- makeregion(glb,regionC)
  regionC <- ans$regionC
  regionD <- ans$regionD
  ans2 <- findunfished(regionC,regionD,glb)
  str(ans2,max.level=2)
}
}
