% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getaav}
\alias{getaav}
\title{getaav calculates annual absolute variation in catch as a percent}
\usage{
getaav(invect)
}
\arguments{
\item{invect}{a vector of catches}
}
\value{
a single scalar value the AAV of the input catches as a percent
}
\description{
getaav calculates the annual absolute change in catch
    for an input vector of catches, which could be across a series
    of years or even across different spatial units for a single
    year (an unusual use).
    The equation used is aav = 100 x sum(|Ct - Ct-1|)/(sum(Ct).
}
\examples{
  catch <- c(1,2,3,4,5,4,3,2,1)
  getaav(catch)  # should equal 0.32
}
