% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docondition.R
\name{getCPUEssq}
\alias{getCPUEssq}
\title{getCPUEssq calculates ssq between historical and conditioned cpue}
\usage{
getCPUEssq(histCE, saucpue, glb)
}
\arguments{
\item{histCE}{the matrix of historical cpue by SAU}

\item{saucpue}{the predicted cpue from the conditioning on historical data}

\item{glb}{the globals object}
}
\value{
a vector of length nsau containing the ssq for each SAU
}
\description{
getCPUEssq takes in the historical CPUE and the conditioned
    zone's cpue and calculates the sum-of squared differences between them
    to assist with the conditioning. This is a greatly simplified version
    of plotcondCPUE, which plots a graph as well as estimating the ssq.
    For speed it is better to use getCPUEssq.
}
\examples{
print("wait on suitable internal data sets")
# histCE=condC$histCE;saucpue=sauZone$cpue;glb=glb
}
