% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getgrad4}
\alias{getgrad4}
\title{getgrad4 applies a linear regression in steps of wid to input}
\usage{
getgrad4(vectce, wid = 4)
}
\arguments{
\item{vectce}{the inout vector of cpue or proportional changes in
cpue}

\item{wid}{the number of years of the input data to use in each
regression}
}
\value{
a vector of length (wid-1) shorter than the input vector
}
\description{
getgrad4 takes an input vector and applies a linear
    regression along the vector in steps of the parameter wid
}
\examples{
x <- c(0.0169,0.1953,0.1102,0.1511,-0.0403,-0.0247,-0.0255,-0.1089,
       -0.1458,-0.2082,0.0289,-0.0267)
grad4 <- getgrad4(x,wid=4)
grad3 <- getgrad4(x,wid=3)
cbind(c(NA,grad4),grad3)
}
