% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getlistvar}
\alias{getlistvar}
\title{getlistvar extracts a vector or matrix from regionC}
\usage{
getlistvar(regC, indexvar, indexvar2 = "")
}
\arguments{
\item{regC}{the constants components of the simulated region}

\item{indexvar}{the name of the variable to be extracted; character}

\item{indexvar2}{the name of the variable within popdef to extract}
}
\value{
either a vector or matrix of values depending on variable
}
\description{
getlistvar extracts a vector or matrix from regionC.
   If a vector of scalars, the names relate to populations, if a
   matrix the columns relate to populations. Only Me, R0, B0, effB0,
   ExB0, effExB0, MSY, MSYDepl, bLML, popq, SaM, SMUname, popdef,
   LML, Maturity, WtL, Emergent, and MatWt are currently valid
   choices. The indexvar = popdef would generate a listing of all
   the constants. If you only want a single constant from popdefs
   then use indexvar2.
}
\examples{
 data(region1)
 data(constants)
 ans <- makeregionC(region=region1,const=constants)
 regionC <- ans$regionC
 getlistvar(regionC,"MSY")
 getlistvar(regionC,"B0")
 getlistvar(regionC,"popdef","AvRec")
}
