% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getnas}
\alias{getnas}
\title{getnas gets the numbers-at-size for all spatial scales}
\usage{
getnas(regD, yr, glob, region)
}
\arguments{
\item{regD}{the dynamic portion of the region}

\item{yr}{which yr from the range available should be summarized}

\item{glob}{the global variables object}

\item{region}{the region constants}
}
\value{
a matrix of numpop + nSMU + 1 columns of numbers-at-size
}
\description{
getnas extracts numbers-at-size for all populations,
    SMUs, and the region. There will therefore be numpop + nSMU + 1
    columns of numbers-at-size fr the particular year given.
}
\examples{
data(region1)
glb <- region1$globals
data(testregD)
nas <- getnas(testregD,yr=1,glob=glb,region=region1)
round(nas[1:30,],1)
}
