% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getnas}
\alias{getnas}
\title{getnas gets the numbers-at-size for all spatial scales}
\usage{
getnas(zoneD, yr, glob, zone)
}
\arguments{
\item{zoneD}{the dynamic portion of the zone}

\item{yr}{which yr from the range available should be summarized}

\item{glob}{the global variables object}

\item{zone}{the zone constants}
}
\value{
a matrix of numpop + nSAU + 1 columns of numbers-at-size
}
\description{
getnas extracts numbers-at-size for all populations,
    SAUs, and the zone. There will therefore be numpop + nSAU + 1
    columns of numbers-at-size fr the particular year given.
}
\examples{
data(zone1)
glb <- zone1$globals
data(testzoneD)
nas <- getnas(testzoneD,yr=1,glob=glb,zone=zone1)
round(nas[1:30,],1)
}
