% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getsauzone}
\alias{getsauzone}
\title{getsauzone summarizes zoneD into MSU and zone}
\usage{
getsauzone(zoneD, glb, B0, ExB0)
}
\arguments{
\item{zoneD}{the zoneD after nyrs of dynamics}

\item{glb}{the globals object}

\item{B0}{is the sau based B0}

\item{ExB0}{is the sau based ExB0}
}
\value{
a list of six matrices of nSAU columns of SAU summaries,
    and one column for the zone
}
\description{
getsauzone rowsums the matrices of matureB, exploitB,
    catch, and recruit from the input zoneD into the separate SAUs
    and the total into the zone. The harvestR is simply the
    respective catch divided by the exploitB, and the cpue are the
    individual population cpue weighted relative to the catch taken
    from each population in either the SAU or the complete zone.
}
\examples{
print("wait on an example")
}
