% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS-HCR.R
\name{getscore}
\alias{getscore}
\title{getscore calculates the scores for the grad1 and grad4 PMs}
\usage{
getscore(grad14)
}
\arguments{
\item{grad14}{the raw values derived from the function getgrad1 or
getgrad4}
}
\value{
a vector of scores to be included in the MCDA
}
\description{
getscore calculates the scores for the grad1 and grad4
    performance measures. It does this by re-scaling the range of
    the PM values and then fitting separate linear regressions to
    the values above and below zero. These enable it to calculate
    the predicted scores.
}
\examples{
data(blockE13)
nyr <- length(blockE13$year)
grad1 <- getgrad1(blockE13$cpue)
score1 <- getscore(grad1)
cbind(blockE13$year[2:nyr],grad1,score1)
}
