% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{getsingleNum}
\alias{getsingleNum}
\title{getsingleNum extracts a single number from an input line of text}
\usage{
getsingleNum(varname, intxt)
}
\arguments{
\item{varname}{the name of the variable to get from intxt}

\item{intxt}{text line to be parsed, usually obtained using readLines}
}
\value{
a single number
}
\description{
getsingleNum obtains a number from an input line
}
\examples{
\dontrun{
 # Not exported, prefix with AbMSE:::
 txtline <- "replicates, 100"
 AbMSE:::getsingleNum("replicates",txtline)
}
}
