% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getsingleNum}
\alias{getsingleNum}
\title{getsingleNum extracts a single number from an input line of text}
\usage{
getsingleNum(varname, intxt, context = "")
}
\arguments{
\item{varname}{the name of the variable to get from intxt}

\item{intxt}{text to be parsed, usually obtained using readLines}

\item{context}{the function name within which getsingleNum is
being used, defaults to empty string}
}
\value{
a single number
}
\description{
getsingleNum obtains a number from an input line. The
    variable context is the a variable declared in each function within
    which getsingleNum is to be called. This cvan then be used in a
    the stop function.
}
\examples{
\dontrun{
 # Not exported, prefix with aMSE:::
 context = "Function Example"
 txtline <- "replicates, 100"
 aMSE:::getsingleNum("replicates",txtline,context=context)
 aMSE:::getsingleNum("eeplicates",txtline,context=context)
}
}
