% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getsum}
\alias{getsum}
\title{getsum sums each of the main dynamics within zoneD}
\usage{
getsum(inmat, index)
}
\arguments{
\item{inmat}{what matrix within zoneD to sum into SAU and zone}

\item{index}{a vector containing an index of populations within SAU}
}
\value{
an nSAU+1 column matrix summarizing each SAU and the zone
}
\description{
getsum is only used by getsauzone to sum each of the main
    dynamic variables within zoneD, and hence is not exported.
}
\examples{
print("wait on an example")
}
