% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getunFished}
\alias{getunFished}
\title{getunFished - extracts all data relating to year 1; unfished.}
\usage{
getunFished(zoneC, zoneD, glb)
}
\arguments{
\item{zoneC}{the constant part of the zone}

\item{zoneD}{the dynamic part of the zone}

\item{glb}{contains the global variables used everywhere}
}
\value{
a list of multiple components relating to the unfished stock
}
\description{
getunFished - extracts all data relating to year 1;
   unfished. Of course, this will only work as long as movezoneYear
   hasn't been called, which would disrupt the contents of the first
   year. But if applied to zone rather than zone1 etc, this should be
   fine. The outputs include: R0, B0, B0crypt, popdef, MSY,
   MSYDepl, LML, ExB0, and Nt
}
\examples{
\dontrun{
  data(zone1)
  glb <- zone1$globals
  data(constants)
  ans <- makezoneC(zone1,constants)
  zoneC <- ans$zoneC
  ans <- makezone(glb,zoneC)
  zoneC <- ans$zoneC
  zoneD <- ans$zoneD
  ans2 <- modzoneC(zoneC,zoneD,glb)
  zoneC <- ans2$zoneC
  unfish <- getunFished(zoneC,zoneD,glb)
  str(unfish,max.level=1)
}
}
