% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getvar}
\alias{getvar}
\title{getvar a replacement for sapply to obtain scalar constants}
\usage{
getvar(regC, invar)
}
\arguments{
\item{regC}{the constants object for the region}

\item{invar}{a character variable eg. "B0" or "R0"}
}
\value{
a numpop vector of the invar constants from regionC
}
\description{
getvar is a replacement for sapply to obtain scalar
    constants from regionC and is significantly faster. It should
    be used to obtain things like B0, R0, MSY, popq, etc. Still
    need to use sapply to pull out vectors.
}
\examples{
data(testregC)
getvar(testregC,"MSY")
getvar(testregC,"B0")
}
