% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getvect}
\alias{getvect}
\title{getvect extracts invar from the popdef vector in regionC}
\usage{
getvect(regC, invar)
}
\arguments{
\item{regC}{the constants object for the region}

\item{invar}{a character variable eg. "steeph", "DLMax"}
}
\value{
a numpop vector of invar from the numpop popdefs in regionC
}
\description{
getvect extracts a numpop vector of invar from the
    popdef vector in regionC. Still need to use sapply to pull out
    complete vectors such as popdef or maturity etc.
}
\examples{
data(testregC)
getvect(testregC,"steeph")
}
