% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getzoneProps}
\alias{getzoneProps}
\title{getzoneProps extracts the depletion level for a given year}
\usage{
getzoneProps(regC, glb, year = 1)
}
\arguments{
\item{regC}{the constants for the region being simulated}

\item{glb}{the global constants}

\item{year}{which years information is wanted}
}
\value{
a list of size objects
}
\description{
getzoneProps extracts the depletion level along with an
    array of other statistics for a given year.
}
\examples{
\dontrun{
 print("wait on internal data sets")
}
}
