% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfunctions.R
\name{getzoneprops}
\alias{getzoneprops}
\title{getzoneprops extracts the depletion level for a given year}
\usage{
getzoneprops(zoneC, zoneD, glb, year = 1)
}
\arguments{
\item{zoneC}{the constants for the zone being simulated}

\item{zoneD}{the dynamic part of the zone being simulated}

\item{glb}{the global constants}

\item{year}{which years information is wanted, default=1}
}
\value{
a matrix of population properties with a column of totals
}
\description{
getzoneprops extracts a set of properties for each
    population and summarizes them for the zone as well. These
    properties include effB0, matureB, legalmatB, propprot, MSY,
    effexB0, exploitB, SpBDepl, ExBDepl, legalDepl, MSYDepl, LML,
    and bLML. See the model documentation for the meaning of each
    of these.
}
\examples{
data(zone)
str(zone,max.level=1)
# round(getzoneprops(zone$zoneC,zone$zoneD,zone$glb),4)
# zoneC=zoneC; zoneD=zoneD;glb=glb;year=1
}
