% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{imperr}
\alias{imperr}
\title{imperr calculates population catches from sau catches with error}
\usage{
imperr(catchsau, exb, sauindex, sigmab = 1e-08)
}
\arguments{
\item{catchsau}{the predicted or aspirational catch per SAU from the Harvest
control rule}

\item{exb}{the exploitable biomass at the end of the previous year. In the
first year of the projections this would be the last year of the
conditioning.}

\item{sauindex}{the SAU index for each population}

\item{sigmab}{the Log-Normal standard deviation of implementation error. The
default value = 1e-08, which effectively means no errors.}
}
\value{
a vector of population catches for the year to be imposed after the
    estimation of exploitable biomass
}
\description{
imperr converts aspirational sau catches into population level
    catches while introducing management implementation error. Here this
    error is implemented as Log-Normal errors on diver intuitions
    concerning the relative abundance in each population. The error is
    imposed separately on the populations in each SAU. Currently this
    function not used in aMSE.
}
\examples{
print("wait on suitable internal data sets")
}
