% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{makeabpop}
\alias{makeabpop}
\title{makeabpop generates full population structure in an unfished state}
\usage{
makeabpop(popparam, midpts, projLML)
}
\arguments{
\item{popparam}{the vector of biological parameters values that define
the specific properties of this population. Obtained from
popdefs, which is produced by definepops}

\item{midpts}{the center values of the size classes dervied from
the zone data file}

\item{projLML}{the LML expected in the projection years 2 - Nyrs;
a vector obtained from the controlfile}
}
\value{
a list of numpop lists of 19 objects as detailed above.
}
\description{
makeabpop generates the full population structure in an
   unfished state. The structure is pre-determined: Me R0 B0 effB0
   ExB0, effExB0 MSY MSYDepl bLML scalece SaM popdef (vector of constants)
   LML G Maturity WtL Emergent Select SelWt MatWt SAUname. See the
   AbMSE documentation to see the full definition of a zone, made
   up of a zoneC and a zoneD. Notice the presence of effB0 and
   effExB0, these relate to the influence of larval dispersal on each
   populations productivity. The effective B0 relates to the unfished
   mature biomass after larval dispersal occurs and the population
   achieves equilibrium.
}
\examples{
print("See the code for makezoneC to see usage of makeabpop")
}
