% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{makeequilzone}
\alias{makeequilzone}
\title{makeequilzone high level function generates equilibrium zone}
\usage{
makeequilzone(
  rundir,
  ctrlfile = "control.csv",
  doproduct = TRUE,
  selectyr = 0,
  uplimH = 0.4,
  incH = 0.005,
  verbose = TRUE
)
}
\arguments{
\item{rundir}{the directory containing the control csv files. It
can/will also act to store results in a manner that will allow them
to be displayed using makehtml.}

\item{ctrlfile}{the main file that controls the particular run. It contains
the name of the data file that is used to biologically condition the
numpop populations}

\item{doproduct}{boolean, should the productivity calculations be made
during the conditioning. Set to FALSE conditionOM}

\item{selectyr}{which year's LML should be used to estimate the LML. This
should be a year index (eg in Tas 1-58). If set = 0, then the LML for
the glb$hyrs, the last year of conditioning data will be used.
default = 0}

\item{uplimH}{defines the upper limit of harvest used when estimating the
productivity (also important when initial depletion is not 1.0). The
default = 0.4}

\item{incH}{defines the interval between H steps when estimating productivity
default = 0.005}

\item{verbose}{Should progress comments be printed to console, default=TRUE}
}
\value{
a list of zoneC, zoneD, glb, constants, saudat,product, ctrl,
    and zone1
}
\description{
makeequilzone is a high level function that merely hides the
    details of generating the original unfished zone after reading in the
    data files, estimates the productivity, and sets up the results
    directory, rundir, ready to receive files.
}
\examples{
print("wait on datafiles")
 #  rundir=rundir;ctrlfile=controlfile; verbose=TRUE;
 #  selectyr=0;uplimH=0.35;incH=0.005;doproduct=FALSE;
}
