% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{makeequilzone}
\alias{makeequilzone}
\title{makeequilzone high level function generates equilibrium zone}
\usage{
makeequilzone(resdir, ctrlfile = "control.csv", cleanslate = FALSE)
}
\arguments{
\item{resdir}{the directory containing the data and control csv files. It
can/will also act to store results in a manner that will allow them
to be displayed using makehtml.}

\item{ctrlfile}{the main file that controls the particular run. It contains
the name of the data file that is used to biologically condition the
numpop populations}

\item{cleanslate}{a boolean determining whether any old results are deleted
from resdir before starting. Default=FALSE}
}
\value{
a list of zoneC, zoneD, glb, constants, product, ctrl, and zone1
}
\description{
makeequilzone is a high level function that merely hides the
    details of generating the original unfished zone after reading in the
    data files, estimates the productivity, and sets up the results
    directory, resdir, ready to receive files.
}
\examples{
print("wait on datafiles")
}
