% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{makeregion}
\alias{makeregion}
\title{makeregion generates the parts of the simulated region}
\usage{
makeregion(glob, regC)
}
\arguments{
\item{glob}{the global constants defined for the current simulation.
These include numpop, nblock, midptsd, Nclass, and Nyrs}

\item{regC}{the regionC object from makeregionC}
}
\value{
a list of the dynamics and the constant components of the
    simulation
}
\description{
makeregion generates the dynamics components of the
    simulated region and completes the constant components of the
    simulated region. The term 'region' refers to the upper level
    of geographical detail that is used. Thus, for example, in
    Tasmania we might simulate a number of statistical blocks with
    multiple populations. In combination, we refer to the total as
    a region. The matrices are Nyrs x numpop and the two
    arrays are N x Nyrs x numpop.
}
\examples{
  data(region1)
  glb <- region1$globals
  data(constants)
  ans <- makeregionC(region1,constants)
  regionC <- ans$regionC
  glb <- ans$glb
  ans2 <- makeregion(glb,regionC)
  str(ans2,max.level=2)
}
