% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{makeZone}
\alias{makeZone}
\title{makeZone the list of numpop populations defines the abalone zone}
\usage{
makeZone(condDat, uplim = 0.4)
}
\arguments{
\item{condDat}{the matrix of biological constants read from the data file}

\item{uplim}{the upper limit on the harvest rate used when estimating
the productivity of each population, defaults to 0.4; a scalar}
}
\value{
a list containing the simulated zone, the production matrix from
    the function doproduction, and the popdefs used to create the zone.
}
\description{
makeZone the list of numpop populations defines the abalone
    zone. This is an S3 class names 'zone'
}
\examples{
\dontrun{
data(condDat)  # replaces the lines above
out <- makeZone(condDat,uplim=0.38)  # Define the Zone
zone <- out$zone
str(zone[[1]],max.level=1)
production <- out$Production
popdefs <- out$popdefs
print(round(t(popdefs),3))
}
}
