% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{makezoneDP}
\alias{makezoneDP}
\title{makezoneDP generates the container for the projection dynamics}
\usage{
makezoneDP(projyr, iter, glob, inzoneD)
}
\arguments{
\item{projyr}{the number of years of the projection}

\item{iter}{the number of replicates}

\item{glob}{the global object}

\item{inzoneD}{the zoneD object after any preliminary depletion}
}
\value{
a large list containing an object ready for the projection dynamics
}
\description{
makezoneDP generates an object designed to j=hold the outputs
    from each replicate within a set of projections. This is identical to
    zoneD except it contains a repaet for each iteration.
}
\examples{
print("wait for a considerd example")
}
