% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{makezoneDR}
\alias{makezoneDR}
\title{makezoneDR generates the container for the projection dynamics}
\usage{
makezoneDR(projyr, iter, glb, inzoneD, multC)
}
\arguments{
\item{projyr}{the number of years of the projection}

\item{iter}{the number of replicates}

\item{glb}{the global object}

\item{inzoneD}{the zoneD object after any preliminary depletion}

\item{multC}{the TAC multiplier from the last year of conditioning}
}
\value{
a large list containing an object ready for the projection dynamics
}
\description{
makezoneDR generates an object designed to hold the outputs
    from each replicate within a set of projections. This is identical to
    zoneD except it contains a repeat for each iteration and now includes
    a cesau, which is the population-catch-weighted cpue for each SAU.
    Variation should be added by addrecvar.
}
\examples{
print("Could add variation to the harvest rates so that when ")
print("prepareprojection was run the range of initial H values would increase ")
}
