% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{modprojC}
\alias{modprojC}
\title{modprojC produces three objects used to condition the zone}
\usage{
modprojC(zoneC, glob, inzone)
}
\arguments{
\item{zoneC}{the constant part of the zone}

\item{glob}{the globals for the simulation}

\item{inzone}{the zone object from readctrlzone}
}
\value{
a list of projSel and projSelWt within the projC object from inzone
}
\description{
modprojC produces an object used to condition the zone when
    projecting it following any conditioning. Once the initial conditions
    for the projection have been attained through an initial depletion of an
    unfished equilibrium, or more particularly by conditioning on historical
    data then the projC object will be used to condition the projections.
}
\examples{
print("wait on new example data")
}
