% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{numbersatsize}
\alias{numbersatsize}
\title{numbersatsize plots details of the numbers-at-size}
\usage{
numbersatsize(resdir, runlabel, glb, zoneD)
}
\arguments{
\item{resdir}{the results directory}

\item{runlabel}{the runlabel, ideally from the ctrl object}

\item{glb}{the globals list}

\item{zoneD}{the dynamic part of the zone, zoneD}
}
\value{
nothing but it does add one plot to the results directory
}
\description{
numbersatsize plots up the initial unfished numbers-
    at-size distribution, omitting the first four size classes to
    avoid the recruitment numbers dominating the plot.
}
\examples{
print("this will be quite long when I get to it")
}
