% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{oneyear}
\alias{oneyear}
\title{oneyear one year's harvest dynamics for one abpop}
\usage{
oneyear(inpopC, inNt, Nclass, inH, yr)
}
\arguments{
\item{inpopC}{a single population from a zoneC, an abpop}

\item{inNt}{the numbers at size for the year previous to the year
of dynamics. These are projected into the active year.}

\item{Nclass}{the number of size classes used to describe growth.
used to define vectors}

\item{inH}{a literal annual harvest rate as a proportion to be
removed as catch during the year, a scalar}

\item{yr}{the year in the dynamics being worked on. The first year
is generated when the zone is defined or when it is initially
depleted. All dynamics are applied from year 2 - Nyrs; scalar}
}
\value{
a list containing ExploitB, MatureB, Catch, Harvest, Nt,
    ce, and CatchN used to update the given pop in yr + 1
}
\description{
oneyear do one year's dynamics for one input population.
   Used to step through populations of a zone. Its dynamics are
   such that it first undergoes growth, then half natural mortality.
   This allows an estimate of exploitable biomass before fishing
   occurs. The remaining dynamics involve the removal of the catch,
   the application of the last half of natural mortality and the
   addition of recruits. Which allows the exploitable biomass to be
   estimated after fishing. The recruitment occurs in oneyearD so
   that larval dispersal can be accounted for. Thus, oneyear requires
   oneyearD.
}
\examples{
print("need to wait on built in data sets")
}
