% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{oneyearD}
\alias{oneyearD}
\title{oneyearD conducts one year's dynamics in the simulation}
\usage{
oneyearD(regC, regD, Ncl, inHt, year, sigmar, npop, deltarec)
}
\arguments{
\item{regC}{the constant portion of the region with a list of
properties for each population}

\item{regD}{the dynamics portion of the region, with matrices and
arrays for the dynamic variables of the dynamics of the
operating model}

\item{Ncl}{the number of size classes used to describe size}

\item{inHt}{a vector of harvest rates taken in the year from each
population}

\item{year}{the year of the dynamics, would start in year 2 as year
1 is the year of initiation.}

\item{sigmar}{the variation in recruitment dynamics, set to 1e-08
when searching for equilibria.}

\item{npop}{the number of populations, the global numpop}

\item{deltarec}{the rate of larval dispersal}
}
\value{
a list containing a revised dynamics list
}
\description{
onyearD conducts one year's dynamics in the simulation
    returning the revised regionD, which will have had a single year
    of activity included in each of its components.
}
\examples{
print("wait for built in data sets")
}
