% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{oneyearD}
\alias{oneyearD}
\title{oneyearD conducts one year's dynamics in the simulation}
\usage{
oneyearD(regC, regD, Ncl, inHt, year, sigmar, npop, movem)
}
\arguments{
\item{regC}{the constant portion of the region with a list of
properties for each population}

\item{regD}{the dynamics portion of the region, with matrices and
arrays for the dynamic variables of the dynamics of the
operating model}

\item{Ncl}{the number of size classes used to describe size}

\item{inHt}{a vector of harvest rates taken in the year from each
population}

\item{year}{the year of the dynamics, would start in year 2 as year
1 is the year of initiation.}

\item{sigmar}{the variation in recruitment dynamics, set to 1e-08
when searching for equilibria.}

\item{npop}{the number of populations, the global numpop}

\item{movem}{the larval dispersal movement matrix}
}
\value{
a list containing a revised dynamics list
}
\description{
onyearD conducts one year's dynamics in the simulation
    returning the revised regionD, which will have had a single year
    of activity included in each of its components.
}
\examples{
 data(constants)
 data(region1)
 ans <- makeregionC(region1,constants) # classical equilibrium
 regionC <- ans$regionC
 glb <- ans$glb
 ans <- makeregion(glb,regionC) # now make regionD
 regionC <- ans$regionC  # region constants
 regionD <- ans$regionD
 numpop <- glb$numpop
 harvest <- rep(0.2,numpop)
 regionD <- oneyearD(regC=regionC,regD=regionD,Ncl=glb$Nclass,
                     inHt=harvest,year=2,sigmar=1e-06,npop=numpop,
                     movem=glb$move)
 str(regionD)
 round(regionD$catchN[60:105,1:5,1],1)
}
