% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{oneyearD}
\alias{oneyearD}
\title{oneyearD conducts one year's dynamics on zoneD in the MSE}
\usage{
oneyearD(zoneC, zoneD, inHt, year, sigmar, Ncl, npop, movem)
}
\arguments{
\item{zoneC}{the constant portion of the zone with a list of
properties for each population}

\item{zoneD}{the dynamics portion of the zone, with matrices and
arrays for the dynamic variables of the dynamics of the
operating model}

\item{inHt}{a vector of harvest rates taken in the year from each
population}

\item{year}{the year of the dynamics, would start in year 2 as year
1 is the year of initiation.}

\item{sigmar}{the variation in recruitment dynamics, set to 1e-08
when searching for equilibria.}

\item{Ncl}{the number of size classes used to describe size}

\item{npop}{the number of populations, the global numpop}

\item{movem}{the larval dispersal movement matrix}
}
\value{
a list containing a revised dynamics list
}
\description{
oneyearD conducts one year's dynamics on zoneD in the MSE
    returning the revised zoneD, which will have had a single year
    of activity included in each of its components. This uses zoneC
    but always within the environment of another function in which
    zoneC (as zoneC) can be found. Used in runthreeH, (and hence
    dodepletion and doproduction) and in testequil
}
\examples{
 print("wait oin revised data")
}
