% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{oneyearsauC}
\alias{oneyearsauC}
\title{oneyearsauC conducts one year's dynamics using catch not harvest}
\usage{
oneyearsauC(
  zoneCC,
  inN,
  popC,
  year,
  Ncl,
  sauindex,
  movem,
  sigmar = 1e-08,
  r0,
  b0,
  exb0,
  rdev
)
}
\arguments{
\item{zoneCC}{the constant portion of the zone with a list of
    properties for each population
##param exb the vector of exploitable biomass for each population for the
##     previous year}

\item{inN}{the numbers-at-length for each population for the previous year}

\item{popC}{a vector of actual catches to be taken in the year from each SAU}

\item{year}{the year of the dynamics the exb and inN would be from year-1}

\item{Ncl}{the number of size classes used to describe size, global Nclass}

\item{sauindex}{the sau index for each population from glb}

\item{movem}{the larval dispersal movement matrix, global move}

\item{sigmar}{the variation in recruitment dynamics, set to 1e-08
when searching for an equilibrium.}

\item{r0}{the unfished R0 level used by oneyearrec}

\item{b0}{the unfished mature biomass B0, used in recruitment and depletion}

\item{exb0}{the unfished exploitable biomass used in depletion}

\item{rdev}{the recruitment deviates for each SAU for the given year}
}
\value{
a list containing a revised dynamics list
}
\description{
oneyearsauC conducts one year's dynamics in the simulation
    using historical SAU catches rather than harvest rates. The harvest rates
    are estimated after first estimating the exploitable biomass.
    returning the revised zoneD, which will have had a single year
    of activity included in each of its components. uses the function
    imperr to introduce implementation error to the actual catches.
}
\examples{
print("Wait on new data")
}
