% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{oneyearsauC}
\alias{oneyearsauC}
\title{oneyearsauC conducts one year's dynamics using catch not harvest}
\usage{
oneyearsauC(
  zoneCC,
  exb,
  inN,
  catchsau,
  year,
  Ncl,
  sauindex,
  movem,
  sigmar = 1e-08,
  sigmaF = 1e-08
)
}
\arguments{
\item{zoneCC}{the constant portion of the zone with a list of
properties for each population}

\item{exb}{the vector of exploitable biomass for each population for the
previous year}

\item{inN}{the numbers-at-length for each population for the previous year}

\item{catchsau}{a vector of catches to be taken in the year from each SAU}

\item{year}{the year of the dynamics the exb and inN would be from year-1}

\item{Ncl}{the number of size classes used to describe size, global Nclass}

\item{sauindex}{the sau index for each population from glb}

\item{movem}{the larval dispersal movement matrix, global move}

\item{sigmar}{the variation in recruitment dynamics, set to 1e-08
when searching for an equilibrium.}

\item{sigmaF}{the variability introduced to the catches by population by
fishers not knowing the distribution of exploitable biomass exactly.
Initial default value=1e-08}
}
\value{
a list containing a revised dynamics list
}
\description{
oneyearsauC conducts one year's dynamics in the simulation
    using historical SAU catches rather than harvest rates. The harvest rates
    are estimated after first estimating the exploitable biomass.
    returning the revised zoneD, which will have had a single year
    of activity included in each of its components. uses the function
    imperr to introduce implementation error to the actual catches.
}
\examples{
print("Wait on new data")
# zoneCC=zone$zoneC; exb=exb ;catchsau=zone$zone1$condC$histCatch[2,];year=2;
# sigmar=1e-08; Ncl=zone$glb$Nclass; sauindex=zone$glb$sauindex;movem=zone$glb$movem; sigmaF=1e-08
}
