% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{plotbysau}
\alias{plotbysau}
\title{plotbysau plots a collection of projections by SAU}
\usage{
plotbysau(zoneDP, glb.resdir)
}
\arguments{
\item{zoneDP}{the dynamic object produced by the projections}

\item{glb}{the object containing the gloabl variables}

\item{resdir}{the results directory}
}
\value{
nothing by it does add 6 plots to resdir
}
\description{
plotbysau generates and store plots in resir
}
\examples{
print("wait on suitable data-sets")
}
