% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plothistcatch}
\alias{plothistcatch}
\title{plothistcatch generates a plot of the historical catches by SAU}
\usage{
plothistcatch(
  yrs,
  histC,
  nsau,
  saunames,
  commonscale = FALSE,
  proportion = FALSE,
  filen = ""
)
}
\arguments{
\item{yrs}{a vector of the years of catch history}

\item{histC}{the matrix of catches by years}

\item{nsau}{the number of SAU}

\item{saunames}{the SAU names as a vector}

\item{commonscale}{should the SAU plots share a common y-axis scale,
default =  FALSE}

\item{proportion}{should the plot be of proportion relative to the maximum
catch through each time-series or as raw tonnes x year, default=FALSE}

\item{filen}{the name of the file used to save the plot, default="", which
sends the plot to a separate window}
}
\value{
nothing but it does generate a plot either to a window or a file
}
\description{
plothistcatch generates a plot of histircal catches by SAU
    and by Zone. It can plot can tonnes x year or proportion. All SAU
    plots can have independent or common y-axis scales.
}
\examples{
print("wait on suitable data sets")
}
