% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{plothistCE}
\alias{plothistCE}
\title{plothistCE plots the historical cpue used for conditioning}
\usage{
plothistCE(zone1, pops = NULL, rundir = "")
}
\arguments{
\item{zone1}{The zonewide properties as in readzonefile(indir,ctrl$zonefile)}

\item{pops}{default = NULL. The plot can be restricted to a specific set
of populations by providing the indices of the populations to be
included. Thus if there were 8 populations the pops=c(1,2,3,8), would
plot the first three and the last population.}

\item{rundir}{the results directory used by makehtml to store plots and
tables.}
}
\value{
nothing, it adds a plot into rundir and modifies resultTable.csv
}
\description{
plothistCE provides a visual representation of the cpue
    taken in each SAU used in the conditioning of the
    Operating Model.
}
\examples{
print("wait until I have altered the internals data sets")
}
