% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotprod}
\alias{plotprod}
\title{plotprod graphs a selected productivity variable}
\usage{
plotprod(
  product,
  xname = "MatB",
  yname = "Catch",
  xlimit = NA,
  xlab = "Mature Biomass t",
  ylab = "Production t",
  font = 7,
  filename = "",
  devoff = TRUE
)
}
\arguments{
\item{product}{the output from doproduction}

\item{xname}{the name of the productivity variable for the x-axis,
defaults to MatB}

\item{yname}{the name of the y-xis variable, default=Catch=Yield}

\item{xlimit}{default=NA, enables the range of the x-axis to be
constrained, for example using xlimit=c(0.15,0.4)}

\item{xlab}{the x-axis label, default=""}

\item{ylab}{the y-axis label, default=""}

\item{font}{the type of font used in the plot. default = 7, bold
Times, 6 is not bold, 1 is sans-serif, 2 bold sans-serif}

\item{filename}{the complete path and filename of where to save the
png plot file. default is empty, meaning no file is produced.}

\item{devoff}{a boolean to allow the plot device to remain open for
the user to add more components. Of course, if this is set to
FALSE then it is up to the user to use dev.off}
}
\value{
invisibly a list of the x and y matrices plotted
}
\description{
plotprod graphs a selected productivity variable from
    the choice of ExB exploitable biomass, MatB mature of spawning
    biomass = Bmsy, AnnH the actual annual harvest rate Hmsy, Catch
    the yield at Bmsy and Hmsy = MSY, Deplet the mature biomass
    depletion level, and RelCE the relative cpue at MSY
}
\examples{
data(product)
plotprod(product)
stat <- findmsy(product)
abline(h=stat[,"Catch"],col=1:6,lwd=2)
abline(v=stat[,"MatB"],col=1:6,lwd=2)
}
