% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{plotproductivity}
\alias{plotproductivity}
\title{plotproductivity characterizes each population's yield curve}
\usage{
plotproductivity(resdir, runlabel, product, glb)
}
\arguments{
\item{resdir}{the results directory}

\item{runlabel}{the runlabel, ideally from the ctrl object}

\item{product}{the productivity 3-D array}

\item{glb}{the globals list}
}
\value{
nothing but it does place five png files into resdir
}
\description{
plotproductivity characterizes each population's yield curve, it
  also describes the total productivity of the zone.
}
\examples{
print("this will be quite long when I get to it")
}
