% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{plotproductivity}
\alias{plotproductivity}
\title{plotproductivity characterizes each population's yield curve}
\usage{
plotproductivity(rundir, product, glb, hsargs)
}
\arguments{
\item{rundir}{the results directory}

\item{product}{the productivity 3-D array}

\item{glb}{the globals list}

\item{hsargs}{the maximum CPUE target input is inside hsargs. This is
compared with the range of CPUE found across the produciton curve to
ensure that the target is not attempting to achieve the impossible. If
there is no maxtarg, this implies no comparison and no adding to the plot}
}
\value{
It produces five png files into rundir and returns sauprod, a matrix
    of B0, Bmsy, MSY, Depletion-msy, and CEmsy for each sau
}
\description{
plotproductivity characterizes each population's yield curve, it
  also describes the total productivity of the zone.
}
\examples{
print("this will be quite long when I get to it")
}
