% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{plotrateofchange}
\alias{plotrateofchange}
\title{plotrateofchange plot the output from getrateofchange function}
\usage{
plotrateofchange(rundir, res, whichvar, glb, console = TRUE)
}
\arguments{
\item{rundir}{the directory in which comparisons are being made. It is best
to be a separate directory form any particular scenario.}

\item{res}{the output object from the getrateofchange function}

\item{whichvar}{which variable within the dyn object to characterize}

\item{glb}{the globals object needed for the number and names of the sau,
and the years and their names.}

\item{console}{should each plot go to the console or be saved to rundir.
default = TRUE ie go to console}
}
\value{
invisibly the filename used without the path
}
\description{
plotrateofchange produces a plot of the percent changes within
    the whichvar selected across all selected scenarios within do_comparison.
    The outcome is placed within the ScenarioPMs tab. The variables it can
    work with from dyn from teh output of do_comparison include: matureB,
    exploitB, midyrexpB, catch, acatch, harvestR, cpue, recruit, deplsB,
    and depleB.
}
\examples{
# syntax
# pickvar <- "acatch"
# res <- getrateofchange(dyn=dyn,whichvar=pickvar,glb=glb)
# filen <- plotrateofchange(rundir=rundir,res=res,whichvar=pickvar,glb=glb)
# rundir=rundir; res=res;whichvar=pickvar;glb=glb;console=TRUE
}
\seealso{
\link{getrateofchange}, \link{do_comparison}
}
