% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{plotzonechangerate}
\alias{plotzonechangerate}
\title{plotzonechangerate plot the output from getzonechangerate function}
\usage{
plotzonechangerate(rundir, res, whichvar, glb, console = TRUE)
}
\arguments{
\item{rundir}{the directory in which comparisons are being made. It is best
to be a separate directory form any particular scenario.}

\item{res}{the output object from the getzonechangerate function}

\item{whichvar}{which variable within the dyn object to characterize}

\item{glb}{the globals object needed for the number and names of the sau,
and the years and their names.}

\item{console}{should each plot go to the console or be saved to rundir.
default = TRUE ie go to console}
}
\value{
invisibly the filename used without the path
}
\description{
plotzonechangerate produces a plot of the percent changes within
    the whichvar by zone selected across all selected scenarios within
    do_comparison. The outcome is placed within the ScenarioPMs tab. The
    variables it can work with from zone from the output of do_comparison
    include: matureB, exploitB, midyrexpB, catch, acatch, harvestR, cpue,
    recruit, deplsB, and depleB.
}
\examples{
# syntax
# pickvar <- "acatch"
# res <- getzonechangerate(zone=zone,whichvar=pickvar,glb=glb)
# filen <- plotrateofchange(rundir=rundir,res=res,whichvar=pickvar,glb=glb)
}
\seealso{
\link{getzonechangerate}, \link{do_comparison}
}
