% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{plotzonedyn}
\alias{plotzonedyn}
\title{plotzonedyn plots zonewide catch, mature biomass, harvestR and CPUE}
\usage{
plotzonedyn(
  rundir,
  scenes,
  zone,
  glbc,
  console = TRUE,
  q90 = TRUE,
  polys = TRUE,
  intens = 127
)
}
\arguments{
\item{rundir}{the directory in which all results are held for a scenario or
comparison of scenarios}

\item{scenes}{the names of the different scenarios being compared}

\item{zone}{a list of the out$outzone objects from each chosen scenario}

\item{glbc}{all of the global objects from out. They should all have the
same reps and years but this will be checked.}

\item{console}{should the plot be sent to the console or saved for use in the
web-page output? default=TRUE, set to FALSE to save it}

\item{q90}{should the 90th quantiles be plotted? default=TRUE. If FALSE then
the 95th quantiles are used.}

\item{polys}{should transparent polygons be plotted = TRUE, or lines = FALSE}

\item{intens}{if polys=TRUE then intens signifies the intensity of colour on
a scale of 0 - 255. 127 is about 50 percent dense.}
}
\value{
nothing but it does generate a plot and saves it if console=FALSE
}
\description{
plotzonedyn generates plots of the zone-wide total catches, the
    mature biomass, the harvest rate, and the CPUE across the years of
    projection.
}
\examples{
print("wait on data sets")
}
\seealso{
{
  \link{poptozone}, \link{plotZone}
}
}
