% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{plotzoneproj}
\alias{plotzoneproj}
\title{plotzoneproj plots the replicates of a single zone-wide variable}
\usage{
plotzoneproj(zoneV, reps, yrs, label = "", addqnts = TRUE, miny = 0)
}
\arguments{
\item{zoneV}{the variable from within the output of aszone}

\item{reps}{the number of replicates to plot. Generaly one would plot all of
them. If not then it might be best to turn addqnts to FALSE}

\item{yrs}{the vector of years as in 1:(inityrs + projyrs)}

\item{label}{the y-axis label, which should obviously reflect which variable
is chosen from the zone summary list.}

\item{addqnts}{should the median and inner 90 percent quantiles be added to
the plot; default = TRUE}

\item{miny}{sets the lower limit of the y-axis, default=0}
}
\value{
if addqnts=TRUE the quantiles are returned invisibly
}
\description{
plotzoneproj takes the results from the function aszone, which
    contains a set of variables zonesB, zoneeB, zoneC, zoneH, zoneR, zonece,
    zonedeplsB,and zonedepleB, and plots whichever is selected. It contains
    the option of also plotting the median and the inner 90 percent quantiles
    of each year's spread of values across replicates
}
\examples{
print("wait on more time")
# zoneV=zoneproj$zoneC;reps=reps;yrs=1:50;miny=0;label="Catches"; addqnts=TRUE
}
