% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{poptosau}
\alias{poptosau}
\title{poptosau converts projected population dynamics to SAU scale results}
\usage{
poptosau(invar, glb)
}
\arguments{
\item{invar}{either zoneDP$ matureB, exploitB, catch, or recruit}

\item{glb}{the global constants object}
}
\value{
a results array of dimension [projyrs, nSAU,reps]
}
\description{
poptosau the MSE dynamics are run at the population level but
    all management decisions are made at the SAU level, hence the results of
    the projections need to be translated into results at the SAU level.
    poptosau uses the sauindex to sum the variables that can be summed, which
    include matureB, exploitB, catch, and recruit, and combines their
    population results to form their respective SAU results. Thus, for an
    input array of dimensions [projyrs,numpop,reps] one receives back an
    array of [projyrs, nSAU,reps]
}
\examples{
print("wait on appropriate internal datasets")
}
