% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{poptozone}
\alias{poptozone}
\title{poptozone translates the zone_pop objects to a single zone object}
\usage{
poptozone(inzone, NAS = NULL, glb, B0, ExB0)
}
\arguments{
\item{inzone}{one of the zone dynamics objects containing replicates, made
up of populations}

\item{NAS}{the numbers-at-size 4D arrays from doprojection; default=NULL so
it can be ignored during conditioning}

\item{glb}{the object containing the global constants}

\item{B0}{the sum of B0 across all populations, use getvar(zoneC,"B0")}

\item{ExB0}{the sum of ExB0 across all populations use getvar(zoneC,"ExB0")}
}
\value{
a list of dynamics variables by zone
}
\description{
poptozone combines the dynamic results for each variable so
    that results by population become results by zone. matureB, exploitB,
    catch, recruit, catchN, and Nt are simple summations of the totals for
    each population into their respective Zone The harvest rate would be
    end of year or beginning of year estimates derived from dividing the
    catch x zone by the exploitable biomass x zone. Similarly the deplsB and
    depleB are the end of year matureB and exploitB divided by their
    respective unfished estimated by Zone obtained using getvar(zoneC,"B0").
}
\examples{
print("wait on suitable internal data sets ")
}
