% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{prepareprojection}
\alias{prepareprojection}
\title{prepareprojection high level function that sets up a projection}
\usage{
prepareprojection(zone1, zoneC, glb, zoneDep, ctrl)
}
\arguments{
\item{zone1}{the original zone1 object from readctrlfile}

\item{zoneC}{the constant part of the zone}

\item{glb}{the global variables}

\item{zoneDep}{the zone after initial depletion}

\item{ctrl}{the ctrl object}
}
\value{
a list of the dynamic zone object as a list of arrays of projyrs x
    populations x replicates, plus the revised projC and revised zoneC
}
\description{
prepareprojection is a high level function that restructures
    zoneC by including the projected selectivity (in case the LML changes),
    and also selectivity x weight-at-size (for computational speed). It
    then generates a new zoneD with room for all replicates as well as the
    aspirational catches from each HS. Finally, it adds recruitment variation
    to each of the replicates and keeps the last year of each iteration of
    the addrecvar function as the start of each replicate projection, with
    zeros in the catch, cpue, and cesau arrays
}
\examples{
print("wait on data files")
}
