% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.r
\name{preparesizecomp}
\alias{preparesizecomp}
\title{preparesizecomp strips out empty columns and identifies samples}
\usage{
preparesizecomp(sizecomp, mincount = 100)
}
\arguments{
\item{sizecomp}{the size-composition of catch data from the dat object
made by readLBMdata.}

\item{mincount}{the minimum number of observations within a year for
inclusion in the analysis, default = 100.}
}
\value{
a cleaned version of the sizecomp matrix
}
\description{
preparesizecomp takes in the size composition of catch data
    and removes empty columns and those with totals of less than the
    defined mincount. The minimum used depends on sampling routines. In
    Tasmania, a sample of 100 is the usual minimum taken from individual
    landings. Single landings may not be representative of much.
}
\examples{
print("wait on data sets")
}
