% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{product}
\alias{product}
\title{product is the productivity curve matrix from doproduction}
\description{
product is the productivity curve matrix from
    doproduction when the example region is generated using the
    inbuilt datasets ctrl, region1, and constants. The slowest
    part of building the whole is to use the modregC function
    to adjust the regionC and generate the production array. To
    save that time in the examples (to avoid time limits on
    examples should this package go to CRAN), then this dataset can
    be used instead. This is a three dimensional array of
    productivity variables.
}
\section{contents}{

\itemize{
  \item harvestrate the initial harvest rates applied
  \item productivity variables ExB, MatB, AnnH, Catch, Deplet, RelCE
  \item population the index of each population
}
}

\examples{
 data(product)
 product[1:20,,1]
}
