% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{readctrlfile}
\alias{readctrlfile}
\title{readctrlfile reads a csv file for controlling aMSE}
\usage{
readctrlfile(indir, infile = "control.csv")
}
\arguments{
\item{indir}{directory in which to find the control file}

\item{infile}{filename of the control file, default="control.csv"}
}
\value{
a list object containing the control variables
}
\description{
readctrlfile complements the readdatafile. The MSE requires
    a data file to condition the operating model but it also needs a
    control file to setup the details of the simulation test being
    conducted. See ctrlfileTemplate./
}
\examples{
\dontrun{
  direct="./../../rcode2/aMSE/data-raw"
  infile="control.csv"
  ctrl <- readctrlfile(indir=direct,infile=infile)
  str(ctrl)
 }
}
