% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{readctrlfile}
\alias{readctrlfile}
\title{readctrlfile checks datadir contains the required csv files}
\usage{
readctrlfile(rundir, infile = "control.csv", datadir = rundir)
}
\arguments{
\item{rundir}{the directory in which all files relating to a
particular run are to be held. If datadir != rundir then the main data
files are kept in datadir}

\item{infile}{default="control.csv", the filename of the control
file present in datadir containing information regarding the
run.}

\item{datadir}{default = rundir. If datadir != rundir then the data files
for a series of scenarios are kept in this directory.}
}
\value{
the control list for the run
}
\description{
readctrlfile checks datadir contains the required csv
    files including the named control file, which then contains
    the names of the region data file, and the population data
    file. The run stops if any are not present or are misnamed.
}
\examples{
\dontrun{
# this has sinoce been modified and needs updating
rundir <- tempdir()
ctrlzonetemplate(rundir)
datafiletemplate(6,rundir,filename="zone1sau2pop6.csv")
ctrl <- readctrlfile(rundir)
ctrl
}
}
