% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{readdatafile}
\alias{readdatafile}
\title{readdatafile reads in a matrix of data defining each population}
\usage{
readdatafile(indir, infile, glb)
}
\arguments{
\item{indir}{directory in which to find the date file}

\item{infile}{character string with filename of the data file}

\item{glb}{the globals variable from the region file, so obviously
the readregionfile needs to run before readdatafile}
}
\value{
a matrix of values defining the PDFs used to define the
    properties of each population. The contents of popdefs
}
\description{
readdatafile expects a matrix of probability density
    function definitions that are used to define the populaitons
    used in the simualtion. These constitute the definition of
    popdefs.
}
\examples{
data(region1)
glb <- region1$globals
glb
data(constants)
constants
ctrlfile <- "control.csv"
ctrl <- readctrlfile(datadir,ctrlfile)
reg1 <- readregionfile(datadir,ctrl$regionfile)
popdefs <- readdatafile(datadir,ctrl$datafile,reg1$globals)
print(popdefs)
}
