% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{readzonefile}
\alias{readzonefile}
\title{readzonefile reads in the constants for the zone}
\usage{
readzonefile(indir, infile)
}
\arguments{
\item{indir}{directory in which to find the zone file}

\item{infile}{character string with filename of the zone file}
}
\value{
eight objects, four vectors, 4 numbers, and a list of five
    objects
\itemize{
  \item SAUnames the labels given to each SAU
  \item SAUpop the number of populations in each SAU, in sequence
  \item minc scaler defining the mid point of the smallest size class
  \item cw scaler defining the class width
  \item randomseed used to ensure that each simulation run starts
      in the same place. Could use getseed() to produce this.
  \item outyear a vector of three, with Nyrs, fixyear, and firstyear
  \item projLML the LML expected in each projection year, a vector
  \item globals a list containing numpop, nSAU, midpts, Nclass, and
      Nyrs
}
}
\description{
with the zone filename from the control file the
    readzonefile will read the data from a csv file arranged with
    a standard layout. Once again this uses the utility functions
    for reading and parsing lines of text. This is illustrated by
    the function makezonefile, which produces an example *.csv
    file that can then be customized to suit your own simulations.
    Each required section contains a series of constants which are
    read in individually,so their labels are equally important.
}
\examples{
\dontrun{
datadir <- "./../../rcode2/aMSE/data-raw/"
ctrlfile <- "control.csv"
ctrl <- readctrlfile(datadir,ctrlfile)
reg1 <- readzonefile(datadir,ctrl$zonefile)
str(reg1)
}
}
