% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{regionfiletemplate}
\alias{regionfiletemplate}
\title{regionfiletemplate generates a template region file}
\usage{
regionfiletemplate(indir, filename = "region1.csv")
}
\arguments{
\item{indir}{directory in which to place the region file}

\item{filename}{the name for the generated region file, a character
string that defaults to region1.csv. It is best to include the
complete path}
}
\value{
nothing, but it creates a template region file within indir
}
\description{
regionfiletemplate generates a standardized region file
    template containing constants tha toperate at a regional level.
    Generate this and then modify the contents to suit
    the system you are attempting to simulate.
}
\examples{
 yourdir <- tempdir()
 regionfiletemplate(yourdir,filename="region1.csv")   #
 region1 <- readregionfile(yourdir,"region1.csv")
 str(region1,max.level=1)
}
