% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{rewritecontrolfile}
\alias{rewritecontrolfile}
\title{rewritecontrolfile generates a revised control file}
\usage{
rewritecontrolfile(indir, zone1)
}
\arguments{
\item{indir}{the directory into which the new file should be written. This
would usually be the same as rundir, the scenario directory}

\item{zone1}{a large object inside the output object from makeequilzone that
contains SAUnames, SAUpop, minc, cw, larvdisp, randomseed, initLML,
condC, projC, globals, ctrl, catches, and projyrs. Obviously there is
some redundency.}
}
\value{
nothing but it does write a file called 'new_controlfile.csv' into
    rundir
}
\description{
rewritecontrolfile is used to generate a new control file after
    the option of reading in parameters from sizemod has been used. The
    rewritten file contains the revised parameter values.
}
\examples{
print("wait on internal data sets")
}
