% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{runthreeH}
\alias{runthreeH}
\title{runthreeH conducts the dynamics with constant catch 3 times}
\usage{
runthreeH(zoneC, zoneD, inHarv, glob, maxiter = 2)
}
\arguments{
\item{zoneC}{the constants components of the simulated zone}

\item{zoneD}{the dynamics portion of the zone, with matrices and
arrays for the dynamic variables of the dynamics of the
operating model}

\item{inHarv}{a vector, length numpop, of annual harvest rates to be held
constant across all years.}

\item{glob}{the globals variable from readzonefile}

\item{maxiter}{default=3; the number of runs through the equilibrium loop.}
}
\value{
a list containing a revised dynamics list, zoneD
}
\description{
runthreeH is used when searching numerically for an
    equilibrium and it conducts the hyrs dynamics three times, each
    time through it replaces year 1 with year hyrs. Thus if hyrs is
    40 it conducts 3 * 39 years of dynamics (117 years). This is
    not exported. It uses zoneC but always it does this inside
    the environment of another function where zoneC can be found
    Used inside dodepletion and doproduction. maxiter may need to be
    increased when we introduce a larger movement rate between populations
    for greenlip, or if the number of conditioning years are fewer than 45.
}
\examples{
print("wait on built in data sets")
# zoneC=zoneC; zoneD=zoneD; glob=glob; inHarv=rep(initH[aH],numpop); maxiter=2
}
\seealso{
{
 \link{dodepletion}, \link{doproduction}
}
}
