% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{runthreeH}
\alias{runthreeH}
\title{runthree conducts the dynamics with constant catch 3 times}
\usage{
runthreeH(regC, regD, glob, inHarv)
}
\arguments{
\item{regC}{the constant portion of the region with a list of
properties for each population}

\item{regD}{the dynamics portion of the region, with matrices and
arrays for the dynamic variables of the dynamics of the
operating model}

\item{glob}{the globals variable from readregionfile}

\item{inHarv}{a vector of annual harvest rates to be held constant
across all years.}
}
\value{
a list containing a revised dynamics list, regionD
}
\description{
runthree is used when searching numerically for an
    equilibrium and it conducts the Nyrs dynamics three times, each
    time through it replaces year 1 with year Nyrs. Thus if Nyrs is
    40 it conducts 3 * 39 years of dynamics (117 years).
}
\examples{
print("wait on built in data sets")
}
