% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{setupregion}
\alias{setupregion}
\title{setupregion makes the region's constant, dynamic, and productivity parts}
\usage{
setupregion(constants, glb, region1)
}
\arguments{
\item{constants}{the population constants derived from readdatafile}

\item{glb}{the global constants out of region1}

\item{region1}{the regional constants}
}
\value{
a list of regionC, regionD and product, the main components
    of the region
}
\description{
setupregion makes the region's constant, dynamic, and
    productivity parts returning them all in a list
}
\examples{
print("this will be quite long when I get to it")
}
