% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{setupzone}
\alias{setupzone}
\title{setupzone makes zone's constant, dynamic, and productivity parts}
\usage{
setupzone(constants, zone1, uplim = 0.4, inc = 0.005)
}
\arguments{
\item{constants}{the population constants derived from readdatafile}

\item{zone1}{the zonal object}

\item{uplim}{the upper limit of harvest rate applied, default=0.4}

\item{inc}{the harvest rate increment at each step, default=0.005}
}
\value{
a list of zoneC, zoneD, product, and glb the main
    components of the zone
}
\description{
setupzone makes the zone's constant, dynamic, and
    productivity parts returning them, along with glb, in a list.
}
\examples{
\dontrun{
data(constants)
data(zone1)
out <- setupzone(constants,zone1)
zoneC <- out$zoneC
glb <- out$glb
str(zoneC[[1]])
str(glb)
}
}
