% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{setupzone}
\alias{setupzone}
\title{setupzone makes zone's constant, dynamic, and productivity parts}
\usage{
setupzone(
  constants,
  zone1,
  doproduct,
  selectyr,
  lowlim = 0.005,
  uplim = 0.4,
  inc = 0.005,
  verbose = TRUE
)
}
\arguments{
\item{constants}{the population constants derived from readdatafile}

\item{zone1}{the zonal object driving the construction}

\item{doproduct}{boolean, should the productivity calculations be made
during the conditioning. defined in do_MSE and makeequilzone}

\item{selectyr}{which year's LML should be used to estimate the LML. This
is set in makeequilzone. If set = 0, the default, then the LML for
the glb$hyrs, the last year of conditioning data will be used.}

\item{lowlim}{defines the lower limit of harvest rate used when estimating
the productivity The default = 0.005 but during testing could be
increased to perhaps 0.2 to speed things up.}

\item{uplim}{the upper limit of harvest rate applied, default=0.4}

\item{inc}{the harvest rate increment at each step, default=0.005}

\item{verbose}{Should progress comments be printed to console, default=TRUE}
}
\value{
a list of zoneC, zoneD, product, and glb the main
    components of the zone
}
\description{
setupzone makes the zone's constant, dynamic, and
    productivity parts returning them, along with glb, in a list. The
    objective of this function is to generate the orignal unfished
    equilibrium zone of nSAU SAU, and numpop populations
}
\examples{
\dontrun{
data(constants)
data(zone1)
out <- setupzone(constants,zone1)
zoneC <- out$zoneC
glb <- out$glb
str(zoneC[[1]])
str(glb)
}
}
