% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparescenarios.R
\name{tabulateproductivity}
\alias{tabulateproductivity}
\title{tabulateproductivity produces tables of the MSY and related statistics}
\usage{
tabulateproductivity(rundir, prods, scenes)
}
\arguments{
\item{rundir}{the full path to the directory holding the all result files}

\item{prods}{the sauprod objects from each scenario in a list}

\item{scenes}{a vector of the names for each scenario}
}
\value{
nothing but it does add nscenario csv files as tables to the rundir
}
\description{
tabulateproductivity writes out csv files for each scenario
    that contain the B0, Bmsy, MSY, Depmsy, and CEmsy for each sau
}
\examples{
print("wait on data sets")
}
