% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS-HCR.R
\name{targscore}
\alias{targscore}
\title{targscore generates the HCR score for the target PM}
\usage{
targscore(vectce, qnt = 0.55)
}
\arguments{
\item{vectce}{the vector of cpue to be assessed}

\item{qnt}{the quantile of the input vector selected as the target,
default = 0.55}
}
\value{
a list of the final year's score, the internals to the calculations,
    and he target and limit reference points
}
\description{
targscore takes in a vector of cpue and define the targ from
    the complete series as well as the limit reference point and an upper
    level that is symmetrically spaced around the
}
\examples{
print("wait on suitable data")
}
