% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineZone.R
\name{testequil}
\alias{testequil}
\title{testequil runs a zone for Nyrs and determines stability}
\usage{
testequil(zoneC, zoneD, glb, inH = 0, verbose = TRUE)
}
\arguments{
\item{zoneC}{the constants components of the simulated zone}

\item{zoneD}{the dynamic components of the simulated zone}

\item{glb}{the global variables}

\item{inH}{a vector of numpop harvest rates}

\item{verbose}{should results go to the console, default=TRUE}
}
\value{
the dynamics component with Nyrs of dynamics
}
\description{
testequil runs a given zone for Nyrs at the given
    harvest rate, and then tests that the last values of matureB,
    exploitB, recruitment, and spawning biomass depletion are the
    same as the first (to three decimal places). It reports this
    to the console if verbose=TRUE. This is used with a harvest rate
    of zero and no variation in recruitment when defining the equilibrium
    zone under the application of the movement matrix.
}
\examples{
\dontrun{  # modzoneC takes too long to run because of doproduction
 data(ctrl)
 data(zone1)
 glb <- zone1$globals
 data(constants)
 ans <- makezoneC(zone1,constants)
 zoneC <- ans$zoneC
 glb <- ans$glb # now contains movement matrix
 ans <- makezone(glb,zoneC)
 zoneC <- ans$zoneC  # zone constants
 zoneD <- ans$zoneD
 ans <- modzoneC(zoneC,zoneD,glb)
 zoneC <- ans$zoneC  # zone constants
 zoneDe <- testequil(zoneC=zoneC,zoneD=zoneD,glb=glb)
}    #zoneC=zoneC; zoneD=zoneD; glb=glb; inH=0.0; verbose=TRUE
}
