% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMSE_utils.R
\name{uniquepairs}
\alias{uniquepairs}
\title{uniquepairs identifies unique combinations of two columns in a matrix}
\usage{
uniquepairs(x, col1 = 1, col2 = 2)
}
\arguments{
\item{x}{the projLML matrix or any minimum two column matrix}

\item{col1}{the column number of the first variable}

\item{col2}{the column number of the second variable}
}
\value{
invisibly a list of the number of unique pairs, a matrix of all
    possible pairs, the row indices of each of the unique pairs in the
    matrix, and the names of all possible combinations
}
\description{
uniquepairs identifies unique combinations of two columns in a
    matrix. This can be used to identify unique pairs of parameters that
    are used to define things like selectivity. In aMSE this is used when
    implementing a slot selectivity so that one needs to identify all years
    in which a change to the LML structure occurs. This can include a
    change in the LML and or the maximum allowable size = MAL
}
\examples{
lmls <- c(145,210,145,210,145,210,145,210,145,185,150,185,150,185,
          150,185,150,185,150,185)
projL <- matrix(lmls,nrow=10,ncol=2,dimnames=list(1:10,c("LML","Max")),
                byrow=TRUE)
outp <- uniquepairs(x=projL,col1=1,col2=2)
outp
  # x=lmlall;col1=1;col2=2
}
