% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{writecompdata}
\alias{writecompdata}
\title{writecompdata writes a size-composition file to indir}
\usage{
writecompdata(indir, lfs, filen = "lf_WZ90-20.csv")
}
\arguments{
\item{indir}{the directory into which the new file should be written. This
would usually be the same as rundir, the scenario directory}

\item{lfs}{the 3-dimensional array of length-composition data to be written
to lf_WZ90-20.csv.}

\item{filen}{the filename of the size-comp data, default='lf_WZ90_20.csv'}
}
\value{
nothing but it does write a file called 'lf_WZ90-20.csv'
    into indir=rundir, which will overwrite the same if already there.
}
\description{
writecompdata writes a size-composition file to indir
}
\examples{
\dontrun{
  data(lfs)
  tmpdir <- tempdir()
  writecompdata(tmpdir,lfs)
  dir(tmpdir)
}
}
