% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{zone}
\alias{zone}
\title{zone a list of 8 major components from makeequilzone function}
\description{
zone is the output from the function makeequilzone and contains
    8 R objects including zoneC, zoneD, glb, constants, saudat, product,
    ctrl, and zone1. This is a big object of about 1 Mb.
}
\examples{
 data(zone)
 str(zone,max.level=1)
}
