% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{zone1}
\alias{zone1}
\title{zone1 the constants common to a zone}
\format{
A list of constants that are uniform across all populations
    in a zone
\describe{
  \item{SAUnames}{the names of each SAU}
  \item{SAUpop}{a vector of how many populations in each SAU}
  \item{minc}{the mid-point of the minimum size class}
  \item{cw}{the size-class width in mm}
  \item{larvdisp}{the rate of larval dispersal as a proportion}
  \item{randomseed}{used if results need repeating}
  \item{initLML}{the LML used to generate teh equilibrium population}
  \item{condC}{a list containing any historical data used to condition the
      initial zone}
  \item{projC}{a list containing the details required to prepare the zoneC
      and zoneD for the projections}
  \item{globals}{a list of global constants, containing numpop,
      nSAU, midpts, Nclass, Nyrs, and larvdisp}
  \item{ctrl}{the list containing control information for the run, including
      the datafile for the constants, the reps, the variation to be included
      when projecting}
  \item{condition}{if zero then no historical conditioning, if >0, then the
      number of years of conditioning catch data, complements condC}
  \item{projection}{if zero then no projection conducted, if >0 then the
      number of years of projection}
}
}
\description{
zone1 contains the constants relating to the whole
    zone rather than the populations. See the listing below.
}
\examples{
 data(zone1)
 str(zone1,max.level=1)
}
