% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{zone1}
\alias{zone1}
\title{zone1 the constants common to a zone}
\format{
A list of constants that are uniform across all populations
    in a zone
\describe{
  \item{SAUnames}{the names of each SAU}
  \item{SAUpop}{a vector of how many populations in each SAU}
  \item{minc}{the mid-point of the minimum size class}
  \item{cw}{the size-class width in mm}
  \item{larvdisp}{the rate of larval dispersal as a proportion}
  \item{randomseed}{used if results need repeating}
  \item{outyear}{a vector on Nyrs, fix year, and start year}
  \item{projLML}{the time series of LML used in the MSE projections}
  \item{histLML}{time series of LML if conditioning the model prior
      to projections}
  \item{globals}{a list of global constants, containing numpop,
      nSAU, midpts, Nclass, Nyrs, and larvdisp}
}
}
\description{
zone1 contains the constants relating to the whole
    zone rather than th epopulations. See th elisting below.
}
\examples{
 data(zone1)
 print(zone1)
}
